==> code/old-code/02_prelim-visualizations.Rmd <==
---
title: "preliminary visualizations"
author: "An Bui"
date: "10/6/2021"
output: 
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
    theme: readable
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 0. set up
```{r message = FALSE, warning = FALSE, echo = FALSE}
# source the cleaning/summarizing script, which sources the set up
source(here::here("code", "01_cleaning-and-summarizing.R"))
```

## summarizing

```{r}
count <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  group_by(sp_code) %>% 
  count() %>% 
  rename(count = n) %>% 
  ungroup() %>% 
  arrange(count) %>% 
  mutate(sp_code = fct_inorder(sp_code)) %>% 
  mutate(left = max(count) - count) %>% 
  pivot_longer(cols = count:left, names_to = "sample", values_to = "values") %>% 
  mutate(label = case_when(
    sample == "left" ~ paste(values, "/30", sep = ""),
    sample == "count" ~ NA
  ),
  label = case_when(
    sample == "left" & sp_code == "CYOS" ~ paste(label, "samples left"),
    TRUE ~ label
  )) %>% 
  mutate(sample = case_when(
    sample == "left" ~ "Left to sample", 
    sample == "count" ~ "Already sampled"
  ), 
  sample = fct_relevel(sample, "Left to sample", "Already sampled"))

all <- ggplot(count, aes(x = values, y = sp_code, fill = sample)) +
  scale_fill_manual(values = c("Already sampled" = "cornflowerblue", "Left to sample" = "coral")) +
  geom_col() +
  geom_text(aes(label = label, x = 31), size = 8, hjust = 0) +
  scale_x_continuous(limits = c(0, 60), breaks = c(0, 10, 20, 30), expand = c(0, 0)) +
  # coord_cartesian(xlim = c(0, 50)) +
  labs(y = "", x = "", fill = "") +
  theme_classic() +
  theme(legend.position = "bottom",
        axis.line = element_blank(),
        axis.ticks.y = element_blank(), 
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        axis.title.x = element_blank(),
        legend.text = element_text(size = 20), 
        legend.justification = "right",
        plot.margin = unit(c(0, 0, 0, 0), units = "cm")) 

already <- ggplot(count %>% filter(sample == "Already sampled"), aes(x = values, y = sp_code)) +
  geom_col(fill = "cornflowerblue") +
  geom_text(aes(label = label, x = 31), size = 8, hjust = 0) +
  scale_x_continuous(limits = c(0, 60), breaks = c(0, 10, 20, 30), expand = c(0, 0)) +
  # coord_cartesian(xlim = c(0, 50)) +
  labs(y = "Species code", x = "", fill = "") +
  theme_classic() +
  theme(legend.position = "bottom",
        axis.line = element_blank(),
        axis.ticks.y = element_blank(), 
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        legend.text = element_text(size = 20), 
        legend.justification = "right",
        plot.margin = unit(c(0, 0, 0, 0), units = "cm")) 

all
already
together <- already + all +
  plot_layout(guides = "collect") & 
  theme(legend.position = 'bottom')
together
```

# 1. FvFm 

## a. analysis of variance

```{r}
fvfm_df <- fvfm_raw %>% 
  # only use FvFm, drop the NAs, change the column name
  select(specimen_ID, '1:Fv/Fm') %>% 
  drop_na() %>%
  rename('fvfm_meas' = '1:Fv/Fm') %>% 
  # make sure that fvfm_meas is numeric
  mutate(fvfm_meas = as.numeric(fvfm_meas)) %>% 
  # join with metadata: spp codes + specimen_ID
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  # join with coarse_traits
  left_join(., coarse_traits, by = "sp_code") %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(sp_code != "POLA")

fvfm_df <- fvfm_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  # join with coarse_traits
  left_join(., coarse_traits, by = "sp_code") %>% 
  filter(sp_code %in% algae_proposal)  %>% 
  filter(sp_code != "POLA")

bartlett.test(fvfm_mean ~ sp_code, data = fvfm_df)
# heterogeneous variance

# analysis of variance
fvfm_sp_aov <- aov(fvfm_mean ~ sp_code, data = fvfm_df)

fvfm_tax_aov <- aov(fvfm_mean ~ taxon_phylum, data = fvfm_df)

summary(fvfm_sp_aov)
summary(fvfm_tax_aov)
# both significant

# Tukey HSD
fvfm_tukey <- TukeyHSD(fvfm_sp_aov)

# compact letter display
fvfm_cld <- multcompLetters4(fvfm_sp_aov, fvfm_tukey)

# output of compact letter display in data frame
fvfm_cld_df <- enframe(fvfm_cld[[1]]$Letters) %>% 
  rename(sp_code = "name")
```

## b. box plot

```{r}
fvfm_boxplot <- ind_traits %>% 
  left_join(., fvfm_cld_df, by = "sp_code") %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = fvfm_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Photosynthetic efficiency",
       x = "Species",
       y = "Fv/Fm") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))

fvfm_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("fvfm-boxplot-", today(), ".jpg", sep = "")),
#        fvfm_boxplot,
#        width = 8, height = 5, dpi = 200)
```

## c. boxplot (PTCA adults vs recruits)

```{r}
ptca_fvfm <- fvfm_ind %>% 
  # rename('fvfm_meas' = '1:Fv/Fm') %>% 
  # mutate(fvfm_meas = as.numeric(fvfm_meas)) %>% 
  # select(specimen_ID, subsample_ID, fvfm_meas) %>% 
  # group_by(specimen_ID) %>% 
  # summarize(mean = mean(fvfm_meas),
  #           var = var(fvfm_meas),
  #           sd = sd(fvfm_meas)) %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "PTCA" & site %in% c("Mohawk", "Arroyo Quemado")) 

var.test(fvfm_mean ~ lifestage, data = ptca_fvfm %>% filter(site == "Arroyo Quemado"))
var.test(fvfm_mean ~ lifestage, data = ptca_fvfm %>% filter(site == "Mohawk"))
# both homogeneous variances

t.test(fvfm_mean ~ lifestage, data = ptca_fvfm %>% filter(site == "Arroyo Quemado"), var.equal = TRUE)
t.test(fvfm_mean ~ lifestage, data = ptca_fvfm %>% filter(site == "Mohawk"), var.equal = TRUE)

var.test(fvfm_mean ~ site, data = ptca_fvfm %>% filter(lifestage == "adult"))
var.test(fvfm_mean ~ site, data = ptca_fvfm %>% filter(lifestage == "recruit"))
# homogeneous variance
t.test(fvfm_mean ~ site, data = ptca_fvfm %>% filter(lifestage == "adult"), var.equal = TRUE)
t.test(fvfm_mean ~ site, data = ptca_fvfm %>% filter(lifestage == "recruit"), var.equal = TRUE)
# not significantly different

ptca_fvfm_plot <- ptca_fvfm %>% 
  ggplot(aes(x = lifestage, y = fvfm_mean)) +
  geom_boxplot(aes(fill = lifestage)) +
  geom_jitter() +
  scale_fill_manual(values = c("#70601d", "#b59b33")) +
  facet_wrap(~site) +
  theme_bw() +
  theme(legend.position = "none") +
  labs(title = "PTCA Fv/Fm",
       caption = "t-test: Arroyo Quemado: t(14) = 3.52, p = 0.003, Mohawk: t(28) = 7.13, p < 0.001") 

ptca_fvfm_plot

# test_plot <- fvfm_raw %>% 
#   rename('fvfm_meas' = '1:Fv/Fm') %>% 
#   mutate(fvfm_meas = as.numeric(fvfm_meas)) %>% 
#   select(specimen_ID, subsample_ID, fvfm_meas) %>% 
#   group_by(specimen_ID) %>% 
#   summarize(mean = mean(fvfm_meas),
#             var = var(fvfm_meas),
#             sd = sd(fvfm_meas)) %>% 
#   left_join(., metadata, by = "specimen_ID") %>% 
#   filter(sp_code == "PTCA" & site %in% c("Mohawk", "Arroyo Quemado")) %>% 
#   mutate(lifestage = case_when(
#     date_collected %in% c("20220405", "20220407") ~ "adult",
#     TRUE ~ "new_recruit"
#   )) %>% 
#   ggplot(aes(x = lifestage, y = mean)) +
#   geom_boxplot(aes(fill = lifestage)) +
#   geom_jitter() +
#   scale_fill_manual(values = c("#70601d", "#b59b33")) +
#   facet_wrap(~site) +
#   theme_bw() +
#   theme(legend.position = "none") +
#   labs(title = "PTCA Fv/Fm",
#        caption = "t-test: Arroyo Quemado: t(14) = 9.4, p < 0.001, Mohawk: t(28) = 7.13, p < 0.001") 
# 
# test_plot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("fvfm-boxplot-lifestage-", todays_date, ".jpg", sep = "")),
#        ptca_fvfm_plot,
#        width = 8, height = 5, dpi = 150)
```

## d. boxplot (PTCA recruits across sites)

```{r}
fvfm_ptca_recruit <- fvfm_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(lifestage == "recruit")

bartlett.test(fvfm_mean ~ site, data = ptca_fvfm)
# not significantly different variances

summary(aov(fvfm_mean ~ site, 
            data = fvfm_ptca_recruit))

fvfm_boxplot_ptca_recruit <- fvfm_ptca_recruit %>% 
  ggplot(aes(x = site, y = fvfm_mean)) +
  geom_boxplot() +
  geom_jitter() +
  theme_bw() +
  labs(caption = "significantly different across sites, ANOVA: F(3) = 14.89, p < 0.001")
fvfm_boxplot_ptca_recruit

# ggsave(here::here("figures/basic-visualizations",
#                   paste("fvfm-boxplot-PTCA_recruits-", todays_date, ".jpg", sep = "")),
#        fvfm_boxplot_ptca_recruit,
#        width = 8, height = 5, dpi = 150)
```

## e. boxplot (PTCA adults across sites)

```{r}
fvfm_ptca_adult <- fvfm_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(lifestage == "adult") %>% 
  filter(site != "Bullito" & year != "2021")

bartlett.test(fvfm_mean ~ site, data = fvfm_ptca_adult)
# not significantly different variances

summary(aov(fvfm_mean ~ site, data = fvfm_ptca_adult))
# significant differences in adults

TukeyHSD(aov(fvfm_mean ~ site, data = fvfm_ptca_adult))

fvfm_boxplot_ptca_adult <- fvfm_ptca_adult %>% 
  ggplot(aes(x = site, y = fvfm_mean)) +
  geom_boxplot() +
  geom_jitter() +
  theme_bw() +
  facet_wrap(~year)
  labs(caption = "ANOVA: F(2) = 3.273, p = 0.06")
fvfm_boxplot_ptca_adult

ggsave(here::here("figures/basic-visualizations",
                  paste("fvfm-boxplot-PTCA_adults-", todays_date, ".jpg", sep = "")),
       fvfm_boxplot_ptca_adult,
       width = 8, height = 5, dpi = 150)

# boxplot for Billie - move this later
plot <- fvfm_raw %>% 
  rename('fvfm_meas' = '1:Fv/Fm') %>% 
  # fill in '1:Fv/Fm' with 1:Y(II) when there is none
  # only an issue with 20220405-MOHK samples
  mutate(fvfm_meas = case_when(
    !is.na((fvfm_meas)) ~ `1:Y (II)`,
    TRUE ~ fvfm_meas
  )) %>% 
  # take out observations with "-" in the measurement
  filter(fvfm_meas != "-") %>% 
  # only use FvFm, drop the NAs, change the column name
  select(specimen_ID, subsample_ID, fvfm_meas) %>% 
  drop_na(specimen_ID) %>% 
  # make sure that fvfm_meas is numeric
  mutate(fvfm_meas = as.numeric(fvfm_meas)) %>% 
  left_join(., metadata_sub, by = "subsample_ID") %>% 
  filter(sp_code == "PTCA" & lifestage == "adult" & site != "Bullito") %>% 
  # weird value, may be mistake in data entry
  filter(fvfm_meas != 0.097) %>% 
  ggplot(aes(x = fvfm_meas, fill = site)) +
  geom_histogram(bins = 40, color = "black") +
  scale_y_continuous(expand = c(0, 0), limits = c(0, 300)) +
  scale_x_continuous(breaks = seq(from = 0.2, to = 1, by = 0.05)) +
  scale_fill_manual(values = cal_palette("superbloom2", n = 3, type = "discrete")) +
  theme_bw() +
  theme(legend.position = c(0.12, 0.85),
        panel.grid.minor.x = element_blank()) +
  labs(x = "FvFm measurement",
       y = "Count",
       fill = "Site",
       title = "Counts of PTCA blade measurements",
       subtitle = "~10 readings per blade, 167 blades, 22 individuals, 1664 readings total")
plot

ggsave(here::here("figures", "basic-visualizations", paste("fvfm-boxplot-01_BB_", today(), ".jpg", sep = "")))
  
adult_plot <- fvfm_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(lifestage == "adult" & site != "Bullito") %>% 
  ggplot(aes(x = site, y = fvfm_mean)) +
  geom_violin(aes(fill = site)) +
  scale_fill_manual(values = cal_palette("superbloom2", n = 3, type = "discrete")) +
  theme_bw() +
  theme(legend.position = "none") +
  labs(x = "Site",
       y = "Mean FvFm per individual",
       title = "Mean PTCA FvFm by individual",
       subtitle = "~10 readings per blade, 167 blades, 22 individuals")
adult_plot
ggsave(here::here("figures", "basic-visualizations", paste("fvfm-boxplot-02_BB_", today(), ".jpg", sep = "")))
```

## f. boxplot (BF BULL vs IVEE)

```{r}
bf_fvfm <- fvfm_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "BF" & site %in% c("Bullito", "Isla Vista"))

var.test(fvfm_mean ~ site, data = bf_fvfm)
# homogeneous variances, use t-test
t.test(fvfm_mean ~ site, data = bf_fvfm, var.equal = TRUE)

bf_plot <- bf_fvfm %>% 
  ggplot(aes(x = site, y = fvfm_mean)) +
  geom_boxplot(aes(fill = site)) +
  geom_jitter() +
  labs(title = "BF Fv/Fm",
       caption = "t-test: t(22) = 2.46, p = 0.02") +
  theme_bw() +
  theme(legend.position = "none")

bf_plot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("fvfm-boxplot-bf-", todays_date, ".jpg", sep = "")),
#        bf_plot,
#        width = 8, height = 5, dpi = 150)
```

# 2. Thickness

## a. analysis of variance

```{r}
thickness_df <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # remove POLA because only 1 observation
  filter(!(sp_code %in% c("POLA")))

bartlett.test(thickness_mm_mean ~ sp_code, data = thickness_df)
# definitely heterogeneous variances

# analysis of variance
thickness_sp_aov <- aov(thickness_mm_mean ~ sp_code, data = thickness_df)

summary(thickness_sp_aov)

# Tukey HSD
thickness_tukey <- TukeyHSD(thickness_sp_aov)

# compact letter display
thickness_cld <- multcompLetters4(thickness_sp_aov, thickness_tukey)

# output of compact letter display in data frame
thickness_cld_df <- enframe(thickness_cld[[1]]$Letters) %>% 
  rename(sp_code = "name")
```

## b. box plot

```{r}
thickness_boxplot <- ind_traits %>% 
  left_join(., thickness_cld_df, by = "sp_code") %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # drop_na(site, mean) %>% 
  group_by(sp_code) %>% 
  ggplot(aes(x = sp_code, y = thickness_mm_mean, group = sp_code)) +
  geom_boxplot(aes(fill = value)) +
  geom_jitter() +
  geom_text(aes(label = value, y = 0.75), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Thallus thickness",
       x = "Species code",
       y = "Thickness (mm)") +
  theme(legend.position = "none")

thickness_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("thickness-boxplot-", todays_date, ".jpg", sep = "")),
#        thickness_boxplot,
#        width = 8, height = 5, dpi = 150)
```

## c. boxplot (BF BULL vs IVEE)

```{r}
bf_thickness <- thickness_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "BF")
  
var.test(thickness_mm_mean ~ site, data = bf_thickness)
# homogeneous variance, can use two-sample t-test

t.test(thickness_mm_mean ~ site, data = bf_thickness, var.equal = TRUE)

thickness_boxplot_bf <- bf_thickness %>% 
  ggplot(aes(x = site, y = thickness_mm_mean)) +
  geom_boxplot(aes(fill = site)) +
  geom_jitter() +
  # geom_text(aes(x = site, y = thickness_mm_mean, label = specimen_ID)) +
  theme_bw() +
  labs(title = "BF thallus thickness",
       x = "Site",
       y = "Thickness (mm)",
       caption = "t-test: t(22) = 2.3, p = 0.03") +
  theme(legend.position = "none")

thickness_boxplot_bf

# ggsave(here::here("figures/basic-visualizations",
#                   paste("thickness-boxplot-bf-", todays_date, ".jpg", sep = "")),
#        thickness_boxplot_bf,
#        width = 8, height = 5, dpi = 150)

```

## d. boxplot (PTCA adults vs recruits)

```{r}
ptca_thickness <- thickness_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "PTCA") 

var.test(thickness_mm_mean ~ lifestage, data = ptca_thickness)
var.test(thickness_mm_mean ~ site, data = ptca_thickness %>% filter(lifestage == "adult"))
# not significantly different variances

summary(aov(thickness_mm_mean ~ lifestage, 
            data = ptca_thickness))
# significantly different between lifestages
t.test(thickness_mm_mean ~ site, data = ptca_thickness %>% filter(lifestage == "adult"), var.equal = TRUE)

thickness_boxplot_ptca <- ptca_thickness %>% 
  ggplot(aes(x = lifestage, y = thickness_mm_mean)) +
  geom_boxplot(aes(fill = lifestage)) +
  scale_fill_manual(values = c("#70601d", "#b59b33")) +
  geom_jitter() +
  theme_bw() +
  facet_wrap(~site) +
  labs(title = "PTCA adult vs recruit thickness",
       y = "mm",
       caption = "significantly different variances") +
  theme(legend.position = "none")
thickness_boxplot_ptca

# ggsave(here::here("figures/basic-visualizations",
#                   paste("thickness-boxplot-lifestage-", todays_date, ".jpg", sep = "")),
#        thickness_boxplot_ptca,
#        width = 8, height = 5, dpi = 150)
```


# 3. Weights

## a. analysis of variance

```{r}
weight_df <- weight_ind %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # take out POLA, only one sample
  filter(!(sp_code %in% c("POLA")))

# analysis of variance
dryw_sp_aov <- aov(total_dry ~ sp_code, data = weight_df)

# dryw_tax_aov <- aov(total ~ taxon_phylum, data = weight_summary %>% filter(!(sp_code %in% c("BO", "CO"))))

summary(dryw_sp_aov)
# summary(dryw_tax_aov)
# species significant, phylum not significant

# Tukey HSD
dryw_tukey <- TukeyHSD(dryw_sp_aov)

# compact letter display
dryw_cld <- multcompLetters4(dryw_sp_aov, dryw_tukey)

# output of compact letter display in data frame
dryw_cld_df <- enframe(dryw_cld[[1]]$Letters) %>% 
  rename(sp_code = "name")

# test homogeneity of variance
# bartlett.test(tdmc ~ sp_code, data = weight_summary %>% filter(!(sp_code %in% c("BO", "CO"))))

# analysis of variance
tdmc_sp_aov <- aov(tdmc_mean ~ sp_code, data = ind_traits %>% filter(sp_code %in% algae_proposal) %>% filter(sp_code != "POLA"))
tdmc_tax_aov <- aov(tdmc_mean ~ pigment_type, data = ind_traits %>% filter(sp_code %in% algae_proposal) %>% filter(sp_code != "POLA"))

summary(tdmc_sp_aov)
summary(tdmc_tax_aov)
# species not significant, phylum significant???

# Tukey HSD
tdmc_tukey <- TukeyHSD(tdmc_sp_aov)

# compact letter display
tdmc_cld <- multcompLetters4(tdmc_sp_aov, tdmc_tukey)

# output of compact letter display in data frame
tdmc_cld_df <- enframe(tdmc_cld[[1]]$Letters) %>% 
  rename(sp_code = "name")
```


## b. box plot

```{r}
dryw_boxplot <- ind_traits %>% 
  filter(sp_code != "BO") %>% 
  left_join(., dryw_cld_df, by = "sp_code") %>% 
  ggplot(aes(x = sp_code, y = total_dry)) +
  geom_boxplot(aes(fill = value)) +
  geom_jitter() +
  geom_text(aes(label = value, y = 9), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Oven dry weight",
       x = "Species code",
       y = "Dry weight (g)") +
  theme(legend.position = "none")

dryw_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("dryw-boxplot-", todays_date, ".jpg", sep = "")),
#        dryw_boxplot,
#        width = 8, height = 5, dpi = 150)

wetw_boxplot <- ggplot(ind_traits, aes(x = sp_code, y = total_wet)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw()  +
  labs(title = "Fresh wet weight",
       x = "Species code",
       y = "Wet weight (g)")

wetw_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("wetw-boxplot-", todays_date, ".jpg", sep = "")),
#        wetw_boxplot,
#        width = 8, height = 5, dpi = 150)

tdmc_boxplot <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(!(sp_code %in% c("POLA"))) %>% 
  left_join(., tdmc_cld_df, by = "sp_code") %>% 
  ggplot(aes(x = sp_code, y = tdmc_mean)) +
  geom_boxplot(aes(fill = value)) +
  geom_jitter() +
  geom_text(aes(label = value, y = 0.9), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_y_continuous(limits = c(0, 1)) +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Thallus dry matter content",
       x = "Species code",
       y = "g dry weight/g wet weight")

tdmc_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("tdmc_boxplot-", todays_date, ".jpg", sep = "")),
#        tdmc_boxplot,
#        width = 8, height = 5, dpi = 150)
```

## c. boxplot (TDMC, BF BULL vs IVEE)

```{r}
bf_tdmc <- ind_traits %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  rename("sp_code" = sp_code.x) %>% 
  filter(sp_code == "BF")

var.test(tdmc_mean ~ site, data = bf_tdmc)
# homogeneous variance
t.test(tdmc_mean ~ site, data = bf_tdmc, var.equal = TRUE)
# not significantly different

tdmc_boxplot_bf <- bf_tdmc %>% 
  ggplot(aes(x = site, y = tdmc_mean)) +
  geom_boxplot(aes(fill = site)) +
  geom_jitter() +
  theme_bw() +
  labs(title = "BF TDMC",
       x = "Site",
       y = "TDMC",
       caption = "t-test: t(22) = 1.85, p = 0.07") +
  theme(legend.position = "none")

tdmc_boxplot_bf

# ggsave(here::here("figures/basic-visualizations",
#                   paste("tdmc-boxplot-bf-", todays_date, ".jpg", sep = "")),
#        tdmc_boxplot_bf,
#        width = 8, height = 5, dpi = 150)
```

## d. boxplot PTCA adults vs recruits

Note 20220507: fix this!!!
```{r}
ptca_leaf <- av_leaf_values %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "PTCA") %>% 
  filter(site %in% c("Arroyo Quemado", "Mohawk"))

var.test(frond_length_mean ~ site, data = ptca_leaf %>% filter(lifestage == "adult"))
var.test(frond_width_mean ~ site, data = ptca_leaf %>% filter(lifestage == "adult"))
var.test(tdmc_mean ~ lifestage, data = ptca_leaf)
var.test(tdmc_mean ~ site, data = ptca_leaf %>% filter(lifestage == "adult"))
# not significantly different variances
t.test(tdmc_mean ~ lifestage, data = ptca_leaf, var.equal = TRUE)
t.test(tdmc_mean ~ site, data = ptca_leaf, var.equal = FALSE)
t.test(frond_width_mean ~ site, data = ptca_leaf, var.equal = TRUE)
# signficantly different

tdmc_boxplot_ptca <- ptca_leaf %>% 
  ggplot(aes(x = lifestage, y = tdmc_mean)) +
  geom_boxplot(aes(fill = lifestage)) +
  # geom_text(aes(x = lifestage, y = tdmc, label = specimen_ID))
  geom_jitter() +
  theme_bw() +
  labs(title = "PTCA TDMC",
       x = "lifestage",
       y = "TDMC",
       caption = "t-test: t(57) = 4.008, p < 0.001") +
  theme(legend.position = "none")

tdmc_boxplot_ptca

# ggsave(here::here("figures/basic-visualizations",
#                   paste("tdmc-boxplot-ptca-", todays_date, ".jpg", sep = "")),
#        tdmc_boxplot_ptca,
#        width = 8, height = 5, dpi = 150)
```


# 4. volume

## a. box plot
```{r}
volume_boxplot <- ggplot(ind_traits %>% filter(sp_code %in% algae_proposal), aes(x = sp_code, y = total_volume)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Volume",
       x = "Species code",
       y = "Volume (mL)")

volume_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("volume-boxplot-", todays_date, ".jpg", sep = "")),
#        volume_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 5. surface area

## a. box plot
```{r}
sa_boxplot <- ggplot(sa_peri_summary, aes(x = sp_code, y = area_total)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Surface area",
       x = "Species code",
       y = bquote("Square millimeter "(mm^2)))

sa_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("sa-boxplot-", todays_date, ".jpg", sep = "")),
#        sa_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 6. perimeter

## a. box plot
```{r}
peri_boxplot <- ggplot(sa_peri_summary, aes(x = sp_code, y = peri_total)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Perimeter",
       x = "Species code",
       y = "Perimeter (mm)")

peri_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("peri_boxplot-", todays_date, ".jpg", sep = "")),
#        peri_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 7. surface area:perimeter ratio

## a. box plot

```{r}
ratio_saperi_boxplot <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = sap_mean)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Surface area:perimeter ratio",
       x = "Species code",
       y = expression(paste(mm^2, "per mm")))

ratio_saperi_boxplot 

# ggsave(here::here("figures/basic-visualizations",
#                   paste("ratio_saperi_boxplot-", todays_date, ".jpg", sep = "")),
#        ratio_saperi_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 8. surface area:volume ratio

## a. box plot
```{r}
ratio_savolume_boxplot <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = sav_mean, fill = sp_code)) +
  geom_boxplot() +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Surface area:volume ratio",
       x = "Species code",
       y = "mm^2/mL") +
  theme(legend.position = "none")

ratio_savolume_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("ratio_savolume_boxplot-", todays_date, ".jpg", sep = "")),
#        ratio_savolume_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 9. maximum height and width

```{r}
height_boxplot <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = maximum_height)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Maximum height",
       x = "Species code",
       y = "cm")

height_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("height_boxplot-", todays_date, ".jpg", sep = "")),
#        height_boxplot,
#        width = 8, height = 5, dpi = 150)

width_boxplot <- ggplot(hw_summary, aes(x = sp_code, y = max_width_cm)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Maximum width",
       x = "Species code",
       y = "cm")

width_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("width_boxplot-", todays_date, ".jpg", sep = "")),
#        width_boxplot,
#        width = 8, height = 5, dpi = 150)
```


# 10. specific thallus area

## a. analysis of variance

```{r}
sta_df <- ind_traits %>% 
  filter(specimen_ID != "20210721-BULL-023") %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(sp_code != "POLA") %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  drop_na(sta_mean)

# test homogeneity of variance
bartlett.test(sta_mean ~ sp_code.x, data = sta_df)
# heterogeneous variances

# analysis of variance
sta_sp_aov <- aov(sta_mean ~ sp_code.x, data = sta_df)
sta_sp_lm <- lm(sta_mean ~ sp_code.x, data = sta_df, na.action = na.pass)
sta_sp_lm.s <- emmeans(sta_sp_lm, specs = "sp_code.x")
pairs(sta_sp_lm.s)
multcomp::cld(sta_sp_lm.s, alpha = 0.95, Letters = LETTERS, signif.sets = TRUE)

# sta_tax_aov <- aov(sta_mm_mg ~ taxon_phylum, data = sta_df)

summary(sta_sp_aov)
# summary(sta_tax_aov)
# both significant

# Tukey HSD
sta_tukey <- TukeyHSD(sta_sp_aov)

# compact letter display
sta_cld <- multcompLetters4(sta_sp_aov, sta_tukey)

# output of compact letter display in data frame
sta_cld_df <- enframe(sta_cld[[1]]$Letters) %>% 
  rename(sp_code = "name")
```

## b. boxplot

thallus area:dry weight

```{r}
sta_boxplot <- sta_df %>% 
  filter(! specimen_ID == "20230620-NAPL-033") %>% 
  # left_join(., sta_cld_df, by = c("sp_code.x" = "sp_code")) %>% 
  ggplot(aes(x = sp_code.x, y = sta_mean, group = sp_code.x)) +
  # geom_boxplot(aes(fill = sp_code.x)) +
  geom_point(aes(color = sp_code.x), position = position_jitter(width = 0.3, seed = 666), alpha = 0.6, shape = 21) +
  scale_color_manual(values = algae_spcode_colors) +
  stat_summary(geom = "pointrange", fun.data = mean_se) +
  # geom_text(aes(label = value, y = 230), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Specific thallus area",
       x = "Species code",
       y = "Thallus area/dry weight (mm^2/mg)") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))

sta_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("sta_boxplot-", today(), ".jpg", sep = "")),
#        sta_boxplot,
#        width = 8, height = 5, dpi = 150)
```

## c. boxplot (BF BULL vs IVEE)

```{r}
var.test(sta_mean ~ site, data = sta_df %>% filter(sp_code.x == "BF"))
# heterogeneous variances
t.test(sta_mean ~ site, data = sta_df %>% filter(sp_code.x == "BF"), var.equal = FALSE)
# no significant difference in means

sta_boxplot_bf <- sta_summary %>% 
  filter(sp_code == "BF") %>% 
  ggplot(aes(x = site, y = sta_mm_mg)) +
  geom_boxplot(aes(fill = site)) +
  geom_jitter() +
  theme_bw() +
  labs(title = "BF STA",
       x = "Site",
       y = "Thallus area/dry weight",
       caption = "Welch t-test: t(8.62) = -0.09, p = 0.93") +
  theme(legend.position = "none")

sta_boxplot_bf

# ggsave(here::here("figures/basic-visualizations",
#                   paste("sta-boxplot-bf-", todays_date, ".jpg", sep = "")),
#        sta_boxplot_bf,
#        width = 8, height = 5, dpi = 150)
```

## d. boxplot (PTCA adults vs recruits)

```{r}
var.test(sta_mean ~ lifestage, data = ptca_leaf)
bartlett.test(sta_mean ~ lifestage, data = ptca_leaf)
leveneTest(sta_mean ~ lifestage, data = ptca_leaf)
# significantly different variances

t.test(sta_mean ~ lifestage, data = ptca_leaf, var.equal = FALSE)

summary(aov(sta_mean ~ lifestage + site, 
            data = ptca_leaf))
# significantly different between lifestages

summary(aov(sta_mm_mg ~ site, 
            data = sta_ptca %>% 
              filter(lifestage == "new_recruit")))
var.test(sta_mm_mg ~ site, data = sta_ptca %>% filter(lifestage == "new_recruit"))
# significantly different STA between sites for new recruits (significantly different variances)
summary(aov(sta_mm_mg ~ site, 
            data = sta_ptca %>% 
              filter(lifestage == "adult")))
var.test(sta_mm_mg ~ site, data = sta_ptca %>% filter(lifestage == "adult"))
# not for adults

sta_boxplot_ptca <- ptca_leaf %>% 
  ggplot(aes(x = lifestage, y = sta_mean)) +
  geom_boxplot(aes(fill = lifestage)) +
  scale_fill_manual(values = c("#70601d", "#b59b33")) +
  geom_jitter() +
  theme_bw() +
  facet_wrap(~site) +
  labs(title = "PTCA adult vs recruit sta",
       y = "mm/mg",
       caption = "significantly different variances") +
  theme(legend.position = "none")
sta_boxplot_ptca
```

## e. boxplot (PTCA)

```{r}
ptca_sta_adults <- ind_traits %>% 
  filter(sp_code == "PTCA") 

bartlett.test(sta_mean ~ site, data = ptca_sta_adults)
# not significantly different variances

summary(aov(sta_mean ~ site, data = ptca_sta_adults, var.equal = TRUE))
# significantly different between sites

sta_boxplot_ptca <- ptca_sta_adults %>% 
  ggplot(aes(x = site, y = sta_mean)) +
  geom_boxplot(aes(fill = site)) +
  # scale_fill_manual(values = c("#70601d", "#b59b33")) +
  geom_jitter() +
  theme_bw() +
  labs(title = "PTCA adults",
       y = "mm/mg",
       caption = "analysis of variance: F(1) = 0.004, p = 0.948") +
  theme(legend.position = "none")
sta_boxplot_ptca
```


# 11. branching order

```{r}
bra_ord_boxplot <- ggplot(bra_ord_summary, aes(x = sp_code, y = mean)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Branching order",
       x = "Species code",
       y = "Branch count")

bra_ord_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("bra_ord_boxplot-", todays_date, ".jpg", sep = "")),
#        bra_ord_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 12. toughness

```{r}
toughness_boxplot <- ggplot(toughness_summary, aes(x = sp_code, y = mean)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.45) +
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "Toughness",
       x = "Species code",
       y = "kg/cm2")

toughness_boxplot

# ggsave(here::here("figures/basic-visualizations",
#                   paste("toughness_boxplot-", todays_date, ".jpg", sep = "")),
#        toughness_boxplot,
#        width = 8, height = 5, dpi = 150)
```

# 13. correlation between STA and TDMC

```{r}
ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(sp_code != "DL") %>% 
  filter(specimen_ID != "20210721-BULL-023") %>% 
  drop_na(tdmc_mean) %>% 
  ggplot(aes(x = sta_mean, y = tdmc_mean, col = sp_code)) +
  geom_point() +
  geom_smooth(aes(col = sp_code), method = "lm", se = FALSE) +
  geom_smooth(method = "lm", se = TRUE, col = "black") +
  theme_bw() +
  labs(title = "Thallus dry matter content predicted by specific thallus area",
       x = "Specific thallus area (mm2/mg)",
       y = "Thallus dry matter content (dry mg/wet mg)", 
       caption = "Linear mixed model: lme(tdmc ~ sta_mm_mg, random = ~ 1|sp_code)")

sta_tdmc_lmer <- lmer(tdmc_mean ~ sta_mean + (1 + sta_mean|sp_code), data = (ind_traits %>% filter(!(sp_code %in% c("CO"))) %>% filter(sp_code %in% algae_proposal) %>% drop_na(tdmc_mean, sta_mean)))
sta_tdmc_lme <- lme(tdmc_mean ~ sta_mean, random = ~ 1|sp_code, data = (ind_traits %>% filter(!(sp_code %in% c("CO"))) %>% filter(sp_code %in% algae_proposal) %>% drop_na(tdmc_mean, sta_mean)))
summary(sta_tdmc_lme)
summary(sta_tdmc_lmer)
testResiduals(sta_tdmc_lme)
check_model(sta_tdmc_lmer)

sta_tdmc_predict <- ggpredict(sta_tdmc_lmer, terms = ~ sta_mean + (1 + sta_mean|sp_code), type = "random") %>% 
  rename("sp_code" = group)

ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(sp_code != "CO") %>% 
  filter(specimen_ID != "20210721-BULL-023") %>% 
  drop_na(tdmc_mean) %>% 
  ggplot(aes(x = sta_mean, y = tdmc_mean, col = sp_code)) +
  geom_point() +
  geom_line(data = sta_tdmc_predict, aes(x = x, y = predicted, col = sp_code), size = 1) +
  geom_ribbon(data = sta_tdmc_predict, aes(x = x, ymax = conf.high, ymin = conf.low), alpha = 0.1) +
  theme_bw() +
  scale_y_continuous(limits = c(0, 0.5)) +
  labs(title = "Thallus dry matter content predicted by specific thallus area",
       x = "Specific thallus area (mm2/mg)",
       y = "Thallus dry matter content (dry mg/wet mg)", 
       caption = "Linear mixed model: lme(tdmc ~ sta_mm_mg, random = ~ 1|sp_code)")

```

# 14. correlation between FvFm and TDMC

```{r}
fvfm_tdmc_lme <- lme(fvfm_mean ~ tdmc_mean, random = ~ 1|sp_code, data = ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(!(sp_code %in% c("DL"))) %>% 
  drop_na(tdmc_mean, fvfm_mean))
fvfm_tdmc_lmer <- lmer(fvfm_mean ~ tdmc_mean + (1|sp_code), data = ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(!(sp_code %in% c("DL"))) %>% 
  drop_na(tdmc_mean, fvfm_mean))
summary(fvfm_tdmc_lme)
testResiduals(fvfm_tdmc_lmer)
check_model(fvfm_tdmc_lmer)
grid <- ref_grid(fvfm_tdmc_lmer, cov.keep = c("tdmc", "sp_code"))
emm <- emmeans(grid, spec = c("tdmc_mean"), level = 0.95) %>% 
  data.frame(.)

tdmc_fvfm_lm <- lm(tdmc ~ fvfm + sp_code, data = sp_int_matrix %>% filter(!(sp_code %in% c("Nandersoniana", "CO"))) %>% drop_na(tdmc))
summary(tdmc_fvfm_lm)
testResiduals(tdmc_fvfm_lm)


ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(!(sp_code %in% c("DL"))) %>% 
  drop_na(tdmc_mean, fvfm_mean) %>% 
  ggplot() +
  geom_point(aes(x = tdmc_mean, y = fvfm_mean, col = sp_code), size = 2) +
  geom_smooth(aes(x = tdmc_mean, y = fvfm_mean, col = sp_code), method = "lm", se = FALSE) +
  geom_line(data = emm, aes(x = tdmc_mean, y = emmean)) +
  geom_ribbon(data = emm, aes(x = tdmc_mean, y = emmean, ymin = lower.CL, ymax = upper.CL), alpha = 0.2) +
  # geom_smooth(data = sp_int_matrix %>% filter(!(sp_code %in% c("Nandersoniana", "CO"))) %>% drop_na(tdmc), aes(x = tdmc, y = fvfm), method = "lm", se = TRUE, col = "black") +
  theme_bw() +
  labs(title = "Thallus dry matter content as predictor of photosynthetic efficiency",
       caption = "Linear model: lme(tdmc ~ fvfm, random = ~ 1|sp_code) \n
                  significant intercept (0.19, p < 0.001), non-significant slope (-0.15, p = 0.36)")
  


```

# 15. correlation between STA and FvFM

```{r}
fvfm_sta_df <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(sp_code != "DL") %>% 
  filter(specimen_ID != "20210721-BULL-023")

fvfm_sta_lme <- lme(fvfm_mean ~ sta_mean, random = ~ 1|sp_code, data = fvfm_sta_df, na.action = na.omit)
fvfm_sta_lmer <- lmer(fvfm_mean ~ sta_mean + (1 + sta_mean|sp_code), data = fvfm_sta_df, na.action = na.omit)
summary(fvfm_sta_lmer)
r.squaredGLMM(fvfm_sta_lmer)
testResiduals(fvfm_sta_lmer)
check_model(fvfm_sta_lmer)
predicted <- ggpredict(fvfm_sta_lmer, terms = c("sta_mean", "sp_code"), type = "fixed")
grid <- ref_grid(fvfm_sta_lmer, cov.keep = c("sta_mean", "sp_code"))
emm <- emmeans(grid, spec = c("sta_mean"), level = 0.95) %>% 
  data.frame(.)


ggplot(data = fvfm_sta_df, aes(x = sta_mean, y = fvfm_mean)) +
  geom_point(aes(color = sp_code)) +
  geom_smooth(aes(color = sp_code), method = "lm", se = FALSE) +
  geom_line(data = emm, aes(x = sta_mean, y = emmean)) +
  geom_ribbon(data = emm, aes(x = sta_mean, y = emmean,
                              ymin = lower.CL, ymax = upper.CL),
              alpha = 0.3) +
  theme_bw() +
  labs(title = "Specific thallus area as predictor of photosynthetic efficiency",
       caption = "Linear model: lme(sta ~ fvfm, random = ~ 1|sp_code) \n
                  significant intercept (0.51, p < 0.001), significant slope (0.001131, p = 0.04)")




ggplot() +
  geom_point(data = sp_int_matrix %>% filter(!(sp_code %in% c("Nandersoniana", "CO"))) %>% drop_na(tdmc), aes(x = sta_mm_mg, y = fvfm, col = sp_code, shape = sp_code), size = 2) +
  geom_smooth(data = sp_int_matrix %>% filter(!(sp_code %in% c("Nandersoniana", "CO"))) %>% drop_na(tdmc), aes(x = sta_mm_mg, y = fvfm, col = sp_code), method = "lm", se = FALSE) +
  geom_line(data = emm, aes(x = sta_mm_mg, y = emmean)) +
  geom_ribbon(data = emm, aes(x = sta_mm_mg, y = emmean, ymin = lower.CL, ymax = upper.CL), alpha = 0.2) +
  # geom_smooth(data = sp_int_matrix %>% filter(!(sp_code %in% c("Nandersoniana", "CO"))) %>% drop_na(tdmc), aes(x = tdmc, y = fvfm), method = "lm", se = TRUE, col = "black") +
  theme_bw() +
  labs(title = "Specific thallus area as predictor of photosynthetic efficiency",
       caption = "Linear model: lme(sta ~ fvfm, random = ~ 1|sp_code) \n
                  significant intercept (0.51, p < 0.001), non-significant slope (0, p = 0.66)")
```

# 16. PTCA

```{r}
ptca_all <- ptca_leaf %>% 
  left_join(., volume_ind, by = "specimen_ID") %>% 
  left_join(., fvfm_ind, by = "specimen_ID") %>% 
  left_join(., ind_height, by = "specimen_ID") %>% 
  dplyr::select(lifestage, specimen_ID, sta_mean, tdmc_mean, total_volume, fvfm_mean, frond_length_mean, frond_width_mean, maximum_height) %>% 
  rename("FvFm" = fvfm_mean,
         "STA" = sta_mean,
         "TDMC" = tdmc_mean,
         "Volume" = total_volume,
         "Frond length" = frond_length_mean,
         "Frond width" = frond_width_mean,
         "Maximum height" = maximum_height) %>% 
  pivot_longer(cols = 3:9, names_to = "trait", values_to = "value") %>% 
  mutate(expand = case_when(
    value = TRUE ~ value*1.2
  )) %>% 
  group_by(trait) %>% 
  mutate(bracket = case_when(
    value = TRUE ~ max(value, na.rm = TRUE)*1.1
  )) %>% 
  ungroup()

ggplot(ptca_all, aes(x = lifestage, y = value)) +
  geom_boxplot(aes(fill = lifestage)) +
  scale_fill_manual(values = c("#70601d", "#b59b33")) +
    geom_blank(aes(y = expand)) +
  geom_signif(comparisons = list(c("adult", "recruit")), 
              map_signif_level = TRUE,
              test = "t.test",
              aes(y_position = bracket)) +
  geom_jitter() +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        axis.text = element_text(size = 14),
        axis.title = element_text(size = 16),
        strip.text = element_text(size = 16)) +
  facet_wrap(~trait, scales = "free_y", nrow = 2)

ptca_all_mat <- ptca_leaf %>% 
  left_join(., volume_ind, by = "specimen_ID") %>% 
  left_join(., fvfm_ind, by = "specimen_ID") %>% 
  left_join(., ind_height, by = "specimen_ID") %>% 
  dplyr::select(specimen_ID, sta_mean, tdmc_mean, total_volume, fvfm_mean, frond_length_mean, frond_width_mean, maximum_height) %>% 
  rename("FvFm" = fvfm_mean,
         "STA" = sta_mean,
         "TDMC" = tdmc_mean,
         "Volume" = total_volume,
         "Frond length" = frond_length_mean,
         "Frond width" = frond_width_mean,
         "Maximum height" = maximum_height) %>% 
  column_to_rownames("specimen_ID")

```


# 17. chlorophyll A


```{r}
chlA_plot <- ind_traits %>% 
  drop_na(chlA_mean) %>% 
  filter(!(date_collected %in% as_date(c("2023-06-27", "2023-06-20"))))

ggplot(data = chlA_plot, aes(x = sp_code, y = chlA_mean)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(aes(color = as_factor(year))) +
  facet_wrap(~pigment_type, scales = "free_x")

ggplot(data = chlA_plot, aes(x = as_factor(year), y = chlA_mean)) +
  geom_boxplot(outlier.alpha = 0) +
  geom_jitter(aes(color = as_factor(year))) +
  facet_wrap(~sp_code, scales = "free")

ggplot(data = chlA_plot, aes(x = chlA_mean, y = fvfm_mean)) +
  geom_point(aes(color = sp_code)) +
  facet_wrap(~pigment_type, scales = "free")

model <- lm(fvfm_mean ~ chlA_mean*sp_code, data = chlA_plot)
simulateResiduals(model, plot = TRUE)
summary(model)
plot(ggeffect(model, terms = "chlA_mean"))

model <- glm(chlA_mean ~ as_factor(year)*sp_code, data = chlA_plot)
simulateResiduals(model, plot = TRUE)
summary(model)
```

# 18. traits for Katrina

Chondracanthus, Rhodymenia, Macrocystis, Pterygophora, and Stephanocystis

```{r}
traits_df <- ind_traits %>% 
  filter(sp_code %in% c("CC", "R", "PTCA", "CYOS")) %>% 
  group_by(sp_code) %>% 
  summarize(sap_mm2_mm = mean(sap_mean, na.rm = TRUE),
            sap_se = se(sap_mean, na.rm = TRUE),
            # SAP sample size
            sap_n = sum(!is.na(sap_mean)),
            chlA_ug_mg = mean(chlA_mean, na.rm = TRUE),
            chlA_se = se(chlA_mean, na.rm = TRUE),
            # chlA sample size
            chlA_n = sum(!is.na(chlA_mean)))

write_csv(traits_df, file = here::here("data", "km_traits.csv"))
```



```{r include = FALSE}
# Use `rmarkdown::render(here::here("code", "02_basic-visualizations.Rmd"))` to knit the document.
```

# 19. isotopes

```{r}
delta_15_n_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = delta_15_n_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "delta 15 N",
       x = "Species",
       y = "delta 15 N") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
delta_15_n_boxplot

delta_13_c_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = delta_13_c_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "delta 13 C",
       x = "Species",
       y = "delta 13 C") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
delta_13_c_boxplot

weight_percent_n_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = weight_percent_n_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "% N",
       x = "Species",
       y = "% N") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
weight_percent_n_boxplot

weight_percent_c_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = weight_percent_c_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "% C",
       x = "Species",
       y = "% C") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
weight_percent_c_boxplot
```

# 20. mass to height

```{r}
mh_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = mass_to_height, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "mass to height ratio",
       x = "Species",
       y = "mass/height (dry mg/cm)") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
mh_boxplot
```


# 21. aspect ratio

```{r}
aspect_ratio_boxplot <- ind_traits %>% 
  group_by(sp_code) %>% 
  filter(sp_code %in% algae_proposal) %>% 
  ggplot(aes(x = sp_code, y = aspect_ratio_mean, group = sp_code)) +
  geom_boxplot(aes(fill = sp_code)) +
  scale_fill_manual(values = algae_spcode_colors) +
  geom_jitter() +
  # geom_text(aes(label = value, y = 0.83), nudge_y = 0.01) + 
  facet_wrap(~ pigment_type, scales = "free_x") +
  scale_x_discrete(guide = guide_axis(n.dodge = 2)) +
  theme_bw() +
  labs(title = "aspect ratio (length/width)",
       x = "Species",
       y = "aspect ratio") +
  theme(legend.position = "none",
        strip.background = element_blank(),
        strip.text = element_text(size = 20),
        axis.text = element_text(size = 20),
        axis.title = element_text(size = 20),
        plot.title = element_text(size = 20))
aspect_ratio_boxplot
```

# 22. distributions

```{r}
distributions <- ind_traits %>% 
  select(specimen_ID, scientific_name,
         maximum_height, mass_to_height, sav_mean, thickness_mm_mean, tdmc_mean, sta_mean, sav_mean, sap_mean, fvfm_mean, aspect_ratio_mean, frond_length_mean, frond_width_mean, total_wet, total_dry) %>% 
  pivot_longer(maximum_height:total_dry,
               names_to = "trait",
               values_to = "value") %>% 
  nest(.by = "trait",
       data = everything()) %>% 
  mutate(length = map(
    data,
    ~ .x %>% 
      drop_na(value) %>% 
      nrow(.)
  )) %>% 
  mutate(bins = map(
    length,
    ~ round((.x^(1/3))*2)
  )) %>% 
  mutate(distribution_plot = pmap(
    list(x = data, y = trait, z = bins),
    function(x, y, z) ggplot(data = x,
                             aes(x = value)) +
      geom_histogram(bins = z) +
      labs(title = paste0(y, " (no transformation)"))
  )) %>% 
  mutate(distribution_plot_log = pmap(
    list(x = data, y = trait, z = bins),
    function(x, y, z) ggplot(data = x,
                             aes(x = log(value))) +
      geom_histogram(bins = z) +
      labs(title = paste0(y, " (natural log transform)"))
  )) %>% 
  mutate(distribution_plot_sqrt = pmap(
    list(x = data, y = trait, z = bins),
    function(x, y, z) ggplot(data = x,
                             aes(x = sqrt(value))) +
      geom_histogram(bins = z) +
      labs(title = paste0(y, " (square root transform)"))
  )) %>% 
  mutate(qq = pmap(
    list(x = data, y = trait),
    function(x, y) ggplot(data = x,
                          aes(sample = value)) +
      geom_qq() +
      geom_qq_line() +
      labs(title = paste0(y, " (no transformation)"))
  )) %>% 
  mutate(qq_log = pmap(
    list(x = data, y = trait),
    function(x, y) ggplot(data = x,
                          aes(sample = log(value))) +
      geom_qq() +
      geom_qq_line() +
      labs(title = paste0(y, " (natural log transform)"))
  )) %>% 
  mutate(qq_sqrt = pmap(
    list(x = data, y = trait),
    function(x, y) ggplot(data = x,
                          aes(sample = log(value))) +
      geom_qq() +
      geom_qq_line() +
      labs(title = paste0(y, " (square root transform)"))
  ))

# maximum height
pluck(distributions, 5, 1) # right skewed
pluck(distributions, 8, 1)
pluck(distributions, 6, 1) # log transform looks normal (ish)
pluck(distributions, 9, 1)
pluck(distributions, 7, 1)
pluck(distributions, 10, 1)

# mass to height
pluck(distributions, 5, 2) # right skewed
pluck(distributions, 8, 2)
pluck(distributions, 6, 2) # log transform looks normal
pluck(distributions, 9, 2)
pluck(distributions, 7, 2)
pluck(distributions, 10, 2)

# surface area to volume
pluck(distributions, 5, 3) # normal ish
pluck(distributions, 8, 3)
pluck(distributions, 6, 3) # log transform looks normal
pluck(distributions, 9, 3)
pluck(distributions, 7, 3)
pluck(distributions, 10, 3)

# thickness
pluck(distributions, 5, 4) # right skewed
pluck(distributions, 8, 4)
pluck(distributions, 6, 4) # log transform looks normal
pluck(distributions, 9, 4)
pluck(distributions, 7, 4)
pluck(distributions, 10, 4)

# TDMC
pluck(distributions, 5, 5) # potentially bimodal
pluck(distributions, 8, 5)
pluck(distributions, 6, 5) 
pluck(distributions, 9, 5)
pluck(distributions, 7, 5)
pluck(distributions, 10, 5)
# no transformations look good

# STA
pluck(distributions, 5, 6) # right skewed
pluck(distributions, 8, 6)
pluck(distributions, 6, 6) 
pluck(distributions, 9, 6) # log transform is better?
pluck(distributions, 7, 6)
pluck(distributions, 10, 6)

# SA:P
pluck(distributions, 5, 7) # right skewed
pluck(distributions, 8, 7)
pluck(distributions, 6, 7) 
pluck(distributions, 9, 7) # log transform is better?
pluck(distributions, 7, 7)
pluck(distributions, 10, 7)

# FvFm
pluck(distributions, 5, 8) # bimodal
pluck(distributions, 8, 8)
pluck(distributions, 6, 8) 
pluck(distributions, 9, 8)
pluck(distributions, 7, 8)
pluck(distributions, 10, 8)
# no transformations look good

# aspect ratio
pluck(distributions, 5, 9) # right skewed
pluck(distributions, 8, 9)
pluck(distributions, 6, 9) 
pluck(distributions, 9, 9) # log transformation looks better?
pluck(distributions, 7, 9)
pluck(distributions, 10, 9)

# frond length
pluck(distributions, 5, 10) # right skewed
pluck(distributions, 8, 10)
pluck(distributions, 6, 10) 
pluck(distributions, 9, 10) # log transformation looks better
pluck(distributions, 7, 10)
pluck(distributions, 10, 10)

# frond width
pluck(distributions, 5, 11) # right skewed
pluck(distributions, 8, 11)
pluck(distributions, 6, 11) 
pluck(distributions, 9, 11)
pluck(distributions, 7, 11)
pluck(distributions, 10, 11)
# no transformations look good

# total wet mass
pluck(distributions, 5, 12) # right skewed
pluck(distributions, 8, 12)
pluck(distributions, 6, 12) # log transform looks normal
pluck(distributions, 9, 12)
pluck(distributions, 7, 12)
pluck(distributions, 10, 12)

# total dry mass
pluck(distributions, 5, 13) # right skewed
pluck(distributions, 8, 13)
pluck(distributions, 6, 13) # log transform looks normal
pluck(distributions, 9, 13)
pluck(distributions, 7, 13)
pluck(distributions, 10, 13)

# traits to transform (potentially)
# max height, mass to height, SAV, thickness, STA, SAP, aspect ratio, 
# frond length, wet mass, dry mass


```
