==> code/old-code/03_prelim-ordinations.Rmd <==
---
title: "ordinations"
author: "An Bui"
date: "10/25/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
# source the cleaning/summarizing script, which sources the set up
source(here::here("code", "01a_trait-cleaning.R"))
```

# 0. Orientation

Necessary data frames are `tbs_matrix` created in the set up script.     

# 1. trait PCoA

```{r}
gower_mat <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # filter(!(sp_code %in% c("CO"))) %>% 
  filter(!(specimen_ID %in% recruits)) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID == "20210721-BULL-023")) %>% 
  # take out standard error
  dplyr::select(!ends_with("_se")) %>% 
  # take out irrelevant columns
  dplyr::select(-c(sp_code, scientific_name, life_habit, posture, branching_yn, lifestage, method, year)) %>% 
  # take out date selected and site
  dplyr::select(-c(date_collected, site)) %>% 
  column_to_rownames("specimen_ID") %>% 
  # convert categorical traits into factors
  mutate(growth_form = as_factor(growth_form),
         pigment_type = as_factor(pigment_type),
         longevity = as_factor(longevity))

# note: PCOA is based on distance matrices, so never actually sees "species"
# add species by calculating weighted averages: from original data matrix of traits by species, average abundance and axis scores for all sites
tbs_gower <- FD::gowdis(gower_mat)
tbs_pcoa <- cmdscale(tbs_gower, eig = TRUE)
ordiplot(tbs_pcoa, display = "sites", type = "text")

tbs_gower_daisy <- daisy(gower_mat, metric = "gower") # same as above
tbs_pcoa_daisy <- cmdscale(tbs_gower_daisy, eig = TRUE)
ordiplot(tbs_pcoa_daisy, display = "sites", type = "text")
biplot.pcoa(ape::pcoa(tbs_gower_daisy))
test_pcoa <- ape::pcoa(tbs_gower_daisy, correction = "lingoes") 
test_pcoa

# tbs_pcoa <- wcmdscale(tbs_gower, eig = TRUE)
# tbs_pcoa <- ape::pcoa(tbs_gower, correction = "none")
# biplot(ape::pcoa(tbs_gower, correction = "none"))
# ordiplot(scores(tbs_pcoa, choices=c(1,2)), type="none", main="PCoA with species weighted averages")
# tbs.wa <- wascores(tbs_pcoa$points[,1:2], tbs_matrix)
# text(tbs.wa, rownames(tbs.wa), cex=0.7, col="red")
# biplot.pcoa(tbs_gower, tbs_pcoa, dir.axis1 = -1)
# tbs_pcoa$species <- wascores(x = tbs_pcoa$points, , expand = TRUE)
p1 <- ordiplot(tbs_pcoa, display = "sites", type = "none")
points1 <- points(p1, "sites", pch = 21, col = "red", bg = "yellow")
text(p1, "sites", col = "blue", cex = 0.9)
# identify(p1, "sites")



pcoa_spp_scores <- scores(tbs_pcoa, display = "sites") %>% 
  as.data.frame() %>% 
  rownames_to_column("specimen_ID") %>% 
  left_join(., metadata_ind, by = "specimen_ID")

pcoa_trait_scores <- scores(tbs_pcoa, choices = c(1,2)) %>% 
  as.data.frame()
ggplot(pcoa_spp_scores, aes(x = Dim1, y = Dim2, color = sp_code, shape = site)) +
  geom_hline(yintercept = 0, color = "grey", lty = 3) +
  geom_vline(xintercept = 0, color = "grey", lty = 3) +
  geom_point(alpha = 0.8, size = 3) +
  theme_bw()
```

# 2. PCA

```{r}
# sp_int_matrix <- ind_traits %>% 
#   filter(sp_code %in% algae_proposal) %>% 
#   filter(!(sp_code %in% c("CO"))) %>% 
#   # weird Nienburgia?
#   filter(!(specimen_ID == "20210721-BULL-023"))

pca_mat <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # filter(!(sp_code %in% c("CO"))) %>% 
  # filter(!(sp_code == "PTCA" & lifestage == "recruit")) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID %in% c("20210721-BULL-023", "20210719-IVEE-009"))) %>% 
  filter(sp_code != "EGME") %>% 
  mutate(growth_form_num = case_when(
    growth_form == "leathery_macrophyte" ~ 1,
    growth_form == "corticated_macrophytes" ~ 2,
    growth_form == "corticated_foliose" ~ 3
  ), 
  pigment_type_num = case_when(
    pigment_type == "red" ~ 1,
    pigment_type == "brown" ~ 2
  ), 
  longevity_num = case_when(
    longevity == "annual" ~ 1,
    longevity == "perennial" ~ 3,
    longevity == "annual or perennial" ~ 2
  )) %>% 
  dplyr::select(specimen_ID, maximum_height, mass_to_height, sav_mean, thickness_mm_mean, tdmc_mean, sta_mean, sav_mean, sap_mean, fvfm_mean, aspect_ratio_mean) %>% 
  column_to_rownames("specimen_ID") %>% 
  drop_na() %>% 
  rename(`Maximum height` = maximum_height,
         `Mass:height` = mass_to_height,
         `SA:V` = sav_mean,
         `Thickness` = thickness_mm_mean,
         `TDMC` = tdmc_mean,
         `STA` = sta_mean,
         `SA:V` = sav_mean,
         `SA:P` = sap_mean,
         `Fv/Fm` = fvfm_mean,
         `Aspect ratio` = aspect_ratio_mean)

pca_mat_scale <- scale(pca_mat)

pca_mat_log <- pca_mat %>% 
  mutate(across(c(`Maximum height`, `Mass:height`, `SA:V`,
                `Thickness`, `STA`, `SA:P`, `Aspect ratio`), log))

pca_mat_with_cn <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # filter(!(sp_code %in% c("CO"))) %>% 
  # filter(!(sp_code == "PTCA" & lifestage == "recruit")) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID %in% c("20210721-BULL-023", "20210719-IVEE-009"))) %>% 
  mutate(growth_form_num = case_when(
    growth_form == "leathery_macrophyte" ~ 1,
    growth_form == "corticated_macrophytes" ~ 2,
    growth_form == "corticated_foliose" ~ 3
  ), 
  pigment_type_num = case_when(
    pigment_type == "red" ~ 1,
    pigment_type == "brown" ~ 2
  ), 
  longevity_num = case_when(
    longevity == "annual" ~ 1,
    longevity == "perennial" ~ 3,
    longevity == "annual or perennial" ~ 2
  )) %>% 
  dplyr::select(specimen_ID, maximum_height, total_volume, sav_mean, thickness_mm_mean, tdmc_mean, sta_mean, sav_mean, sap_mean, fvfm_mean, thickness_mm_mean, weight_percent_n_mean, weight_percent_c_mean) %>% 
  column_to_rownames("specimen_ID") %>% 
  drop_na()

# , growth_form_num, pigment_type_num, longevity_num 
# frond_length_mean, frond_width_mean
```

Got a lot of help from David Zeleny's [blog post](https://www.davidzeleny.net/anadat-r/doku.php/en:pca_examples).

```{r}
# tbs matrix filtered for species in proposal
# doesn't include Nienburgia or Corallina
# sp_int_matrix <- tbs_matrix %>% 
#   filter(sp_code %in% algae_proposal) %>% 
#   filter(!(sp_code %in% c("CO", "Nandersoniana")))
# 
# filtered_matrix <- tbs_matrix %>% 
#   filter(!(sp_code %in% c("MAPY", "CO", "Nandersoniana")))

# matrix_choice <- pca_mat

# only include numeric variables
# tbs_pca <- rda(matrix_choice[, 2:10], scale = TRUE)
tbs_pca <- rda(pca_mat_log)
screeplot(tbs_pca, bstick = TRUE)

# extract proportion explained for each axis
summary(tbs_pca)
loadings <- scores(tbs_pca, display = 'species', scaling = 0, choices = c(1, 2, 3))

loadings_df <- loadings %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("trait") %>% 
  arrange(PC1) %>% 
  mutate(trait = fct_inorder(trait))

loadings_plot_aes <- list(
  geom_col(color = "#000000", fill = "orange", linewidth = 0.5),
  geom_vline(xintercept = 0),
  scale_x_continuous(limits = c(-0.89, 1), breaks = seq(from = -0.75, to = 1, by = 0.25))
)

loadings_plot_theme <- function() {
  theme_bw() +
    theme(panel.grid = element_blank(),
          axis.text = element_text(size = 20),
          axis.title = element_blank(),
          # plot.title.position = "plot",
          plot.title = element_text(hjust = 0.5, size = 20))
}

pc1_plot <- ggplot(data = loadings_df, 
       aes(x = PC1,
           y = trait)) +
  loadings_plot_aes +
  loadings_plot_theme() +
  labs(title = "PC1")

pc2_plot <- ggplot(data = loadings_df, 
       aes(x = PC2,
           y = trait)) +
  loadings_plot_aes +
  loadings_plot_theme() +
  labs(title = "PC2")

pc3_plot <- ggplot(data = loadings_df, 
       aes(x = PC3,
           y = trait)) +
  loadings_plot_aes +
  loadings_plot_theme() +
  labs(title = "PC3")

loadings_plot <- pc1_plot / pc2_plot / pc3_plot

ggsave(here::here(
  "figures",
  "prelim-ordinations",
  paste0("loadings_", today(), ".jpg")),
  loadings_plot,
  width = 14,
  height = 14,
  units = "cm",
  dpi = 300
)

# simple biplot (compare with ggplot output to make sure it's right)
biplot(tbs_pca)

# species points
PCAscores <- scores(tbs_pca, display = "sites", choices = c(1, 2)) %>% 
  as.data.frame() %>% 
  rownames_to_column("specimen_ID") %>% 
  left_join(metadata_ind, by = "specimen_ID") %>% 
  left_join(., coarse_traits, by = "sp_code") %>% 
  mutate(scientific_name = factor(scientific_name, levels = algae_proposal_sciname_factors))

PCAscores_13 <- scores(tbs_pca, display = "sites", choices = c(1, 3)) %>% 
  as.data.frame() %>% 
  rownames_to_column("specimen_ID") %>% 
  left_join(metadata_ind, by = "specimen_ID") %>% 
  left_join(., coarse_traits, by = "sp_code") %>% 
  mutate(scientific_name = factor(scientific_name, levels = algae_proposal_sciname_factors))

# trait vectors
PCAvect <- scores(tbs_pca, display = "species", choices = c(1, 2)) %>% 
  as.data.frame()

PCAvect_13 <- scores(tbs_pca, display = "species", choices = c(1, 3)) %>% 
  as.data.frame()

# plot PCA
plot_PCA_12 <- ggplot() +
  coord_cartesian() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_point(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name), size = 1) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  scale_color_manual(values = algae_colors) +
  geom_segment(data = PCAvect, aes(x = 0, y = 0, xend = PC1, yend = PC2), arrow = arrow(length = unit(0.2, "cm")), linewidth = 0.5) +
  geom_text_repel(data = PCAvect, aes(x = PC1, y = PC2, label = rownames(PCAvect)), size = 8, alpha = 0.8,
            color = "red") +
  # scale_x_continuous(limits = c(-1.75, 3.75)) +
  # scale_y_continuous(limits = c(-2, 3.25)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 4),
         shape = guide_legend(ncol = 4)) +
  theme(legend.position = "bottom", 
        legend.box = "vertical", 
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 10)) +
  labs(x = "PC1 (37.6%)",
       y = "PC2 (30.4%)",
       title = "PC1 and PC2",
       color = "Scientific name") 
plot_PCA_12

plot_PCA_13 <- ggplot() +
  coord_cartesian() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_point(data = PCAscores_13, aes(x = PC1, y = PC3, color = scientific_name), size = 1) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  scale_color_manual(values = algae_colors) +

  geom_segment(data = PCAvect_13, aes(x = 0, y = 0, xend = PC1, yend = PC3), arrow = arrow(length = unit(0.2, "cm")), linewidth = 0.5) +
  geom_text_repel(data = PCAvect_13, aes(x = PC1, y = PC3, label = rownames(PCAvect)), size = 8, alpha = 0.8,
            color = "red") +
  # scale_x_continuous(limits = c(-1.5, 2.3)) +
  # scale_y_continuous(limits = c(-3, 2)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 4),
         shape = guide_legend(ncol = 4)) +
  theme(legend.position = "bottom", 
        legend.box = "vertical", 
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 10)) +
  labs(x = "PC1 (37.6%)",
       y = "PC3 (18.5%)",
       title = "PC1 and PC3",
       color = "Scientific name") 
plot_PCA_13

pca_together <- (plot_PCA_12 | plot_PCA_13) +
  plot_layout(guides = "collect", ncol = 2, widths = c(1, 1)) &
  theme(legend.position = "bottom")
pca_together

# get loading contributions
loadings <- scores(tbs_pca, display = 'species', scaling = 0)
sort(abs(loadings[,1]), decreasing = TRUE)
sort(abs(loadings[,2]), decreasing = TRUE)

ggsave(here::here("figures", "prelim-ordinations",
                  paste("PCA_together_cont-", today(), ".jpg", sep = "")),
       pca_together,
       width = 18, height = 12, units = "cm", dpi = 300)
# 
# ggsave(here::here("figures", "prelim-ordinations",
#                   paste("PCA_spp-all-12-", today(), ".jpg", sep = "")),
#        plot_PCA_12,
#        width = 6, height = 6, dpi = 300)
# 
# ggsave(here::here("figures", "prelim-ordinations",
#                   paste("PCA_spp-all-13-", today(), ".jpg", sep = "")),
#        plot_PCA_13,
#        width = 6, height = 6, dpi = 300)
```

# 3. Corrplot

## not scaled

```{r}
M <- pca_mat %>% 
  cor(method = "pearson")

p <- cor.mtest(pca_mat,
          conf.level = 0.95,
          method = "pearson")

corr_traits <- corrplot(corr = M, 
                        p.mat = p$p,
                        diag = FALSE,
                        method = "circle", 
                        # addgrid.col = NA,
                        type = "lower", 
                        insig = "blank", 
                        addCoef.col = "black", 
                        col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
                        sig.level = 0.05)

jpeg(here::here("figures", 
               "prelim-ordinations",
               paste0("corrplot-traits_", today(), ".jpeg")),
     width = 14, height = 14, units = "cm", res = 300)
corrplot(corr = M, 
         p.mat = p$p,
         diag = FALSE,
         method = "circle", 
         # addgrid.col = NA,
         type = "lower", 
         insig = "blank", 
         addCoef.col = "black", 
         col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
         sig.level = 0.05)
dev.off()
```

## scaled

```{r}
M <- pca_mat_scaled %>% 
  cor(method = "pearson")

p <- cor.mtest(pca_mat_scaled,
          conf.level = 0.95,
          method = "pearson")

corr_traits <- corrplot(corr = M, 
                        p.mat = p$p,
                        diag = FALSE,
                        method = "circle", 
                        # addgrid.col = NA,
                        type = "lower", 
                        insig = "blank", 
                        addCoef.col = "black", 
                        col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
                        sig.level = 0.05)

jpeg(here::here("figures", 
               "prelim-ordinations",
               paste0("corrplot-traits_scaled_", today(), ".jpeg")),
     width = 14, height = 14, units = "cm", res = 300)
corrplot(corr = M, 
         p.mat = p$p,
         diag = FALSE,
         method = "circle", 
         # addgrid.col = NA,
         type = "lower", 
         insig = "blank", 
         addCoef.col = "black", 
         col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
         sig.level = 0.05)
dev.off()
```

## log transformed

natural log

```{r}
M <- pca_mat_log %>% 
  cor(method = "pearson")

p <- cor.mtest(pca_mat_log,
          conf.level = 0.95,
          method = "pearson")

corr_traits <- corrplot(corr = M, 
                        p.mat = p$p,
                        diag = FALSE,
                        method = "circle", 
                        # addgrid.col = NA,
                        type = "lower", 
                        insig = "blank", 
                        addCoef.col = "black", 
                        col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
                        sig.level = 0.05)

jpeg(here::here("figures", 
               "prelim-ordinations",
               paste0("corrplot-traits_log_", today(), ".jpeg")),
     width = 14, height = 14, units = "cm", res = 300)
corrplot(corr = M, 
         p.mat = p$p,
         diag = FALSE,
         method = "circle", 
         # addgrid.col = NA,
         type = "lower", 
         insig = "blank", 
         addCoef.col = "black", 
         col= colorRampPalette(c("#F21A00", "#FFFFFF", "#3B9AB2"))(200), 
         sig.level = 0.05)
dev.off()
```

# 4. "Tradeoffs"

## pairs plot

```{r}
pairs_plot <- GGally::ggpairs(pca_mat_log, 
                      upper = list(continuous = wrap("cor", method = "spearman")))
pairs_plot
```

potentially interesting pairs:  

- SAV and thickness (-0.762)  
- thickness and N (-0.749)  
- thickness and C (-0.588)   
- STA and C (0.597)

## summarizing

```{r}
tradeoff_df <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # filter(!(sp_code %in% c("CO"))) %>% 
  # filter(!(sp_code == "PTCA" & lifestage == "recruit")) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID %in% c("20210721-BULL-023", "20210719-IVEE-009", "20210719-IVEE-010")))

tradeoff_summ <- tradeoff_df %>% 
  group_by(scientific_name) %>% 
  summarize(mean_thickness = mean(thickness_mm_mean, na.rm = TRUE),
            sd_thickness = sd(thickness_mm_mean, na.rm = TRUE), 
            mean_sav = mean(sav_mean, na.rm = TRUE),
            sd_sav = sd(sav_mean, na.rm = TRUE),
            mean_sap = mean(sap_mean, na.rm = TRUE),
            sd_sap = mean(sap_mean, na.rm = TRUE),
            mean_vol = mean(total_volume, na.rm = TRUE),
            sd_vol = sd(total_volume, na.rm = TRUE),
            mean_tdmc = mean(tdmc_mean, na.rm = TRUE),
            sd_tdmc = sd(tdmc_mean, na.rm = TRUE),
            mean_fvfm = mean(fvfm_mean, na.rm = TRUE),
            sd_fvfm = sd(fvfm_mean, na.rm = TRUE),
            mean_c = mean(weight_percent_c_mean, na.rm = TRUE),
            sd_c = sd(weight_percent_c_mean, na.rm = TRUE),
            mean_n = mean(weight_percent_n_mean, na.rm = TRUE),
            sd_n = sd(weight_percent_n_mean, na.rm = TRUE),
            mean_mh = mean(mass_to_height, na.rm = TRUE),
            sd_mh = sd(mass_to_height, na.rm = TRUE),
            mean_mh = mean(maximum_height, na.rm = TRUE),
            sd_mh = sd(maximum_height, na.rm = TRUE),
            mean_sta = mean(sta_mean, na.rm = TRUE),
            sd_sta = sd(sta_mean, na.rm = TRUE))
```


## a. SAV and thickness

```{r}
thick_sav <- ggplot(tradeoff_summ,
                    aes(x = mean_thickness,
                        y = mean_sav)) +
  geom_errorbarh(aes(xmin = mean_thickness - sd_thickness,
                     xmax = mean_thickness + sd_thickness),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_sav - sd_sav,
                    ymax = mean_sav + sd_sav),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits,
             aes(x = thickness_mm_mean,
                 y = sav_mean,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 1, y = 7500, label = "Spearman ρ = -0.72", size = 10) +
  labs(x = "Mean thickness (mm)",
       y = "Mean surface area:volume ratio (mm \U00B2/mL") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

thick_sav

cor.test(ind_traits$thickness_mm_mean, ind_traits$sav_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("thickness-sav_", today(), ".jpg", sep = "")),
#        thick_sav,
#        width = 8, height = 8, dpi = 150)

```

## b. SAV and SAP

```{r}
sap_sav <- ggplot(tradeoff_summ,
                  aes(x = mean_sap,
                      y = mean_sav)) +
  geom_errorbarh(aes(xmin = mean_sap - sd_sap,
                     xmax = mean_sap + sd_sap),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_sav - sd_sav,
                    ymax = mean_sav + sd_sav),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits,
             aes(x = sap_mean,
                 y = sav_mean,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 60, y = 7500, label = "Spearman ρ = -0.194", size = 10) +
  labs(x = "Mean surface area:perimeter ratio (mm)",
       y = "Mean surface area:volume ratio (mm \U00B2/mL") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

sap_sav

cor.test(ind_traits$sap_mean, ind_traits$sav_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("sap-sav_", today(), ".jpg", sep = "")),
#        sap_sav,
#        width = 8, height = 8, dpi = 150)
```

## c. SAP and total volume

Pearson's r = 0.409

```{r}
sap_vol <- ggplot(tradeoff_summ,
                  aes(x = mean_sap,
                      y = mean_vol)) +
  geom_errorbarh(aes(xmin = mean_sap - sd_sap,
                     xmax = mean_sap + sd_sap),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_vol - sd_vol,
                    ymax = mean_vol + sd_vol),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits %>% filter(specimen_ID != "20220824-MOHK-001"),
             aes(x = sap_mean,
                 y = total_volume,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 60, y = 1000, label = "Spearman ρ = 0.48", size = 10) +
  labs(x = "Mean surface area:perimeter ratio (mm)",
       y = "Mean volume (mL)") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

sap_vol

cor.test(ind_traits$sap_mean, ind_traits$total_volume,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("sap-vol_", today(), ".jpg", sep = "")),
#        sap_vol,
#        width = 8, height = 8, dpi = 150)
```

## d. TDMC and fvfm

Pearson's r = -0.446

```{r}
tdmc_fvfm <- ggplot(tradeoff_summ,
                    aes(x = mean_tdmc,
                        y = mean_fvfm)) +
  geom_errorbarh(aes(xmin = mean_tdmc - sd_tdmc,
                     xmax = mean_tdmc + sd_tdmc),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_fvfm - sd_fvfm,
                    ymax = mean_fvfm + sd_fvfm),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits,
             aes(x = tdmc_mean,
                 y = fvfm_mean,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 0.75, y = 0.8, label = "Spearman ρ = -0.42", size = 10) +
  labs(x = "Mean thallus dry matter content (dry mass:fresh mass)",
       y = "Mean photosynthetic efficiency") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

tdmc_fvfm

cor.test(ind_traits$tdmc_mean, ind_traits$fvfm_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("tdmc-fvfm_", today(), ".jpg", sep = "")),
#        tdmc_fvfm,
#        width = 8, height = 8, dpi = 150)
```

## e. N and STA

```{r}
n_sta <- ggplot(tradeoff_summ,
                    aes(x = mean_n,
                        y = mean_sta)) +
  # geom_errorbarh(aes(xmin = mean_n - sd_n,
  #                    xmax = mean_n + sd_n),
  #                alpha = 0.4) +
  # geom_errorbar(aes(ymin = mean_sta - sd_sta,
  #                   ymax = mean_sta + sd_sta),
  #               alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits,
             aes(x = weight_percent_n_mean,
                 y = sta_mean,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  # annotate(geom = "text", x = 20, y = 0.8, label = "Spearman ρ = 0.789", size = 10) +
  labs(x = "Mean %C",
       y = "Mean specific thallus area (mm \U00B2/g)") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

n_sta

cor.test(ind_traits$weight_percent_c_mean, ind_traits$fvfm_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("c-fvfm_", today(), ".jpg", sep = "")),
#        c_fvfm,
#        width = 8, height = 8, dpi = 150)
```

## f. mass:height and tdmc

```{r}
mh_tdmc <- ggplot(tradeoff_summ %>% 
                    filter(scientific_name != "Egregia menziesii"),
                    aes(x = mean_mh,
                        y = mean_tdmc)) +
  geom_errorbarh(aes(xmin = mean_mh - sd_mh,
                     xmax = mean_mh + sd_mh),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_tdmc - sd_tdmc,
                    ymax = mean_tdmc + sd_tdmc),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  # geom_point(data = ind_traits %>% 
  #                   filter(scientific_name != "Egregia menziesii"),
  #            aes(x = mass_to_height,
  #                y = tdmc_mean,
  #                color = scientific_name),
  #            alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 400, y = 0.9, label = "Spearman ρ = 0.77", size = 10) +
  labs(x = "Mean mass:height ratio",
       y = "Mean TDMC") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

mh_tdmc

cor.test(ind_traits$mass_to_vol, ind_traits$tdmc_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("mv-tdmc_", today(), ".jpg", sep = "")),
#        mv_tdmc,
#        width = 8, height = 8, dpi = 150)
```

## g. mass:volume and STA

```{r}
mv_sta <- ggplot(tradeoff_summ,
                    aes(x = mean_mv,
                        y = mean_sta)) +
  geom_errorbarh(aes(xmin = mean_mv - sd_mv,
                     xmax = mean_mv + sd_mv),
                 alpha = 0.4) +
  geom_errorbar(aes(ymin = mean_sta - sd_sta,
                    ymax = mean_sta + sd_sta),
                alpha = 0.4) +
  geom_point(aes(fill = scientific_name),
             size = 4,
             stroke = 1,
             shape = 21) +
  geom_point(data = ind_traits,
             aes(x = mass_to_vol,
                 y = sta_mean,
                 color = scientific_name),
             alpha = 0.2) +
  scale_fill_manual(values = algae_colors) +
  scale_color_manual(values = algae_colors) +
  annotate(geom = "text", x = 1000, y = 700, label = "Spearman ρ = -0.59", size = 10) +
  labs(x = "Mean mass:volume ratio",
       y = "Mean STA") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 18))

mv_sta

cor.test(ind_traits$mass_to_vol, ind_traits$sta_mean,
         method = "spearman")

# ggsave(here::here("figures", "tradeoffs",
#                   paste("mv-sta_", today(), ".jpg", sep = "")),
#        mv_sta,
#        width = 8, height = 8, dpi = 150)
```

## h. SMA

```{r}
library(lmodel2)
library(smatr)
library(ggpmisc)
library(ggExtra)

ind_traits_filtered <- ind_traits %>% 
  filter(scientific_name != "Egregia menziesii")

mh_tdmc <- lmodel2(mass_to_height ~ tdmc_mean,
    data = ind_traits_filtered)

ggMarginal(
  ggplot(data = ind_traits_filtered,
       aes(x = tdmc_mean,
           y = mass_to_height)) +
  geom_point() +
  stat_ma_line(method = "SMA") +
  scale_y_continuous(limits = c(0, 2500))
)

sta_tdmc <- lmodel2(log(sta_mean) ~ log(tdmc_mean),
    data = ind_traits_filtered,
    nperm = 999)

sta_tdmc

ggplot(data = ind_traits_filtered,
       aes(x = log(tdmc_mean),
           y = log(sta_mean))) +
  geom_point(alpha = 0.1) +
  stat_ma_line(method = "SMA") +
  theme_bw() +
  theme(panel.grid = element_blank()) 

sma(mass_to_height ~ tdmc_mean,
    data = ind_traits_filtered)

plot(mh_tdmc, method = "SMA")

thickness_sav <- sma(log(thickness_mm_mean) ~ log(sav_mean),
    data = ind_traits_filtered,
    nperm = 999)

thickness_sav

ggplot(data = ind_traits_filtered,
       aes(x = log(thickness_mm_mean),
           y = log(sav_mean))) +
  geom_point() +
  stat_ma_line(method = "SMA") +
  labs(x = "log Thallus thickness (mm)",
       y = "log Thallus surface area:volume ratio") +
  theme_bw() +
  theme(panel.grid = element_blank()) 


plot(residuals(thickness_sav))

plot(thickness_sav)

sta_sav <- lmodel2(scale(sta_mean) ~ scale(sav_mean),
                   data = ind_traits_filtered)

sma(scale(sta_mean) ~ scale(sav_mean),
    data = ind_traits_filtered)

sta_sav

plot(sta_sav)


sta_sav <- ggplot(data = ind_traits_filtered,
       aes(x = scale(sav_mean),
           y = scale(sta_mean))) +
  geom_point() +
  stat_ma_line(method = "SMA") +
  labs(x = "scale surface area:volume ratio (mm\U00B2/mL)",
       y = "scale specific thallus area (mm\U00B2/g)") +
  theme_bw() +
  theme(panel.grid = element_blank()) +
  coord_cartesian(ylim = c(0, 16)) +
  annotate(geom = "text",
           x = 4, y = 15,
           label = "R\U00B2 = 0.05, p < 0.001")


mh_sta <- lmodel2(scale(mass_to_height) ~ scale(sta_mean),
                  data = ind_traits_filtered,
                  nperm = 999)

sma(scale(mass_to_height) ~ scale(sta_mean),
    data = ind_traits_filtered)

mh_sta <- ggplot(data = ind_traits_filtered,
       aes(x = scale(sta_mean),
           y = scale(mass_to_height))) +
  geom_point() +
  stat_ma_line(method = "SMA") +
  labs(x = "scaled log specific thallus area (mm\U00B2/g)",
       y = "scaled mass:height ratio (g/cm)") +
  theme_bw() +
  theme(panel.grid = element_blank()) +
  coord_cartesian(ylim = c(-0.5, 7.6)) +
  annotate(geom = "text",
           x = 12, y = 7,
           label = "R\U00B2 = 0.02, p = 0.01")

plot_grid(mh_sta, sta_sav, ncol = 2)

# p-values from correlation in matrix form
p_df <- p$p

# thank you rguha for this solution: https://stackoverflow.com/questions/24554642/what-is-the-best-way-to-remove-items-above-a-diagonal-in-a-data-frame-in-r
# get rid of the upper triangle
p_df[upper.tri(p_df)] <- NA

pairwise_comparisons <- p_df %>% 
  # turn into data frame
  as.data.frame() %>% 
  # make the rownames into "trait1" or the first trait in the pairwise comparison
  rownames_to_column("trait1") %>% 
  # when p < 0.05, keep the p, and when it's not then replace with NA
  mutate(across(where(is.numeric), ~ case_when(. < 0.05 ~ ., TRUE ~ NA))) %>% 
  # make the data frame longer and turn the column names into "trait2" or the second trait in the comparison
  pivot_longer(cols = `Maximum height`:`Aspect ratio`,
               names_to = "trait2",
               values_to = "corr") %>%
  # drop any NAs in the p-values (the upper triangle of the matrix)
  drop_na(corr) %>% 
  # only include p-values that are more than 0 (the matrix has the same-same pair as 0)
  filter(corr > 0) %>% 
  mutate(trait1_column = case_match(
    trait1,
    "SA:V" ~ "sav_mean",
    "Thickness" ~ "thickness_mm_mean",
    "TDMC" ~ "tdmc_mean",
    "STA" ~ "sta_mean",
    "SA:P" ~ "sap_mean",
    "Fv/Fm" ~ "fvfm_mean",
    "Aspect ratio" ~ "aspect_ratio_mean"
  )) %>% 
  mutate(trait2_column = case_match(
    trait2,
    "Maximum height" ~ "maximum_height",
    "Mass:height" ~ "mass_to_height",
    "SA:V" ~ "sav_mean",
    "Thickness" ~ "thickness_mm_mean",
    "TDMC" ~ "tdmc_mean",
    "STA" ~ "sta_mean",
    "SA:P" ~ "sap_mean",
    "Fv/Fm" ~ "fvfm_mean"
  )) %>% 
  mutate(formula = paste(trait1_column, "~", trait2_column, sep = " "))

log_ind_traits <- ind_traits_filtered %>% 
      mutate(across(c(maximum_height, mass_to_height, sav_mean,
                      thickness_mm_mean, sta_mean, sap_mean, aspect_ratio_mean), 
                    log))

scaled_ind_traits <- ind_traits_filtered %>% 
      mutate(across(where(is.numeric), scale))

pairwise_sma <- function(model_formula, trait1, trait2, data) {
  df <- if(data == "normal") {
    ind_traits_filtered
  } else if(data == "scaled") {
    scaled_ind_traits
  } else if(data == "log") {
    log_ind_traits
  } 
  
  # lmodel2
  lmodel2_obj <- lmodel2(model_formula,
                         data = df)
  # sma
  sma_obj <- sma(model_formula,
      data = df)
  
  # plot
  plot <- ggplot(data = df,
                 aes(x = {{ trait1 }},
                     y = {{ trait2 }})) +
    geom_point(alpha = 0.1) +
    stat_ma_line(method = "SMA") +
    # labs(title = model_formula) +
    theme_bw() +
    theme(panel.grid = element_blank())
  
  # return all
  return(list(lmodel2_obj, sma_obj, plot))
  
  
}

pair_sav_height <- pairwise_sma(
  model_formula = "sav_mean ~ maximum_height", 
  trait1 = sav_mean,
  trait2 = maximum_height,
  data = "log"
)

pair_sav_mh <- pairwise_sma(
  model_formula = "sav_mean ~ mass_to_height", 
  trait1 = sav_mean,
  trait2 = mass_to_height,
  data = "log"
)

# fix this later with new mass data entry, though not sure why things worked with 
# other models...
pair_thickness_mh <- pairwise_sma(
  model_formula = "thickness_mm_mean ~ mass_to_height",
  trait1 = thickness_mm_mean,
  trait2 = mass_to_height,
  data = "log"
)

pair_thickness_sav <- pairwise_sma(
  model_formula = "thickness_mm_mean ~ sav_mean", 
  trait1 = thickness_mm_mean,
  trait2 = sav_mean,
  data = "log"
)

pair_tdmc_height <- pairwise_sma(
  model_formula = "tdmc_mean ~ maximum_height", 
  trait1 = tdmc_mean,
  trait2 = maximum_height,
  data = "log"
)

pair_tdmc_mh <- pairwise_sma(
  model_formula = "tdmc_mean ~ mass_to_height", 
  trait1 = tdmc_mean,
  trait2 = mass_to_height,
  data = "log"
)

pair_tdmc_thickness <- pairwise_sma(
  model_formula = "tdmc_mean ~ thickness_mm_mean", 
  trait1 = tdmc_mean,
  trait2 = thickness_mm_mean,
  data = "log"
)

pair_sta_height <- pairwise_sma(
  model_formula = "sta_mean ~ maximum_height", 
  trait1 = sta_mean,
  trait2 = maximum_height,
  data = "log"
)

pair_sta_mh <- pairwise_sma(
  model_formula = "sta_mean ~ mass_to_height", 
  trait1 = sta_mean,
  trait2 = mass_to_height,
  data = "log"
)

pair_sta_sav <- pairwise_sma(
  model_formula = "sta_mean ~ sav_mean", 
  trait1 = sta_mean,
  trait2 = sav_mean,
  data = "log"
)

pair_sta_thickness <- pairwise_sma(
  model_formula = "sta_mean ~ thickness_mm_mean", 
  trait1 = sta_mean,
  trait2 = thickness_mm_mean,
  data = "log"
)

pair_sta_tdmc <- pairwise_sma(
  model_formula = "sta_mean ~ tdmc_mean", 
  trait1 = sta_mean,
  trait2 = tdmc_mean,
  data = "log"
)

pair_sap_height <- pairwise_sma(
  model_formula = "sap_mean ~ maximum_height", 
  trait1 = sap_mean,
  trait2 = maximum_height,
  data = "log"
)

pair_sap_sav <- pairwise_sma(
  model_formula = "sap_mean ~ sav_mean", 
  trait1 = sap_mean,
  trait2 = sav_mean,
  data = "log"
)

pair_sap_thickness <- pairwise_sma(
  model_formula = "sap_mean ~ thickness_mm_mean", 
  trait1 = sap_mean,
  trait2 = thickness_mm_mean,
  data = "log"
)

pair_sap_tdmc <- pairwise_sma(
  model_formula = "sap_mean ~ tdmc_mean", 
  trait1 = sap_mean,
  trait2 = tdmc_mean,
  data = "log"
)

pair_sap_sta <- pairwise_sma(
  model_formula = "sap_mean ~ sta_mean", 
  trait1 = sap_mean,
  trait2 = sta_mean,
  data = "log"
)     

pair_fvfm_height <- pairwise_sma(
  model_formula = "fvfm_mean ~ maximum_height", 
  trait1 = fvfm_mean,
  trait2 = maximum_height,
  data = "log"
)     

pair_fvfm_sav <- pairwise_sma(
  model_formula = "fvfm_mean ~ sav_mean", 
  trait1 = fvfm_mean,
  trait2 = sav_mean,
  data = "log"
)    

pair_fvfm_tdmc <- pairwise_sma(
  model_formula = "fvfm_mean ~ tdmc_mean", 
  trait1 = fvfm_mean,
  trait2 = tdmc_mean,
  data = "log"
)

pair_fvfm_sta <- pairwise_sma(
  model_formula = "fvfm_mean ~ sta_mean", 
  trait1 = fvfm_mean,
  trait2 = sta_mean,
  data = "log"
)

pair_fvfm_sap <- pairwise_sma(
  model_formula = "fvfm_mean ~ sap_mean", 
  trait1 = fvfm_mean,
  trait2 = sap_mean,
  data = "log"
)

pair_ar_height <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ maximum_height", 
  trait1 = aspect_ratio_mean,
  trait2 = maximum_height,
  data = "log"
)

# this also doesn't work!
pair_ar_mh <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ mass_to_height",
  trait1 = aspect_ratio_mean,
  trait2 = mass_to_height,
  data = "log"
)

pair_ar_sav <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ sav_mean", 
  trait1 = aspect_ratio_mean,
  trait2 = sav_mean,
  data = "log"
)

pair_ar_thickness <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ thickness_mm_mean", 
  trait1 = aspect_ratio_mean,
  trait2 = thickness_mm_mean,
  data = "log"
)

pair_ar_tdmc <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ tdmc_mean", 
  trait1 = aspect_ratio_mean,
  trait2 = tdmc_mean,
  data = "log"
)

pair_ar_fvfm <- pairwise_sma(
  model_formula = "aspect_ratio_mean ~ fvfm_mean", 
  trait1 = aspect_ratio_mean,
  trait2 = fvfm_mean,
  data = "log"
)

mh_row_scaled <- plot_grid(
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

mh_row <- plot_grid(
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

sav_row_scaled <- plot_grid(
  NULL,
  pair_sav_mh[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

sav_row <- plot_grid(
  pair_sav_height[[3]],
  pair_sav_mh[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

thickness_row_scaled <- plot_grid(
  NULL,
  pair_thickness_mh[[3]],
  pair_thickness_sav[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

thickness_row <- plot_grid(
  NULL,
  NULL,
  pair_thickness_sav[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

tdmc_row_scaled <- plot_grid(
  NULL,
  pair_tdmc_mh[[3]],
  NULL,
  pair_tdmc_thickness[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

tdmc_row <- plot_grid(
  pair_tdmc_height[[3]],
  pair_tdmc_mh[[3]],
  NULL,
  pair_tdmc_thickness[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

sta_row_scaled <- plot_grid(
  NULL,
  pair_sta_mh[[3]],
  pair_sta_sav[[3]],
  pair_sta_thickness[[3]],
  pair_sta_tdmc[[3]],
  NULL,
  NULL,
  NULL,
  nrow = 1
)

sta_row <- plot_grid(
  pair_sta_height[[3]],
  pair_sta_mh[[3]],
  pair_sta_sav[[3]],
  pair_sta_thickness[[3]],
  pair_sta_tdmc[[3]],
  NULL,
  NULL,
  NULL,
  nrow = 1
)

sap_row_scaled <- plot_grid(
  NULL,
  NULL,
  pair_sap_sav[[3]],
  pair_sap_thickness[[3]],
  pair_sap_tdmc[[3]],
  pair_sap_sta[[3]],
  NULL,
  NULL,
  nrow = 1
)

sap_row <- plot_grid(
  pair_sap_height[[3]],
  NULL,
  pair_sap_sav[[3]],
  NULL,
  pair_sap_tdmc[[3]],
  pair_sap_sta[[3]],
  NULL,
  NULL,
  nrow = 1
)

fvfm_row_scaled <- plot_grid(
  pair_fvfm_height[[3]],
  NULL,
  pair_fvfm_sav[[3]],
  NULL,
  pair_fvfm_tdmc[[3]],
  pair_fvfm_sta[[3]],
  pair_fvfm_sap[[3]],
  NULL,
  nrow = 1
)

fvfm_row <- plot_grid(
  pair_fvfm_height[[3]],
  NULL,
  pair_fvfm_sav[[3]],
  NULL,
  pair_fvfm_tdmc[[3]],
  pair_fvfm_sta[[3]],
  pair_fvfm_sap[[3]],
  NULL,
  nrow = 1
)

ar_row_scaled <- plot_grid(
  NULL,
  NULL,
  pair_ar_sav[[3]], 
  pair_ar_thickness[[3]],
  NULL,
  NULL,
  NULL,
  NULL,
  nrow = 1
)

ar_row <- plot_grid(
  pair_ar_height[[3]],
  NULL,
  pair_ar_sav[[3]], 
  pair_ar_thickness[[3]],
  pair_ar_tdmc[[3]],
  NULL,
  NULL,
  pair_ar_fvfm[[3]],
  nrow = 1
)

sma_grid_scaled <- plot_grid(mh_row_scaled,
                             sav_row_scaled,
                             thickness_row_scaled,
                             tdmc_row_scaled,
                             sta_row_scaled,
                             sap_row_scaled,
                             fvfm_row_scaled,
                             ar_row_scaled,
                             nrow = 8)

sma_grid <- plot_grid(mh_row,
                      sav_row,
                      thickness_row,
                      tdmc_row,
                      sta_row,
                      sap_row,
                      fvfm_row,
                      ar_row,
                      nrow = 8)

cowplot::save_plot(here::here(
  "figures",
  "tradeoffs",
  paste0("sma-grid_", today(), ".jpg")),
  base_height = 10,
  sma_grid_scaled)

ggsave(filename = here::here(
  "figures",
  "tradeoffs",
  paste0("sma-grid_", today(), ".jpg")),
  plot = sma_grid,
  width = 18,
  height = 18,
  units = "cm",
  dpi = 300
)

```



# 5. Clustering

Need two data frames:  

1. species x trait for continuous traits  
2. species x trait for categorical traits

## a. getting species x trait matrices

```{r}
cont_spp_mat <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  # filter(!(sp_code %in% c("CO"))) %>% 
  # filter(!(sp_code == "PTCA" & lifestage == "recruit")) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID %in% c("20210721-BULL-023"))) %>% 
  select(scientific_name, maximum_height, total_volume, sav_mean, thickness_mm_mean, tdmc_mean, sta_mean, sav_mean, sap_mean, fvfm_mean, thickness_mm_mean) %>% 
  group_by(scientific_name) %>% 
  mutate(across(c(maximum_height, total_volume, sav_mean, thickness_mm_mean, tdmc_mean, sta_mean, sav_mean, sap_mean, fvfm_mean, thickness_mm_mean), ~mean(., na.rm = TRUE))) %>% 
  unique() %>% 
  column_to_rownames("scientific_name")

cat_spp_mat <- cat_data %>% 
  # all traits
  select(scientific_name,
         size_cm, thickness, position_to_benthos, articulated, stipe,
         midrib, branching, branch_shape, blades, blade_category,
         coenocyte, attachment, tissue_complexity, growth, calcification) %>%
  # filter(scientific_name %in% rownames(cont_spp_mat)) %>% 
  column_to_rownames("scientific_name") %>% 
  mutate(across(.cols = thickness:calcification, as_factor))
```

### i. sidebar: doing PCA with mean trait values for each species

```{r}
tbs_mean_pca <- rda(cont_spp_mat, scale = TRUE)
screeplot(tbs_mean_pca, bstick = TRUE)

# extract proportion explained for each axis
summary(tbs_mean_pca)
loadings <- scores(tbs_mean_pca, display = 'species', scaling = 0, choices = c(1, 2, 3))

# extract correlation of variables with PC1
sort(abs(loadings[,1]), decreasing = TRUE)
# extract correlation of variables with PC2
sort(abs(loadings[,2]), decreasing = TRUE)

# simple biplot (compare with ggplot output to make sure it's right)
biplot(tbs_mean_pca)

# species points
PCAscores_mean <- scores(tbs_mean_pca, display = "sites", choices = c(1, 2)) %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("scientific_name") %>% 
  left_join(., coarse_traits, by = "scientific_name") %>% 
  mutate(scientific_name = factor(scientific_name, levels = algae_proposal_sciname_factors))

PCAscores_mean_13 <- scores(tbs_mean_pca, display = "sites", choices = c(1, 3)) %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("scientific_name") %>% 
  left_join(., coarse_traits, by = "scientific_name") %>% 
  mutate(scientific_name = factor(scientific_name, levels = algae_proposal_sciname_factors))

# trait vectors
PCAvect_mean <- scores(tbs_mean_pca, display = "species", choices = c(1, 2)) %>% 
  as.data.frame()

PCAvect_13_mean <- scores(tbs_mean_pca, display = "species", choices = c(1, 3)) %>% 
  as.data.frame()

# plot PCA
plot_PCA_mean_12 <- ggplot() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_point(data = PCAscores_mean, 
             aes(x = PC1, 
                 y = PC2, 
                 color = scientific_name), 
             size = 2) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  scale_color_manual(values = algae_colors) +
  geom_segment(data = PCAvect_mean, 
               aes(x = 0, 
                   y = 0, 
                   xend = PC1, 
                   yend = PC2), 
               arrow = arrow(length = unit(0.2, "cm"))) +
  geom_text(data = PCAvect_mean, 
            aes(x = PC1, 
                y = PC2, 
                label = rownames(PCAvect)), 
            size = 5, 
            alpha = 0.8) +
  scale_x_continuous(limits = c(-1.75, 1.75), 
                     expand = c(0, 0)) +
  scale_y_continuous(limits = c(-1.4, 2.1), 
                     expand = c(0, 0)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 2),
         shape = guide_legend(ncol = 2)) +
  theme(legend.position = "none", 
        legend.box = "vertical", 
        legend.margin = margin(),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 12),
        panel.grid = element_blank()) +
  labs(x = "PC1 (33%)",
       y = "PC2 (30%)",
       title = "PC1 and PC2",
       color = "Scientific name") 
plot_PCA_mean_12

plot_PCA_mean_13 <- ggplot() +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_point(data = PCAscores_mean_13, aes(x = PC1, y = PC3, color = scientific_name), size = 2) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  scale_color_manual(values = algae_colors) +

  geom_segment(data = PCAvect_13_mean, aes(x = 0, y = 0, xend = PC1, yend = PC3), arrow = arrow(length = unit(0.2, "cm"))) +
  geom_text(data = PCAvect_13_mean, aes(x = PC1, y = PC3, label = rownames(PCAvect)), size = 5, alpha = 0.8) +
  scale_x_continuous(limits = c(-2.05, 2.05), 
                     expand = c(0, 0)) +
  scale_y_continuous(limits = c(-1.1, 3), 
                     expand = c(0, 0)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 2),
         shape = guide_legend(ncol = 2)) +
  theme(legend.position = "none", 
        legend.box = "vertical", 
        legend.margin = margin(),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 12),
        panel.grid = element_blank()) +
  labs(x = "PC1 (33%)",
       y = "PC3 (15%)",
       title = "PC1 and PC3",
       color = "Scientific name") 
plot_PCA_mean_13

pca_mean_together <- (plot_PCA_mean_12 | plot_PCA_mean_13) +
  plot_layout(widths = c(1, 1)) 
pca_mean_together

# ggsave(here("figures", "prelim-ordinations",
#             paste0("PCA_together_means_", today(), ".jpg")),
#        width = 16,
#        height = 8.5,
#        units = "cm",
#        dpi = 300)

```


## b. gower dissimilarity matrices

```{r}
cont_gower <- gowdis(cont_spp_mat)
cat_gower <- gowdis(cat_spp_mat)
```

## c. clustering

continuous

```{r}
cont_groups <- cluster::pam(x = cont_gower,
                            k = 3,
                            metric = "euclidean")

pairwise.perm.manova(cont_gower, 
                     fact = cont_groups$clustering, 
                     p.method = "none")

# 4 is too many groups
# groups 1 and 4 are not different from each other
# cont_groups <- cluster::pam(x = cont_gower,
#                             k = 4,
#                             metric = "euclidean")
# 
# pairwise.perm.manova(cont_gower,
#                      fact = cont_groups$clustering,
#                      p.method = "none")
```

categorical

```{r}
cat_groups <- cluster::pam(x = cat_gower,
                            k = 2,
                            metric = "euclidean")

pairwise.perm.manova(cat_gower, 
                     fact = cat_groups$clustering, 
                     p.method = "none")

# 3 is too many groups
# groups 2 and 3 are not different
# cat_groups <- cluster::pam(x = cat_gower,
#                             k = 3,
#                             metric = "euclidean")
# 
# pairwise.perm.manova(cat_gower, 
#                      fact = cat_groups$clustering, 
#                      p.method = "none")
```

getting clusters out

```{r}
cont_clusters <- cont_groups$clustering %>% 
  enframe() %>% 
  rename(cont_cluster = value,
         scientific_name = name) %>% 
  # make sure clusters are factors
  mutate(cont_cluster = factor(cont_cluster)) 

cat_clusters <- cat_groups$clustering %>% 
  enframe() %>% 
  rename(cat_cluster = value,
         scientific_name = name) %>% 
  # make sure clusters are factors
  mutate(cat_cluster = factor(cat_cluster)) 
```


## d. visualizing

```{r}
cont_nmds <- metaMDS(cont_gower) 

cont_nmds_scores <- scores(cont_nmds,
                           choices = c(1, 2),
                           display = c("sites", "species"),
                           tidy = TRUE) %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("scientific_name") %>% 
  mutate(label = case_when(
    scientific_name == "Dictyota binghamiae; Dictyota flabellata; Dictyota coriacea" ~ "Dictyota spp.",
    scientific_name == "Chondracanthus corymbiferus; Chondracanthus exasperatus" ~ "Chondracanthus spp.",
    TRUE ~ scientific_name
  )) %>% 
  left_join(., lte_spp, by = "scientific_name") %>% 
  select(!(life_cycle:eco_zone)) %>% 
  left_join(., cont_clusters, by = "scientific_name") %>% 
  left_join(., cat_clusters, by = "scientific_name")


cont_nmds_plot <- ggplot(cont_nmds_scores,
                          aes(x = NMDS1, y = NMDS2)) +
  geom_point(aes(color = cont_cluster,
                 shape = taxon_phylum), 
             size = 1,
             alpha = 0.9) +
  geom_convexhull(aes(fill = cont_cluster),
                  alpha = 0.6) +
  scale_x_continuous(limits = c(-0.25, 0.25),
                     breaks = c(-0.2, 0, 0.2)) +
  scale_y_continuous(limits = c(-0.25, 0.25),
                     breaks = c(-0.2, 0, 0.2)) +
  scale_color_manual(values = c(cluster1, cluster2, cluster3)) +
  scale_fill_manual(values = c(cluster1, cluster2, cluster3)) +
  annotate(geom = "text", x = -0.2, y = 0.15, label = "1", color = cluster1, size = 10) +
  annotate(geom = "text", x = -0.1, y = -0.1, label = "2", color = cluster2, size = 10) +
  annotate(geom = "text", x = 0.2, y = 0.08, label = "3", color = cluster3, size = 10) +
  labs(title = "PAM clusters with continuous traits") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 16))
  
cont_nmds_plot

cat_nmds_plot <- ggplot(cont_nmds_scores,
                          aes(x = NMDS1, y = NMDS2)) +
  geom_point(aes(color = cat_cluster,
                 shape = taxon_phylum), 
             size = 1,
             alpha = 0.9) +
  geom_convexhull(aes(fill = cat_cluster),
                  alpha = 0.6) +
  scale_x_continuous(limits = c(-0.25, 0.25),
                     breaks = c(-0.2, 0, 0.2)) +
  scale_y_continuous(limits = c(-0.25, 0.25),
                     breaks = c(-0.2, 0, 0.2)) +
  scale_color_manual(values = c(`1` = cat_cluster1, `2` = cat_cluster2)) +
  scale_fill_manual(values = c(`1` = cat_cluster1, `2` = cat_cluster2)) +
  annotate(geom = "text", x = 0.2, y = -0.2, label = "1", color = cat_cluster1, size = 10) +
  annotate(geom = "text", x = -0.2, y = 0, label = "2", color = cat_cluster2, size = 10) +
  labs(title = "PAM clusters with categorical traits") +
  theme_bw() +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        text = element_text(size = 12))
  
cat_nmds_plot

# ggsave(here::here("figures", "prelim-ordinations",
#                   paste("cont-nmds-plot_", today(), ".jpg", sep = "")),
#        cont_nmds_plot,
#        width = 6, height = 6, units = "cm", dpi = 300)
# 
# ggsave(here::here("figures", "prelim-ordinations",
#                   paste("cat-nmds-plot_", today(), ".jpg", sep = "")),
#        cat_nmds_plot,
#        width = 6, height = 6, units = "cm", dpi = 300)
```

## e. table

```{r}
FitFlextableToPage <- function(ft, pgwidth = 6){
  
  ft_out <- ft %>% autofit()
  
  ft_out <- width(ft_out, width = dim(ft_out)$widths*pgwidth /(flextable_dim(ft_out)$widths))
  return(ft_out)
}

cont_clusters_table <- cont_spp_mat %>% 
  rownames_to_column("scientific_name") %>% 
  left_join(., cont_clusters, by = "scientific_name") %>% 
  arrange(cont_cluster) %>% 
  relocate(cont_cluster, .after = "scientific_name") %>% 
  rename(cluster = cont_cluster) %>% 
  mutate(across(maximum_height:fvfm_mean, ~ round(., digits = 2))) %>% 
  flextable() %>% 
  bg(i = ~ cluster == 1, j = c("scientific_name", "cluster"), bg = cluster1) %>% 
  bg(i = ~ cluster == 2, j = c("scientific_name", "cluster"), bg = cluster2) %>% 
  bg(i = ~ cluster == 3, j = c("scientific_name", "cluster"), bg = cluster3) %>% 
  autofit() %>% 
  FitFlextableToPage(pgwidth = 10)
cont_clusters_table

# save_as_docx(cont_clusters_table,
#              path = here::here("tables",
#                                paste0("cont-trait-clusters_", today(), ".docx")))
```

# 6. Functional diversity

```{r}
benthics <- read_csv(
  here("data",
       "SBC-LTER-benthics",
       "Annual_All_Species_Biomass_at_transect_20240501.csv")
) 

lte <- read_csv(
  here("data",
       "SBC-LTER-benthics",
       "LTE_All_Species_Biomass_at_transect_20240501.csv")
)

benthics_clean <- benthics %>% 
  clean_names() %>%
  # replace NA sp_code with Nandersoniana
  mutate(sp_code = case_when(
    scientific_name == "Nienburgia andersoniana" ~ "Nandersoniana",
    TRUE ~ sp_code
  )) %>%
  filter(sp_code %in% algae_proposal) %>% 
  # replace all -99999 values with NA
  mutate(dry_gm2 = replace(dry_gm2, dry_gm2 < 0, NA),
         wm_gm2 = replace(wm_gm2, wm_gm2 < 0, NA),
         density = replace(density, density < 0, NA)) %>%
  # create a sample_ID for each sampling date at each treatment at each site
  unite("sample_ID", site, year, transect, remove = FALSE) %>%
  # change to lower case
  mutate_at(c("group", "mobility", "growth_morph", "site"), str_to_lower) 

lte_clean <- lte %>% 
  clean_names() %>%
  # filter(sp_code != "MAPY") %>% 
  # replace NA sp_code with Nandersoniana
  mutate(sp_code = case_when(
    scientific_name == "Nienburgia andersoniana" ~ "Nandersoniana",
    TRUE ~ sp_code
  )) %>%
  # filter(sp_code %in% algae_proposal) %>% 
  # replace all -99999 values with NA
  mutate(dry_gm2 = replace(dry_gm2, dry_gm2 < 0, NA),
         wm_gm2 = replace(wm_gm2, wm_gm2 < 0, NA)) %>%
  mutate(across(c(site, treatment, group, mobility, growth_morph, site), str_to_lower)) %>% 
  filter(group == "algae") %>% 
  # create a sample_ID for each sampling date at each treatment at each site
  unite("sample_ID", site, date, treatment, remove = FALSE) 

site_by_spp <- benthics_clean %>% 
  select(sample_ID, scientific_name, dry_gm2) %>% 
  pivot_wider(names_from = "scientific_name",
              values_from = "dry_gm2") %>% 
  column_to_rownames("sample_ID") %>% 
  select(rownames(cont_spp_mat)) %>% 
  mutate(sum = rowSums(.[1:16])) %>% 
  filter(sum > 0) %>% 
  select(-sum)

site_by_spp_lte <- lte_clean %>% 
  select(sample_ID, scientific_name, dry_gm2) %>% 
  pivot_wider(names_from = "scientific_name",
              values_from = "dry_gm2") %>% 
  column_to_rownames("sample_ID") %>% 
  select(rownames(cat_spp_mat)) %>% 
  mutate(sum = rowSums(.[1:16])) %>% 
  filter(sum > 0) %>% 
  select(-sum)

site_meta <- benthics_clean %>% 
  # select metadata columns
  select(sample_ID, year, month, date, site, transect, vis) %>% 
  # get rid of duplicates
  unique() %>% 
  # replace all -99999 values with NA
  mutate(vis = replace(vis, vis < 0, NA)) 

site_meta_lte <- lte_clean %>% 
  # select metadata columns
  select(sample_ID, year, month, date, site, treatment) %>% 
  # get rid of duplicates
  unique() 

zero_sites <- benthics_clean %>% 
  select(sample_ID, scientific_name, dry_gm2) %>% 
  pivot_wider(names_from = "scientific_name",
              values_from = "dry_gm2") %>% 
  column_to_rownames("sample_ID") %>% 
  select(rownames(cont_spp_mat)) %>% 
  mutate(sum = rowSums(.[1:16])) %>% 
  filter(!(sum > 0)) %>% 
  rownames_to_column("sample_ID") %>% 
  select(sample_ID)

# cont_spp_mat is the spp x trait matrix


```

run this if you need to regenerate the fd stuff:

```{r}
fd <- dbFD(x = cont_spp_mat, # traits
           a = site_by_spp, # abundance
           stand.x = TRUE # standardize all traits to mean 0
           )

fd_cat_lte <- dbFD(x = cat_spp_mat,
                   a = site_by_spp_lte,
                   corr = "none")

str(fd)

write_rds(x = fd,
          file = here("data", "rds-objects",
                      paste0("fd_", today(), ".rds")))
```

run this if you just need to read it in!

```{r}
fd <- read_rds(here("data", 
                    "rds-objects",
                    "fd_2024-09-26.rds"))

test <- fd_fric(cont_spp_mat, site_by_spp)

fric <- fd$FRic %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("sample_ID") %>% 
  rename(funcrich = value)

fdiv <- fd$FDiv %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("sample_ID") %>% 
  rename(funcdiv = value)

fdis <- fd$FDis %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("sample_ID") %>% 
  rename(funcdis = value)

cwm <- fd$CWM %>% 
  as_tibble(rownames = NA) %>% 
  rownames_to_column("sample_ID") %>% 
  rename_with(~str_c(., "_cwm"), .cols = maximum_height:fvfm_mean)
  
```


1. calculate biomass of species that have traits  
2. plot biomass on y-axis, functional richness or evenness on x-axis

```{r}

survey_biomass <- site_by_spp %>% 
  mutate(sum = rowSums(.[1:16])) %>% 
  rownames_to_column("sample_ID") %>% 
  left_join(., site_meta, by = "sample_ID") %>% 
  left_join(., fric, by = "sample_ID") %>% 
  left_join(., fdiv, by = "sample_ID") %>% 
  left_join(., fdis, by = "sample_ID") %>% 
  left_join(., cwm, by = "sample_ID")

ggplot(data = survey_biomass,
       aes(x = funcrich,
           y = sum)) +
  geom_point(aes(color = site),
             alpha = 0.4) +
  geom_smooth(method = "lm",
              se = FALSE,
              linewidth = 2) +
  labs(title = "smooth line for visualization only") +
  theme_bw() +
  theme(panel.grid = element_blank())

mod <- glmmTMB(sum ~ funcrich + (1|site/transect) + (1|year),
               data = survey_biomass %>% filter(!(is.na(funcrich))),
               na.action = na.omit)

simulateResiduals(mod) %>% plot()

plotResiduals(simulateResiduals(mod), form = survey_biomass$funcrich, 
              quantreg = FALSE)

ggpairs(data = survey_biomass,
        columns = c("sum", "funcrich", "funcdiv",
                    "funcdis", "maximum_height_cwm",
                    "total_volume_cwm", "sav_mean_cwm",
                    "thickness_mm_mean_cwm", "tdmc_mean_cwm",
                    "sta_mean_cwm", "sap_mean_cwm",
                    "fvfm_mean_cwm"))
```





# 7. PTCA adults vs recruits

```{r}
tbsample_matrix_ptca <- av_leaf_values %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "PTCA") %>% 
  filter(site %in% c("Arroyo Quemado", "Mohawk")) %>% 
  left_join(., volume_ind, by = "specimen_ID") %>% 
  left_join(., fvfm_ind, by = "specimen_ID") %>% 
  left_join(., ind_height, by = "specimen_ID") %>% 
  dplyr::select(site, lifestage, specimen_ID, thickness_mm_mean, tdmc_mean, total_volume, fvfm_mean, frond_length_mean, frond_width_mean, maximum_height) %>% 
  rename("FvFm" = fvfm_mean,
         "STA" = thickness_mm_mean,
         "TDMC" = tdmc_mean,
         "Volume" = total_volume,
         "Frond length" = frond_length_mean,
         "Frond width" = frond_width_mean,
         "Maximum height" = maximum_height)

tbsample_mat <- av_leaf_values %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  filter(sp_code == "PTCA") %>% 
  filter(site %in% c("Arroyo Quemado", "Mohawk")) %>% 
  left_join(., volume_ind, by = "specimen_ID") %>% 
  left_join(., fvfm_ind, by = "specimen_ID") %>% 
  left_join(., ind_height, by = "specimen_ID") %>% 
  dplyr::select(specimen_ID, thickness_mm_mean, tdmc_mean, total_volume, fvfm_mean, frond_length_mean, frond_width_mean, maximum_height) %>% 
  rename("FvFm" = fvfm_mean,
         "STA" = thickness_mm_mean,
         "TDMC" = tdmc_mean,
         "Volume" = total_volume,
         "Frond length" = frond_length_mean,
         "Frond width" = frond_width_mean,
         "Maximum height" = maximum_height) %>% 
  column_to_rownames("specimen_ID") %>% 
  drop_na()

tbsample_meta <- tbsample_matrix_ptca %>% 
  select(specimen_ID, lifestage, site) %>% 
  filter(specimen_ID %in% rownames(tbsample_mat))

ptca_pca <- rda(tbsample_mat, scale = TRUE)
summary(ptca_pca)
biplot(ptca_pca)

# species points
PCAscores_12_ptca <- scores(ptca_pca, display = "sites", choices = c(1, 2)) %>% 
  as.data.frame() %>% 
  rownames_to_column("specimen_ID") %>% 
  left_join(., tbsample_meta, by = "specimen_ID")

PCAscores_13_ptca <- scores(tbs_pca, display = "sites", choices = c(1, 3)) %>% 
  as.data.frame() %>% 
  rownames_to_column("specimen_ID") %>% 
  left_join(., tbsample_meta, by = "specimen_ID")

# trait vectors
PCAvect_12_ptca <- scores(ptca_pca, display = "species", choices = c(1, 2)) %>% 
  as.data.frame()

PCAvect_13_ptca <- scores(ptca_pca, display = "species", choices = c(1, 3)) %>% 
  as.data.frame()

# plot PCA
plot_PCA_12_ptca <- ggplot() +
  coord_cartesian() +
  geom_point(data = PCAscores_12_ptca, aes(x = PC1, y = PC2, color = lifestage, shape = site), size = 3, alpha = 0.9) +
  scale_color_manual(values = c("#70601d", "#b59b33")) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  # scale_color_manual(values = pal) +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_segment(data = PCAvect_12_ptca, aes(x = 0, y = 0, xend = PC1, yend = PC2), arrow = arrow(length = unit(0.2, "cm")), size = 1) +
  geom_text_repel(data = PCAvect_12_ptca, aes(x = PC1, y = PC2, label = rownames(PCAvect_12_ptca)), size = 6, alpha = 0.8) +
  # scale_x_continuous(limits = c(-1, 2.1)) +
  # scale_y_continuous(limits = c(-2.7, 1.6)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 2),
         shape = guide_legend(ncol = 2)) +
  theme(legend.position = "bottom", 
        legend.box = "vertical", 
        legend.margin = margin(),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 12),
        legend.key = element_rect(size = 11), 
        panel.grid = element_blank()) +
  labs(x = "PC1 (59.07%)",
       y = "PC2 (19.14%)",
       color = "Life stage") 
plot_PCA_12_ptca

plot_PCA_13 <- ggplot() +
  coord_cartesian() +
  geom_point(data = PCAscores_13, aes(x = PC1, y = PC3, color = scientific_name, shape = site), size = 3, alpha = 0.8) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  # scale_color_manual(values = pal) +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_segment(data = PCAvect_13, aes(x = 0, y = 0, xend = PC1, yend = PC3), arrow = arrow(length = unit(0.2, "cm")), size = 1) +
  geom_text(data = PCAvect_13, aes(x = PC1, y = PC3, label = rownames(PCAvect)), size = 5, alpha = 0.8) +
  scale_x_continuous(limits = c(-1, 2.1)) +
  scale_y_continuous(limits = c(-2.7, 1.6)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 2),
         shape = guide_legend(ncol = 2)) +
  theme(legend.position = "bottom", 
        legend.box = "vertical", 
        legend.margin = margin(),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 12)) +
  labs(x = "PC1 (42.4%)",
       y = "PC3 (12.4%)",
       title = "PC1 and PC3",
       color = "Scientific name") 
plot_PCA_13

pca_together <- (plot_PCA_12 | plot_PCA_13) +
  plot_layout(guides = "collect", ncol = 2, widths = c(1, 1)) &
  theme(legend.position = "bottom")
pca_together
```

# 8. environmental conditions

```{r}
env_mat <- CWM_fixed_filtered %>% 
  dplyr::select(transect_ID, mean_urchins, mean_subs, mean_kelp, mean_umol) %>% 
  column_to_rownames("transect_ID") %>% 
  drop_na()

corrplot(cor(env_mat), method = "circle", type = "lower", insig = "blank", addCoef.col = "black")


env_pca <- rda(env_mat, scale = TRUE)

summary(env_pca)
biplot(env_pca)

PCAscores_12_env <- scores(env_pca, display = "sites", choices = c(1, 2)) %>% 
  as.data.frame() %>% 
  rownames_to_column("transect_ID") %>% 
  left_join(., community_metadata_transect, by = "transect_ID")

PCAvect_12_env <- scores(env_pca, display = "species", choices = c(1, 2)) %>% 
  as.data.frame()

plot_PCA_12_env <- ggplot() +
  coord_cartesian() +
  geom_point(data = PCAscores_12_env, aes(x = PC1, y = PC2, color = site, shape = site), size = 3, alpha = 0.7) +
  # scale_color_manual(values = c("#70601d", "#b59b33")) +
  # geom_text(data = PCAscores, aes(x = PC1, y = PC2, color = scientific_name, label = specimen_ID)) +
  # scale_color_manual(values = pal) +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  geom_segment(data = PCAvect_12_env, aes(x = 0, y = 0, xend = PC1, yend = PC2), arrow = arrow(length = unit(0.2, "cm")), size = 1) +
  geom_text_repel(data = PCAvect_12_env, aes(x = PC1, y = PC2, label = rownames(PCAvect_12_env)), size = 6, alpha = 0.8) +
  # scale_x_continuous(limits = c(-1, 2.1)) +
  # scale_y_continuous(limits = c(-2.7, 1.6)) +
  theme_bw() +
  guides(colour = guide_legend(ncol = 1),
         shape = guide_legend(ncol = 1)) +
  theme(legend.position = c(0.9, 0.2),
        legend.box = "vertical", 
        legend.margin = margin(),
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 16),
        title = element_text(size = 20),
        legend.text = element_text(size = 12),
        legend.key = element_rect(size = 11), 
        panel.grid = element_blank()) +
  labs(x = "PC1 (31.26%)",
       y = "PC2 (28.56%)") 
plot_PCA_12_env

adonis2(env_mat ~ site, data = CWM_fixed_filtered, method = "euclidean")
```
