==> code/old-code/04_benthic-data.Rmd <==
---
title: "Abundant species dynamics from benthic surveys"
author: "An Bui"
date: "1/31/2022"
output: 
  html_document:
    toc: yes
    toc_depth: 4
    toc_float:
      collapsed: no
    theme: readable
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# 0. set up
```{r source, warning = FALSE, message = FALSE}
# source the cleaning/summarizing script, which sources the set up
source(here::here("code", "01a_trait-cleaning.R"))
```

# 1. Biomass

## a. Guiding questions

What's the percent community biomass that can be attributed to:  
1. _Laminariales_, and  
2. the 11 abundant species from Miller et al.?  

Data question: what is the difference between the percent cover data and the percent cover that's in the column in the biomass data?  

Stray thoughts: introduced/native species trait column?

## b. Test case: SCTW from 3 different sampling points

To make sure the code was working, I tested this out with "SCTW_2010-07-29", "SCTW_2016-07-20", and "SCTW_2020-07-29". The issue comes from scenarios where species were counted at more than one transect - for example, at SCTW_2020-07-29, EC was at both transect 2 and 3, but I only want the dry/wet biomass for EC for the whole survey.  

When I create the data frame, there are then "duplicate" observations for total species biomass for EC because it was present at both transects. To fix this, I filtered the data frame for unique observations of the total and percent biomass calculations and double checked the total dry percent to make sure everything added up to 1.  

### i. Calculations

This is just a test to make sure the calculations worked.

```{r calc-test, include = FALSE, eval = FALSE}
# test <- biomass %>% 
#   filter(sample_ID %in% c("SCTW_2010-07-29", "SCTW_2016-07-20", "SCTW_2020-07-29")) %>% 
#   select(sample_ID, site, year, date, sp_code, dry_gm2, wm_gm2) %>% 
#   # create a new column where total biomass for both surveys is calculated
#   group_by(site, year) %>% 
#   mutate(total_dry = sum(dry_gm2),
#          total_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where total biomass for the species for the whole survey is calculated
#   group_by(site, year, sp_code) %>% 
#   mutate(total_sp_dry = sum(dry_gm2),
#          total_sp_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where percent of total biomass per survey for each species is calculated
#   mutate(percent_sp_dry = total_sp_dry/total_dry,
#          percent_sp_wet = total_sp_wet/total_wet) %>% 
#   # double check to make sure the percents worked out...
#   select(sample_ID, site, year, date, sp_code, 
#          # values that need to be unique: 
#          total_sp_dry, total_sp_wet, percent_sp_dry, percent_sp_wet) %>% 
#   unique() %>% 
#   group_by(sample_ID) %>% 
#   mutate(total_percent = sum(percent_sp_dry),
#          total_percent_wet = sum(percent_sp_wet)) %>% 
#   # join with coarse traits data frame, which includes taxonomy
#   left_join(., coarse_traits, by = "sp_code")
```

### ii. Plots

These are tests of the plots.

```{r plot-test, include = FALSE, eval = FALSE}
# # plot of proportion total biomass by phylum faceted by sampling date
# ggplot(test) +
#   aes(x = taxon_phylum, y = percent_sp_dry) +
#   # calculating sum of proportion within stat_summary
#   stat_summary(aes(fill = taxon_phylum),
#                fun = "sum", geom = "bar") +
#   # adding labels for sum values
#   stat_summary(aes(label = round(..y.., 2)), 
#                fun = "sum", geom = "text", 
#                size = 4, vjust = -0.5) +
#   scale_fill_manual(values = c("Chlorophyta" = chloro_col, 
#                                "Ochrophyta" = ochro_col, 
#                                "Rhodophyta" = rhodo_col)) +
#   scale_y_continuous(limits = c(0, 1), expand = c(0, 0)) +
#   facet_wrap(~sample_ID) +
#   theme_classic() +
#   theme(legend.position = "none") +
#   labs(x = "Phylum",
#        y = "Proportion total biomass",
#        title = "Biomass by phylum for SCTW")
# 
# # pretend timeseries
# ggplot(test) +
#   aes(x = date, y = percent_sp_dry) +
#   # calculating sum of proportion within stat_summary
#   stat_summary(aes(col = taxon_phylum),
#                fun = "sum", geom = "line", size = 1) +
#   stat_summary(aes(col = taxon_phylum),
#                fun = "sum", geom = "point", size = 2) +
#   scale_color_manual(values = c("Chlorophyta" = chloro_col, 
#                                 "Ochrophyta" = ochro_col, 
#                                 "Rhodophyta" = rhodo_col)) +
#   scale_y_continuous(limits = c(0, 1), expand = c(0, 0)) +
#   theme_bw() +
#   # theme(legend.position = "none") +
#   labs(x = "Sampling date",
#        y = "Proportion total biomass",
#        title = "Biomass by phylum through time for SCTW",
#        col = "Phylum")
```

## c. Biomass data set

Since this worked, I felt comfortable doing this for the whole `biomass` data frame.  

### i. Calculations

**Note:** as of 2022-03-04, the code to create the `prop_biomass` data frame is in the `00_set-up.R` script.  

### ii. Plots

Made some plot functions to keep things tidy.  

**Note:** FIX THIS. YOU NEED TO REDO THESE PLOTS BECAUSE THE DATA FRAME THEY'RE BUILT ON DOESN'T TAKE THE AVERAGE AT THE SITE ANYMORE.  
```{r plot-fxns, include = FALSE, eval = FALSE}
# phylum.plot <- function(site_choice) {
#   # choose a subtitle from sites_full
#   subtitle <- pluck(sites_full, site_choice)
#   
#   ggplot(prop_biomass %>% filter(site == {{ site_choice }})) +
#   aes(x = date, y = percent_sp_dry) +
#   # calculating sum of proportion within stat_summary
#   stat_summary(aes(col = taxon_phylum),
#                fun = "sum", geom = "line", size = 1) +
#   stat_summary(aes(col = taxon_phylum),
#                fun = "sum", geom = "point", size = 2) +
#   scale_color_manual(values = c("Chlorophyta" = chloro_col, 
#                                 "Ochrophyta" = ochro_col, 
#                                 "Rhodophyta" = rhodo_col)) +
#   scale_y_continuous(limits = c(0, 1), expand = c(0, 0)) +
#   theme_bw() +
#   # theme(legend.position = "none") +
#   labs(x = "Sampling date",
#        y = "Proportion total biomass",
#        title = "Biomass by phylum through time",
#        subtitle = subtitle,
#        col = "Phylum")
# }
# 
# growth.form.plot <- function(site_choice) {
#   # choose a subtitle from sites_full
#   subtitle <- pluck(sites_full, site_choice)
#   
#   ggplot(prop_biomass %>% filter(site == {{ site_choice }})) +
#     aes(x = date, y = percent_sp_dry) +
#     # calculating sum of proportion within stat_summary
#     stat_summary(aes(col = growth_form),
#                  fun = "sum", geom = "line", size = 1) +
#     stat_summary(aes(col = growth_form),
#                  fun = "sum", geom = "point", size = 2) +
#     scale_y_continuous(limits = c(0, 1), expand = c(0, 0)) +
#     theme_bw() +
#     labs(x = "Sampling date",
#          y = "Proportion total biomass",
#          title = "Biomass by growth form through time",
#          subtitle = subtitle,
#          col = "Growth form")
# }
# 
# spp.interest.plot <- function(site_choice) {
#   # choose a subtitle from sites_full
#   subtitle <- pluck(sites_full, site_choice)
#   
#   ggplot(prop_biomass %>% filter(site == {{ site_choice }})) +
#     aes(x = date, y = percent_sp_dry) +
#     # calculating sum of proportion within stat_summary
#     stat_summary(aes(col = algae_interest),
#                  fun = "sum", geom = "line", size = 1) +
#     stat_summary(aes(col = algae_interest),
#                  fun = "sum", geom = "point", size = 2) +
#     scale_y_continuous(limits = c(0, 1), expand = c(0, 0)) +
#     theme_bw() +
#     labs(x = "Sampling date",
#          y = "Proportion total biomass",
#          title = "Species of interest",
#          subtitle = subtitle, 
#          col = " ")
# }
# 
# species.plot <- function(site_choice) {
#   # choose a title from sites_full
#   title <- pluck(sites_full, site_choice)
#   
#   df <- prop_biomass %>% ungroup() %>% filter(site == {{ site_choice }})
#   
#   plot_ly(data = df, 
#           x = ~date, y = ~percent_sp_dry, 
#           color = ~scientific_name, colors = c(cal_palette(name = "tidepool", type = "continuous")), 
#           mode = "markers+lines") %>% 
#     add_trace(mode = "markers+lines", type = "scatter") %>% 
#     layout(title = title, 
#            xaxis = list(title = "Date", nticks = 8), 
#            yaxis = list(title = "Proportion total biomass", range = c(-0.01, 1)),
#            legend = list(title = list(text = '<b> Scientific name </b>')))
# }

```

```{r plots, include = FALSE, eval = FALSE}
# phylum.plot("bull")
# phylum.plot("aque")
# phylum.plot("ahnd")
# phylum.plot("napl")
# phylum.plot("ivee")
# phylum.plot("golb")
# phylum.plot("abur")
# phylum.plot("mohk")
# phylum.plot("carp")
# phylum.plot("scdi")
# phylum.plot("sctw")
# 
# growth.form.plot("bull")
# growth.form.plot("aque")
# growth.form.plot("ahnd")
# growth.form.plot("napl")
# growth.form.plot("ivee")
# growth.form.plot("golb")
# growth.form.plot("abur")
# growth.form.plot("mohk")
# growth.form.plot("carp")
# growth.form.plot("scdi")
# growth.form.plot("sctw")
# 
# spp.interest.plot("bull")
# spp.interest.plot("aque")
# spp.interest.plot("ahnd")
# spp.interest.plot("napl")
# spp.interest.plot("ivee")
# spp.interest.plot("golb")
# spp.interest.plot("abur")
# spp.interest.plot("mohk")
# spp.interest.plot("carp")
# spp.interest.plot("scdi")
# spp.interest.plot("sctw")
# 
# # these use plotly so that you can look at species on their own
# species.plot("bull")
# species.plot("aque")
# species.plot("ahnd")
# species.plot("napl")
# species.plot("ivee")
# species.plot("golb")
# species.plot("abur")
# species.plot("mohk")
# species.plot("carp")
# species.plot("scdi")
# species.plot("sctw")
```

# 2. Tables of species contribution to biomass

This is a big table of percent species contribution to community biomass for each site across all sampling points. It's useful for figuring out where to sample what. `05_species-selection-tables.R` is a separate script with this code, essentially to play around with sites.

```{r full-table, include = FALSE, eval = FALSE}
# table_df <- biomass %>% 
#   select(site, scientific_name, dry_gm2, wm_gm2) %>% 
#   mutate(site = fct_relevel(site, "bull", "aque", "ahnd", "napl", "ivee", "golb", "abur",
#                             "mohk", "carp", "scdi", "sctw")) %>% 
#   # filter out MAPY
#   filter(scientific_name != "Macrocystis pyrifera") %>% 
#   # create a new column where total biomass for the whole survey is calculated
#   group_by(site) %>% 
#   mutate(total_dry = sum(dry_gm2),
#          total_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where total biomass for the species for the whole survey is calculated
#   group_by(site, scientific_name) %>% 
#   mutate(total_sp_dry = sum(dry_gm2),
#          total_sp_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where percent of total biomass per survey for each species is calculated
#   mutate(percent_sp_dry = total_sp_dry/total_dry,
#          percent_sp_wet = total_sp_wet/total_wet) %>% 
#   # double check to make sure the percents worked out...
#   select(site, scientific_name, 
#          # values that need to be unique: 
#          total_sp_dry, total_sp_wet, percent_sp_dry, percent_sp_wet) %>% 
#   unique() %>% 
#   group_by(site) %>% 
#   mutate(total_percent = sum(percent_sp_dry),
#          total_percent_wet = sum(percent_sp_wet)) %>% 
#   select(scientific_name, site, percent_sp_dry) %>% 
#   mutate(percent_sp_dry = round(percent_sp_dry, 4)) %>% 
#   mutate(percent_whole_sp_dry = percent_sp_dry*100) %>% 
#   select(-percent_sp_dry) %>% 
#   mutate(site = str_to_upper(site)) %>% 
#   pivot_wider(names_from = "site", values_from = "percent_whole_sp_dry") %>% 
#   left_join(., coarse_traits, by = "scientific_name") %>% 
#   slice(-2) %>% 
#   select(scientific_name, sp_code, BULL, AQUE, AHND, NAPL, IVEE, GOLB, ABUR,
#                             MOHK, CARP, SCDI, SCTW)
# 
# table <- table_df %>% 
#   gt() %>% 
#   data_color(columns = 3,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 3]), 0))) %>% 
#   data_color(columns = 4,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 4]), 0))) %>% 
#   data_color(columns = 5,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 5]), 0))) %>% 
#   data_color(columns = 6,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 6]), 0))) %>% 
#   data_color(columns = 7,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 7]), 0))) %>% 
#   data_color(columns = 8,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 8]), 0))) %>% 
#   data_color(columns = 9,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 9]), 0))) %>% 
#   data_color(columns = 10,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 10]), 0))) %>% 
#   data_color(columns = 11,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 11]), 0))) %>% 
#   data_color(columns = 12,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 12]), 0))) %>% 
#   data_color(columns = 13,
#              colors = scales::col_numeric(palette = gradient_palette, 
#                                           domain = c(max(table_df[, 13]), 0))) %>% 
#   fmt_missing(columns = 3:13,
#               missing_text = "--") %>% 
#   tab_spanner(
#     label = "Sites",
#     columns = c(BULL, AQUE, AHND, NAPL, IVEE, GOLB, ABUR,
#                 MOHK, CARP, SCDI, SCTW)
#   ) %>% 
#   tab_spanner(
#     label = "Species",
#     columns = c("scientific_name", "sp_code")
#   ) %>% 
#   cols_label(
#     scientific_name = "Scientific name",
#     sp_code = "Species code",
#   )
#   %>% gtsave("sp_site_table-selection.png", path = here::here("tables"))
```

```{r filter-table, include = FALSE, eval = FALSE}
# filtered_table_df <- biomass %>% 
#   select(site, scientific_name, dry_gm2, wm_gm2) %>% 
#   mutate(site = fct_relevel(site, "bull", "aque", "ahnd", "napl", "ivee", "golb", "abur",
#                             "mohk", "carp", "scdi", "sctw")) %>% 
#   mutate(spp_interest = case_when(
#     scientific_name %in% c("Desmarestia ligulata", 
#                    "Stephanocystis osmundacea", 
#                    "Sargassum horneri", 
#                    "Sargassum muticum", 
#                    "Eisenia arborea", 
#                    "Egregia menziesii", 
#                    "Laminaria farlowii", 
#                    "Pterygophora californica") ~ scientific_name)) %>% 
#   mutate(spp_interest = case_when(
#     is.na(spp_interest) ~ "other species",
#     TRUE ~ as.character(spp_interest)
#   )) %>% 
#   # filter out MAPY
#   filter(scientific_name != "Macrocystis pyrifera") %>% 
#   # create a new column where total biomass for the whole survey is calculated
#   group_by(site) %>% 
#   mutate(total_dry = sum(dry_gm2),
#          total_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where total biomass for the species for the whole survey is calculated
#   group_by(site, spp_interest) %>% 
#   mutate(total_sp_dry = sum(dry_gm2),
#          total_sp_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where percent of total biomass per survey for each species is calculated
#   mutate(percent_sp_dry = total_sp_dry/total_dry,
#          percent_sp_wet = total_sp_wet/total_wet) %>% 
#   # double check to make sure the percents worked out...
#   select(site, spp_interest, 
#          # values that need to be unique: 
#          total_sp_dry, total_sp_wet, percent_sp_dry, percent_sp_wet) %>% 
#   unique() %>% 
#   group_by(site) %>% 
#   mutate(total_percent = sum(percent_sp_dry),
#          total_percent_wet = sum(percent_sp_wet)) %>% 
#   select(spp_interest, site, percent_sp_dry) %>% 
#   mutate(percent_sp_dry = round(percent_sp_dry, 4)) %>% 
#   mutate(percent_whole_sp_dry = percent_sp_dry*100) %>% 
#   select(-percent_sp_dry) %>% 
#   mutate(site = str_to_upper(site)) %>% 
#   pivot_wider(names_from = "site", values_from = "percent_whole_sp_dry") %>% 
#   left_join(., coarse_traits, by = "scientific_name") %>% 
#   slice(-2) %>% 
#   select(scientific_name, sp_code, BULL, AQUE, AHND, NAPL, IVEE, GOLB, ABUR,
#                             MOHK, CARP, SCDI, SCTW)
```

# 3. Urchin density at each site

Plots using `stat_summary` of mean urchin density at each site and a timeseries:

```{r urchin-plot}
urchin_plot <- ggplot(urchins %>% filter(site %in% c("aque", "napl", "ivee", "mohk", "carp")), aes(x = reorder(site, mean_urchins), y = mean_urchins)) +
  stat_summary(aes(fill = site), 
               fun = "mean", geom = "col") +
  stat_summary(aes(group = site),
               fun.data = mean_se, geom = "errorbar", width = 0.5) +
  labs(x = "Site", y = "Mean urchin density (number/m2)",
       title = "Purple urchin density") +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(0, 35)) +
  theme_bw() +
  theme(legend.position = "none")

urchin_plot

urchins_time_plot <- urchins %>% 
  filter(site %in% c("bull", "aque", "napl", "ivee", "mohk", "carp")) %>% 
  filter(year > 2015) %>% 
  ggplot(aes(x = year, y = mean_urchins)) +
  # stat_summary(aes(col = site), 
  #              fun = "sum", geom = "point", size = 3) +
  # stat_summary(aes(col = site),
  #              fun = "sum", geom = "line") +
  stat_summary(aes(col = site),
               fun = "mean", geom = "point", size = 3) +
  stat_summary(aes(col = site),
               fun = "mean", geom = "line") +
  stat_summary(aes(col = site),
               fun.data = mean_se, geom = "errorbar", width = 0.5) +
  labs(x = "Year", y = "Mean urchin density (number/m2)",
       title = "Purple urchin density through time") +
  scale_y_continuous(expand = c(0, 0)) +
  # coord_cartesian(ylim = c(-0.1, 15)) +
  theme_bw()

urchins_time_plot
```

Plot with pairwise comparisons of means:

```{r urchin-aov}
# analysis of variance
urchin_aov <- urchins %>% 
  filter(site %in% c("aque", "napl", "ivee", "mohk", "carp")) %>% 
  # filter(year > 2015) %>% 
  aov(na.omit(mean_urchins) ~ site, data = .)

# Tukey HSD
urchin_tukey <- TukeyHSD(urchin_aov)

# compact letter display
urchin_cld <- multcompLetters4(urchin_aov, urchin_tukey)

# output of compact letter display in data frame
urchin_cld_df <- enframe(urchin_cld[[1]]$Letters)

# summary data frame for plotting
urchin_plot_summary <- urchins %>% 
  filter(site %in% c("aque", "ivee", "mohk", "napl", "carp")) %>% 
  # filter(year > 2015) %>% 
  group_by(site) %>% 
  summarize(mean = mean(mean_urchins, na.rm = TRUE),
            se = se(mean_urchins), 
            sd = sd(mean_urchins)) %>% 
  left_join(., urchin_cld_df, by = c("site" = "name")) %>% 
  mutate(value = fct_relevel(value, "a", "b", "bc", "c")) %>% 
  arrange(value) %>% 
  mutate(site = fct_relevel(site, "carp", "mohk", "napl", "aque", "ivee"))

# new plot
urchin_plot_new <- ggplot(urchin_plot_summary, aes(x = site, y = mean)) +
  geom_col(aes(fill = value)) +
  scale_fill_manual(values = c("a" = "#4e1f70", 
                               "b" = "#814aa8",
                               "bc" = "#a06cc4",
                               "c" = "#cba0e8")) +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.5) +
  geom_text(aes(label = value, y = mean + se + 10), size = 5) + 
  labs(x = "Site", y = "Mean urchin density (number/m2)",
       title = "Purple urchin density") +
  scale_y_continuous(expand = c(0, 0),
                     limits = c(0, 200)) +
  # coord_cartesian(ylim = c(0, 12)) +
  theme_bw() +
  theme(legend.position = "none",
        axis.title = element_text(size = 18),
        axis.text = element_text(size = 18),
        plot.title = element_text(size = 20),
        panel.grid = element_blank())

urchin_plot_new
```


# 4. Sand cover

Plot using `stat_summary`:

```{r sand-plot}
# sand_plot <- ggplot(substrate %>% filter(common_name == "Sand"), aes(x = site, y = percent_cover)) +
#   stat_summary(aes(fill = site), 
#                fun = "mean", geom = "col") +
#   stat_summary(aes(group = site), 
#                fun.data = mean_se, geom = "errorbar", width = 0.5) +
#   # facet_wrap(~site) +
#   labs(x = "Site", y = "Mean sand percent cover (%)",
#        title = "Sand cover across sites") +
#   scale_y_continuous(expand = c(0, 0)) +
#   coord_cartesian(ylim = c(0, 90)) +
#   theme_bw() +
#   theme(legend.position = "none")
# 
# sand_plot

# sand_time_plot <- ggplot(substrate %>% filter(common_name == "Sand"), aes(x = year, y = percent_cover)) +
#   stat_summary(aes(col = site), 
#                fun = "mean", geom = "point", size = 3) +
#   stat_summary(aes(col = site),
#                fun = "mean", geom = "line") +
#   stat_summary(aes(col = site), 
#                fun.data = mean_se, geom = "errorbar", width = 0.5) +
#   # facet_wrap(~site) +
#   labs(x = "Site", y = "Mean sand percent cover (%)",
#        title = "Sand cover across sites through time") +
#   scale_y_continuous(expand = c(0, 0)) +
#   coord_cartesian(ylim = c(-1, 105)) +
#   theme_bw() 
# 
# sand_time_plot
```

In case saving plots is of interest: 

```{r plots-together, include = FALSE, eval = FALSE}
# plots_together <- urchin_plot / sand_plot

# ggsave(here::here("figures/habitat-characteristics", 
#                   paste("habitat-characteristics-", todays_date, ".jpg", sep = "")), 
#        plots_together)
# 
# plots_together <- urchins_time_plot / sand_time_plot
# 
# ggsave(here::here("figures/habitat-characteristics", 
#                   paste("habitat-characteristics-timeseries", todays_date, ".jpg", sep = "")), 
#        plots_together)
```

Plot with pairwise comparisons of means:

```{r plots-pairwise}
# filter out sand
sand <- substrate %>% 
  # filter(common_name == "Sand") %>% 
  filter(site %in% c("bull", "aque", "ivee", "mohk", "napl", "carp"))

# analysis of variance
sand_aov <- aov(na.omit(sand$percent_cover) ~ sand$site)

# Tukey HSD
sand_tukey <- TukeyHSD(sand_aov)

# compact letter display
sand_cld <- multcompLetters4(sand_aov, sand_tukey)

# results into data frame
sand_cld_df <- enframe(sand_cld[[1]]$Letters)

# summary data frame
sand_plot_summary <- sand %>% 
  group_by(site) %>% 
  summarize(mean = mean(percent_cover, na.rm = TRUE),
            se = se(percent_cover)) %>% 
  left_join(., sand_cld_df, by = c("site" = "name")) %>% 
  mutate(site = fct_relevel(site, "bull", "aque", "ivee", "mohk", "napl", "carp"))

# new plot
sand_plot_new <- ggplot(sand_plot_summary, aes(x = site, y = mean)) +
  geom_col(aes(fill = value)) +
  scale_fill_manual(values = c("a" = "#bda36f",
                               "b" = "#8c733f",
                               "c" = "#614815")) +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.5) +
  geom_text(aes(label = value, y = mean + se + 1)) + 
  labs(x = "Site", y = "Mean sand percent cover (%)",
       title = "Sand cover across sites") +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(0, 24)) +
  theme_bw() +
  theme(legend.position = "none",
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 14),
        plot.title = element_text(size = 16),
        panel.grid = element_blank())

sand_plot_new
```

Together:
```{r}
plots_together <- urchin_plot_new / sand_plot_new

# ggsave(here::here("figures/habitat-characteristics",
#                   paste("habitat-characteristics-", todays_date, ".jpg", sep = "")),
#        plots_together)
```

# 5. Kelp cover

Plot using `stat_summary`:

```{r kelp-plots}
biomass %>% 
  filter(sp_code == "MAPY") %>% 
  filter(!(site %in% c("scdi", "sctw"))) %>% 
  # replace all -99999 values with 0
  mutate(wm_gm2 = replace(wm_gm2, wm_gm2 < 0, 0)) %>% 
  ggplot(aes(x = reorder(site, wm_gm2), y = wm_gm2)) +
  stat_summary(aes(fill = site), 
               fun = "mean", geom = "col") +
  stat_summary(aes(group = site),
               fun.data = mean_se, geom = "errorbar", width = 0.5) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(0, 7000)) +
  theme_bw() +
  theme(legend.position = "none") +
  labs(x = "Site", y = "Kelp biomass (wet g/m2)",
       title = "Kelp biomass across sites")

biomass %>% 
  filter(sp_code == "MAPY") %>% 
  filter(site == "carp") %>% 
  # filter(!(site %in% c("scdi", "sctw"))) %>% 
  # replace all -99999 values with 0
  mutate(wm_gm2 = replace(wm_gm2, wm_gm2 < 0, 0)) %>% 
  ggplot(aes(x = year, y = wm_gm2)) +
  stat_summary(aes(col = site), 
               fun = "mean", geom = "point", size = 3) +
  stat_summary(aes(col = site),
               fun = "mean", geom = "line") +
  stat_summary(aes(col = site), 
               fun.data = mean_se, geom = "errorbar", width = 0.5) +
  theme_bw() 
```

Plot with pairwise comparisons:  

**Note:** does it make sense that mean kelp biomass is the same across all sites?

```{r kelp-pairwise}
mapy <- biomass %>% 
  # filter(!(site %in% c("sctw", "scdi"))) %>% 
  filter(site %in% sites_proposal) %>% 
  filter(sp_code == "MAPY") %>% 
  group_by(site, year) %>% 
  summarize(total = mean(dry_gm2, na.rm = TRUE)) %>% 
  drop_na(total)

mapy_aov <- aov(total ~ site, data = mapy)

mapy_tukey <- TukeyHSD(mapy_aov)

mapy_cld <- multcompLetters4(mapy_aov, mapy_tukey)

mapy_cld_df <- enframe(mapy_cld[[1]]$Letters)

mapy_plot_summary <- mapy %>% 
  group_by(site) %>% 
  summarize(mean = mean(total, na.rm = TRUE),
            se = se(total)) %>% 
  left_join(., mapy_cld_df, by = c("site" = "name")) %>% 
  mutate(site = fct_relevel(site, "bull", "aque", "napl", "ivee", "mohk", "carp"))

mapy_plot_new <- ggplot(mapy_plot_summary, aes(x = site, y = mean)) +
  geom_col() +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.5) +
  geom_text(aes(label = value, y = mean + se + 70)) + 
  labs(x = "Site", y = "Mean biomass (g dry/m2)",
       title = "MAPY biomass across sites") +
  # scale_y_continuous(expand = c(0, 0)) +
  # coord_cartesian(ylim = c(0, 1600)) +
  theme_bw()

mapy_plot_new
```

# 6. Understory biomass

```{r understory-plot}
macrobio <- biomass %>% 
  filter(site %in% c("bull", "aque", "ivee", "mohk", "napl", "carp")) %>% 
  filter(sp_code != "MAPY") %>% 
  group_by(site, year) %>% 
  summarize(total = sum(dry_gm2))

macrobio_aov <- aov(macrobio$total ~ macrobio$site)

macrobio_tukey <- TukeyHSD(macrobio_aov)

macrobio_cld <- multcompLetters4(macrobio_aov, macrobio_tukey)

macrobio_cld_df <- enframe(macrobio_cld[[1]]$Letters)

macrobio_plot_summary <- macrobio %>% 
  group_by(site) %>% 
  summarize(mean = mean(total, na.rm = TRUE),
            se = sd(total)/sqrt(length(total))) %>% 
  left_join(., macrobio_cld_df, by = c("site" = "name")) 
  # mutate(site = fct_relevel(site, "napl", "carp", "aque", "ivee", "mohk"))

macrobio_plot_new <- ggplot(macrobio_plot_summary, aes(x = site, y = mean)) +
  geom_col() +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.5) +
  geom_text(aes(label = value, y = mean + se + 50)) + 
  labs(x = "Site", y = "Mean biomass (g dry/m2)",
       title = "Understory macroalgal biomass across sites") +
  scale_y_continuous(expand = c(0, 0)) +
  # coord_cartesian(ylim = c(0, 1600)) +
  theme_bw()

macrobio_plot_new

# all algae

all_macro <- biomass %>% 
  filter(site %in% c("aque", "ivee", "mohk", "napl", "carp")) %>% 
  group_by(site, year) %>% 
  summarize(total = sum(dry_gm2))

all_macro_aov <- aov(all_macro$total ~ all_macro$site)

all_macro_tukey <- TukeyHSD(all_macro_aov)

all_macro_cld <- multcompLetters4(all_macro_aov, all_macro_tukey)

all_macro_cld_df <- enframe(all_macro_cld[[1]]$Letters)

all_macro_plot_summary <- all_macro %>% 
  group_by(site) %>% 
  summarize(mean = mean(total, na.rm = TRUE),
            se = sd(total)/sqrt(length(total))) %>% 
  left_join(., all_macro_cld_df, by = c("site" = "name")) 
  # mutate(site = fct_relevel(site, "napl", "carp", "aque", "ivee", "mohk"))

all_macro_plot_new <- ggplot(all_macro_plot_summary, aes(x = site, y = mean)) +
  geom_col() +
  geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.5) +
  geom_text(aes(label = value, y = mean + se + 80)) + 
  labs(x = "Site", y = "Mean biomass (g dry/m2)",
       title = "Whole community biomass across sites") +
  scale_y_continuous(expand = c(0, 0)) +
  # coord_cartesian(ylim = c(0, 4500)) +
  theme_bw()

all_macro_plot_new
```

# 7. Community metrics

## a. wrangling

Going to make a site by species matrix for each sampling date. There are a couple of anomalies: CARP_2000-09-08 and NAPL_2000-09-18 don't have any observations, and AQUE_2019-07-23 is missing the sand cover data because sand cover was taken on 2019-07-24. Going to fix that later.   

```{r}
community_matrix <- biomass %>% 
  filter(sp_code %in% unique(metadata_ind$sp_code)) %>% 
  filter(!(sp_code == "MAPY")) %>% 
  filter(site %in% sites_proposal_new) %>% 
  select(sample_ID, scientific_name, dry_gm2) %>% 
  pivot_wider(names_from = scientific_name,
              values_from = dry_gm2) %>% 
  column_to_rownames("sample_ID") %>% 
  mutate(sum = rowSums(across(where(is.numeric)))) %>% 
  filter(sum > 0)

# 647 transects, 595 with more than 0

community_metadata <- biomass %>% 
  filter(sp_code %in% unique(metadata_ind$sp_code)) %>% 
  select(sample_ID, year, month, date, site, transect) %>% 
  unique() %>% 
  filter(sample_ID %in% rownames(community_matrix))
```


## b. Species richness/diversity

### i. Richness

```{r sp-rich}
sp_rich <- specnumber(community_matrix) %>% 
  enframe() %>% 
  left_join(., community_metadata, by = c("name" = "sample_ID")) %>% 
  rename(richness = value)

sp_rich_site_mod <- glmmTMB(richness ~ year + (1|site) + (1|transect),
                            data = sp_rich,
                            family = poisson(link = "log"))

simulateResiduals(sp_rich_site_mod) %>% plot()

summary(sp_rich_site_mod)

ggpredict(sp_rich_site_mod,
          terms = c("year")) %>% 
  plot(show_data = TRUE)
```

### ii. Diversity
```{r sp-div}
sp_div <- diversity(community_matrix, index = "shannon") %>% 
  enframe() %>% 
  left_join(., community_metadata, by = c("name" = "sample_ID")) %>% 
  rename(shandiv = value)

sp_div_site_mod <- glmmTMB(shandiv ~ year + (1|site) + (1|transect),
                            data = sp_div)

simulateResiduals(sp_div_site_mod) %>% plot()

summary(sp_div_site_mod)

ggpredict(sp_div_site_mod,
          terms = c("year")) %>% 
  plot(show_data = TRUE)
```


## c. SIMPER analysis

Wanted to know which species significantly contributed to differences in community composition.

```{r simper-analysis}
#sim <- with(community_metadata, simper(community_matrix, site)) 

# write_rds(sim, here::here("data", "rds-objects", "simper.rds"))

sim <- read_rds(here::here("data", "rds-objects", "simper.rds"))

head(summary(sim))
```

## d. NMDS 

### i. Jaccard (presence/absence)

```{r nmds-jacc}
community_mds_jacc <- metaMDS(community_matrix, distance = "jaccard")

community_dist_jacc <- vegdist(community_matrix, method = "jaccard")

stressplot(community_mds_jacc)

community_mds_sites_jacc <- scores(community_mds_jacc, display = "sites") %>% 
  as.data.frame() %>% 
  rownames_to_column("sample_ID") %>% 
  left_join(., community_metadata, by = "sample_ID") 

community_mds_spp_jacc <- scores(community_mds_jacc, display = "species") %>% 
  as.data.frame() %>% 
  rownames_to_column("sp_code") %>% 
  left_join(., coarse_traits, by = "sp_code") %>% 
  filter(sp_code %in% algae_proposal)

perm_jacc <- adonis2(community_dist_jacc ~ year, 
                     data = community_metadata,
                     strata = community_metadata$site)

perm_jacc

disper_jacc <- betadisper(community_dist_jacc, community_metadata$year)

anova(disper_jacc)

TukeyHSD(disper_jacc)

fit_jacc <- community_metadata  %>% 
  dplyr::select(sample_ID, site, 
                mean_urchins, max_subs, mean_kelp, mean_umol) %>% 
  mutate(across(.cols = 3:4, ~replace(., is.na(.), 0))) %>% 
  envfit(community_mds_jacc, ., perm = 999, na.rm = TRUE)

fit_vect_jacc <- scores(fit_jacc, display = "vectors") %>% 
  as.data.frame()

nmds_plot_jacc <- ggplot() +
  # x and y axes
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  # points for sites
  geom_point(data = community_mds_sites_jacc, aes(x = NMDS1, y = NMDS2, color = site)) +
  stat_ellipse(data = community_mds_sites_jacc, aes(x = NMDS1, y = NMDS2, color = site), level = 0.95, size = 1) +
  # # arrows for species
  # geom_segment(data = community_mds_spp_jacc, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm")), color = "blue") +
  # geom_text(data = community_mds_spp_jacc, aes(x = NMDS1, y = NMDS2, label = scientific_name), color = "blue") +
  # # arrows for environmental characteristics (vector fitting)
  geom_segment(data = fit_vect_jacc, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm"))) +
  geom_text(data = fit_vect_jacc, aes(x = NMDS1, y = NMDS2, label = rownames(fit_vect_jacc))) +
  # theme
  theme_bw() +
  guides(colour = guide_legend(nrow = 3)) +
  theme(legend.position = "bottom", legend.box = "vertical", legend.margin = margin()) 

nmds_plot_jacc
```


### ii. Euclidean

```{r nmds-bray}
community_mds_euclidean <- metaMDS(community_matrix_transect_sub, distance = "euclidean")

community_dist_euclidean <- vegdist(community_matrix_transect_sub, method = "euclidean")

stressplot(community_mds_euclidean)

community_mds_sites_euclidean <- scores(community_mds_euclidean, display = "sites") %>% 
  as.data.frame() %>% 
  rownames_to_column("transect_ID") %>% 
  left_join(., community_metadata_transect_sub, by = "transect_ID") 

community_mds_spp_euclidean <- scores(community_mds_euclidean, display = "species") %>% 
  as.data.frame() %>% 
  rownames_to_column("sp_code") %>% 
  left_join(., coarse_traits, by = "sp_code") %>% 
  filter(sp_code %in% algae_proposal)

perm_euclidean <- adonis2(community_matrix_transect ~ site, data = community_metadata_transect, method = "euclidean")

perm_euclidean

disper_euclidean <- betadisper(community_dist_euclidean, community_metadata_transect_sub$site)

anova(disper_euclidean)

TukeyHSD(disper_euclidean)

fit_euclidean <- community_metadata_transect_sub %>% 
  dplyr::select(transect_ID, site, mean_urchins, mean_subs, mean_kelp, mean_umol) %>% 
  mutate(across(.cols = 3:4, ~replace(., is.na(.), 0))) %>% 
  envfit(community_mds_euclidean, ., perm = 999, na.rm = TRUE)

fit_vect_euclidean <- scores(fit_euclidean, display = "vectors") %>% 
  as.data.frame()

nmds_plot_euclidean <- ggplot() +
  geom_point(data = community_mds_sites_euclidean, aes(x = NMDS1, y = NMDS2, color = site, shape = site)) +
  # scale_color_manual(values = pal) +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  stat_ellipse(data = community_mds_sites_euclidean, aes(x = NMDS1, y = NMDS2, color = site), level = 0.95, size = 1) +
  geom_segment(data = community_mds_spp_euclidean, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm"))) +
  geom_text(data = community_mds_spp_euclidean, aes(x = NMDS1, y = NMDS2, label = scientific_name)) +
  geom_segment(data = fit_vect_euclidean, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm")), col = "blue") +
  geom_text(data = fit_vect_euclidean, aes(x = NMDS1, y = NMDS2, label = rownames(fit_vect_euclidean)), col = "blue") +
  theme_bw() +
  guides(colour = guide_legend(ncol = 1)) +
  theme(legend.box = "vertical", legend.margin = margin()) 

nmds_plot_euclidean
```

## e. rank abundance

```{r}
ra_bull <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("bull"))

ra_aque <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("aque"))

ra_napl <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("napl"))

ra_ivee <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("ivee"))

ra_mohk <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("mohk"))

ra_carp <- rankabundance(x = community_matrix, y = community_metadata, factor = "site", level = c("carp"))
```

## f. biomass through time

```{r}
prop_biomass %>% 
  filter(site == "ivee") %>% 
  filter(sp_code == "CYOS") %>% 
  ggplot(aes(x = year, y = percent_sp_dry, col = transect)) +
  stat_summary(aes(col = as.factor(transect)), geom = "point", fun = "mean") +
  stat_summary(aes(col = as.factor(transect)), geom = "line", fun = "mean") +
  stat_summary(aes(col = as.factor(transect)), geom = "errorbar", fun.data = mean_se)

prop_biomass %>% 
  filter(site %in% sites_proposal) %>% 
  filter(sp_code == "CYOS") %>% 
  ggplot(aes(x = year, y = percent_sp_dry)) +
  stat_summary(aes(col = site), geom = "point", fun = "mean") +
  stat_summary(aes(col = site), geom = "line", fun = "mean")

prop_biomass %>% 
  filter(site %in% sites_proposal) %>% 
  filter(sp_code == "DL") %>% 
  ggplot(aes(x = year, y = percent_sp_dry)) +
  stat_summary(aes(col = site), geom = "point", fun = "mean") +
  stat_summary(aes(col = site), geom = "line", fun = "mean")
```



# 8. Trait calculations

## a. proportion of traits

"Mean proportion of biomass" for each species per trait across sampling dates: matrix where row is sampling date, and columns are traits spread across, and cells are total biomass that can be "assigned" to that trait. The `prop_biomass` data frame is already set up.

### i. growth form

```{r prop-traits}
prop_traits <- prop_biomass %>% 
  filter(site %in% c("aque", "napl", "ivee", "mohk", "carp")) %>%
  select(sample_ID, site, year, sp_code, growth_form,
         dry_gm2, wm_gm2, total_dry, total_wet, percent_sp_dry, percent_sp_wet) %>% 
  # create a new column where total biomass for the whole survey is calculated
  # group_by(site, year) %>% 
  # mutate(total_dry = sum(dry_gm2),
  #        total_wet = sum(wm_gm2)) %>% 
  # ungroup() %>% 
  # calculate biomass within each growth form
  group_by(sample_ID, growth_form) %>% 
  mutate(total_gf_dry = sum(dry_gm2),
         total_gf_wet = sum(wm_gm2)) %>% 
  ungroup() %>% 
  select(sample_ID, growth_form, total_gf_dry, total_dry, total_gf_wet, total_wet) %>% 
  unique() %>% 
  group_by(sample_ID, growth_form) %>% 
  summarize(percent_gf_dry = total_gf_dry/total_dry,
         percent_gf_wet = total_gf_wet/total_wet) %>% 
  ungroup() %>% 
  # double check: should sum to 1
  # group_by(sample_ID) %>%
  # mutate(test = sum(unique(percent_gf_dry)))
  # make the data frame wider
  select(sample_ID, growth_form, percent_gf_dry) %>% 
  unique() %>% 
  pivot_wider(names_from = "growth_form", values_from = "percent_gf_dry")
```

GLM???

```{r prop-GLM, include = FALSE, eval = FALSE}
# df <- biomass %>% 
#   # filter sites of interest
#   filter(site %in% c("bull", "aque", "ahnd", "napl", "ivee", "golb", 
#            "abur", "mohk", "carp")) %>% 
#   # only include understory species
#   filter(sp_code != "MAPY") %>% 
#   # join with trait data
#   left_join(., coarse_traits, by = "sp_code") %>% 
#   group_by(sample_ID, site, year, date, growth_form) %>% 
#   summarize(mean_gf_dry = mean(dry_gm2, na.omit = TRUE),
#          mean_gf_wet = mean(wm_gm2, na.omit = TRUE),
#          total_gf_dry = sum(dry_gm2),
#          total_gf_wet = sum(wm_gm2)) %>% 
#   ungroup() %>% 
#   # create a new column where total biomass for the whole survey is calculated
#   group_by(site, year) %>% 
#   mutate(total_dry = sum(total_gf_dry),
#          total_wet = sum(total_gf_wet)) %>% 
#   ungroup() %>% 
#   # create new column where percent growth form biomass is calculated for each survey
#   mutate(percent_gf_dry = total_gf_dry/total_dry,
#          percent_gf_wet = total_gf_wet/total_wet) %>% 
#   # join with metadata data frame
#   left_join(., community_metadata, by = "sample_ID") %>% 
#   filter(growth_form == "leathery_macrophyte") %>% 
#   # filter(total_gf_dry > 0) %>% 
#   drop_na()
# 
# ggplot(df, aes(x = urchin_density, y = mean_gf_dry)) +
#   geom_point() +
#   geom_smooth(method = "lm", se = FALSE) 
# 
# ggplot(df, aes(x = max_subs, y = mean_gf_dry)) +
#   geom_point() +
#   geom_smooth(method = "lm", se = FALSE) 
# 
# ggplot(df %>% filter(mean_gf_dry > 0), aes(x = mean_gf_dry)) +
#   geom_histogram()
# 
# # GLM with tweedie distribution for leathery macrophytes
# model1 <- glmmTMB(mean_gf_dry ~ urchin_density + max_subs + urchin_density*max_subs + (1|site.x) + (1|year.x), 
#               data = df, family = tweedie(link = "log"))
# 
# summary(model1)
# 
# simul.resid <- simulateResiduals(model1)
# 
# plot(simul.resid)
# 
# 
# testZeroInflation(model1)
# # no zero inflation - which makes sense, because it's a tweedie distribution
# testDispersion(simul.resid)
# # no over dispersion
# 
# # GLM with gamma distribution and a log link function, but without the interaction between urchins and sand
# model2 <- glmmTMB(total_gf_dry ~ urchin_density + max_subs + (1|site.x) + (1|year.x), 
#               data = df, family = Gamma(link = "log"))
# 
# summary(model2)
# 
# AIC(model1, model2)
# 
# predict.glm(model1, type = "response")
# 
# # GLM with beta
# model1.prop <- glmmTMB(percent_gf_dry ~ urchin_density + max_subs + mean_kelp + (1|site.x) + (1|year.x),
#                        data = df, family = beta_family(link = "logit"))
# 
# summary(model1.prop)
# 
# emmeans(model1, "urchin_density")
```

# 9. Community-weighted means

These are calculated using the `functcomp` function in the `FD` package. The function requires a matrix with functional traits and a matrix of species abundances (or presence-absence). In the trait matrix, rows are species and columns are traits. In the species matrix, rows are samples and columns are species.  

The first step is to get the `coarse_traits` data frame into a functional form (lol).

```{r CWM-setup}
trait_matrix <- coarse_traits %>% 
  select(scientific_name, growth_morph, growth_form, pigment_type, 
         height_class, longevity, posture, branching_yn) %>% 
  filter(scientific_name != "Macrocystis pyrifera") %>% 
  column_to_rownames("scientific_name")

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(trait_matrix) == colnames(community_matrix)
```

Then the community matrix has to be a matrix, not a data frame (why?).
```{r turn-into-matrix}
comm_matrix <- as.matrix(community_matrix)
```


Then, you're ready to calculate community-weighted means.

```{r fixed-CWM}
CWM_cat <- functcomp(trait_matrix, comm_matrix, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("sample_ID") %>% 
  left_join(., community_metadata, by = "sample_ID") %>% 
  # to get the glmm to work!!!!!!
  mutate(across(.cols = everything(), .fns = ~ replace(., . == 1, 0.999)))
```

For categorical traits, the output is the exact same as the `prop_traits` data frame up top - each cell is the proportion of the community biomass that is attributed to a specific category (for example, within `growth_morph`, at CARP in 2000, there is 92% `solitary` and 8 % `aggregated`).

Now to plot?

## a. GLMs

Kill me!

```{r lm-glm}
lm_plot <- ggplot(CWM_cat, 
                  aes(x = urchin_density, y = growth_form_leathery_macrophyte)) +
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(-0.001, 1)) +
  geom_point(aes(col = site, shape = site), alpha = 0.5) +
  geom_smooth(method = "lm", se = FALSE, aes(col = site)) +
  geom_smooth(method = "lm", col = "black") +
  labs(x = "Urchin density", y = "Proportion biomass (dry g/m2)",
       title = "Proportion leathery macrophytes") +
  theme_bw() +
  theme(plot.title = element_text(size = 16),
        axis.title = element_text(size = 14),
        axis.text = element_text(size = 14),
        panel.grid = element_blank())
# missing values for mean sand

lm_plot

ggplot(CWM_cat, aes(x = growth_form_leathery_macrophyte)) +
  geom_histogram(binwidth = 0.1)

# full model
model_lm_01a <- glmmTMB(growth_form_leathery_macrophyte ~ urchin_density + max_subs + kelp_dry + urchin_density*max_subs + (1|site) + (1|year), 
              data = CWM_cat, ziformula = ~ urchin_density + max_subs + urchin_density*max_subs, family = beta_family(link = "logit"))
# significant effects of sand and kelp and maybe random intercept
# in zero-inflation: significant effect of urchin density and kelp 
# interpretation: proportion of community biomass that is "leathery macrophyte" is predicted by sand and kelp, such that sand has a positive (???) relationship with proportion of community biomass, and kelp has a positive (???) relationship with proportion of community biomass.
# however, having no LMs in the community is predicted by urchin density and kelp biomass 

# sand and kelp and intercepts only
model_lm_01b <- glmmTMB(growth_form_leathery_macrophyte ~ urchin_density + max_subs + kelp_dry + (1|site) + (1|year), 
              data = CWM_cat, ziformula = ~ urchin_density + max_subs + kelp_dry, family = beta_family(link = "logit"))

# plot residuals
model_lm_01a_resid <- simulateResiduals(model_lm_01a)
plot(model_lm_01a_resid)

model_lm_01b_resid <- simulateResiduals(model_lm_01b)
plot(model_lm_01b_resid)

# look at the model
model_lm_01a
summary(model_lm_01a)

model_lm_01b
summary(model_lm_01b)
performance::r2_zeroinflated(model_lm_01b, method = "correlation")

AIC(model_lm_01a, model_lm_01b)

# plot predicted values on top of predictor
temp <- ggpredict(model_lm_01a, terms = c("urchin_density", "site"), type = "fixed") %>% 
  rename("site" = group) %>% 
  rename("urchin_density" = x)

# no data in the global ggplot call
ggplot() +
  # just some expansions of the axes to make them go from 0 - 1
  scale_x_continuous(expand = c(0, 0)) +
  scale_y_continuous(expand = c(0, 0)) +
  coord_cartesian(ylim = c(-0.001, 1)) +
  # raw data
  geom_point(data = CWM_cat, aes(x = urchin_density, y = growth_form_leathery_macrophyte), alpha = 0.5) +
  # lines from ggpredict
  # confidence interval
  geom_ribbon(data = temp, aes(x = urchin_density, y = predicted, ymin = conf.low, ymax = conf.high, fill = site), alpha = 0.2) +
  # line
  geom_line(data = temp, aes(x = urchin_density, y = predicted, color = site)) +
  # axis labels
  labs(x = "Urchin density", y = "Proportion biomass (dry g/m2)",
       title = "Proportion leathery macrophytes") +
  # facet using group
  facet_wrap(~site)
```

## b. Checking correlations

Checking to see if urchins, sand, kelp, and irradiance are at all correlated.

```{r correlations}
ggplot(CWM_cat, aes(x = urchin_density, y = max_subs)) +
  geom_point(aes(col = site)) +
  geom_smooth()

# create correlation matrix
cor_mat <- cor(CWM_cat[, c("urchin_density", "max_subs", "mean_daily_umol", "kelp_dry")], use = "na.or.complete")
# plot correlation matrix
corrplot(cor_mat, type = "lower", addCoef.col = "black")
# 1) sand and urchins are slightly negatively correlated, 
# 2) irradiance is moderately positively correlated with urchin density, 
# 3) urchin density is moderately negatively correlated with kelp biomass, 
# 4) sand cover is slightly negatively correlated with kelp biomass, and 
# 5) average daily irradiance is slightly negatively correlated with kelp

# check for collinearity in the model
check_collinearity(model_lm_01c)

# plot VIF for each variable (if >5, inflates)
plot(check_collinearity(model_lm_01c))
```

# 10. CWM for categorical and continuous traits

## a. fixed CWM using `FD::dbFD()` 

Use `community_matrix` and `tbspp_matrix` from set up script.

```{r fixed-CWM-again}
cwm_commat <- community_matrix_transect %>% 
  dplyr::select(c(rownames(tbspp_matrix))) %>% 
  # take out 0 sites
  mutate(total = rowSums(across(where(is.numeric)))) %>% 
  filter(total > 0) %>% 
  dplyr::select(-total) %>% 
  as.matrix()

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(tbspp_matrix) == colnames(cwm_commat)

CWM_fixed <- functcomp(x = tbspp_matrix, a = cwm_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("transect_ID") %>% 
  rename_with(.fn = ~ paste0(., "_fixed", sep = ""), .cols = 2:28) %>% 
  left_join(., community_metadata_transect, by = "transect_ID")

scaled_coef <- CWM_fixed %>% 
  dplyr::select(transect_ID, mean_urchins, max_subs, mean_subs, mean_kelp, mean_umol) %>% 
  column_to_rownames("transect_ID") %>% 
  stdize() %>% 
  rownames_to_column("transect_ID")

CWM_fixed <- CWM_fixed %>% 
  left_join(., scaled_coef, by = "transect_ID")

CWM_fixed_filtered <- CWM_fixed %>% 
  drop_na(mean_umol, sta_mean_fixed) %>% 
  unite("site_transect", site, transect, remove = FALSE)

```

### i. Thickness
```{r thickness}
# plot 
ggplot(CWM_fixed_filtered, aes(x = z.mean_urchins, y = thickness_mm_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(aes(col = site), method = "lm") +
  theme_bw()

ggplot(CWM_fixed_filtered, aes(x = mean_urchins, y = thickness_mm_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(aes(col = site), method = "lm") +
  theme_bw()

ggplot(CWM_fixed_filtered, aes(x = thickness_mm_mean_fixed)) +
  geom_histogram(binwidth = 0.03)

model_thickness_01a <- glmmTMB(thickness_mm_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|transect) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass) 

dredge(model_thickness_01a) %>% 
  gt()

model_thickness_01b <- glmmTMB(thickness_mm_mean_fixed ~ z.mean_umol + (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass)

AIC(model_thickness_01a, model_thickness_01b)

# plot residuals
model_thickness_01a_resid <- simulateResiduals(model_thickness_01a)
plot(model_thickness_01a_resid)

model_thickness_01b_resid <- simulateResiduals(model_thickness_01b)
plot(model_thickness_01b_resid)

summary(model_thickness_01a)
summary(model_thickness_01b)

# check collinearity
check_collinearity(model_thickness_01a)
plot(check_collinearity(model_thickness_01a))
```

### ii. LDMC/TDMC
```{r ldmc}
ggplot(CWM_fixed_filtered, aes(x = site, y = tdmc_mean_fixed)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.5)

tdmc_aov <- aov(tdmc_mean_fixed ~ site, data = CWM_fixed_filtered)
summary(tdmc_aov)
TukeyHSD(tdmc_aov)

ggplot(CWM_fixed_filtered, aes(x = mean_urchins, y = tdmc_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(method = "lm")

ggplot(CWM_fixed, aes(x = tdmc_mean_fixed)) +
  geom_histogram(binwidth = 0.01)

model_tdmc_01a <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

plot(check_collinearity(model_tdmc_01a))

# no urchins
model_tdmc_01b <- glmmTMB(tdmc_mean_fixed ~ z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no substrate
model_tdmc_01c <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no kelp
model_tdmc_01d <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_umol + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no irradiance
model_tdmc_01e <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_urchins*z.mean_kelp +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no sand and no kelp
model_tdmc_01f <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_umol + z.mean_urchins*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no sand and no kelp and no interaction
model_tdmc_01g <- glmmTMB(tdmc_mean_fixed ~ z.mean_urchins + z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# plot residuals
plot(simulateResiduals(model_tdmc_01a))
plot(simulateResiduals(model_tdmc_01b))
plot(simulateResiduals(model_tdmc_01c))
check_model(model_tdmc_01c)
plot(simulateResiduals(model_tdmc_01d))
plot(simulateResiduals(model_tdmc_01e))
plot(simulateResiduals(model_tdmc_01f))
plot(simulateResiduals(model_tdmc_01g))

summary(model_tdmc_01a)
summary(model_tdmc_01b)
summary(model_tdmc_01c)
summary(model_tdmc_01g)

AIC(model_tdmc_01a, model_tdmc_01b, model_tdmc_01c, model_tdmc_01d, model_tdmc_01e, model_tdmc_01f, model_tdmc_01g)

predicted_fixed <- ggpredict(model_tdmc_01g, terms = ~ z.mean_urchins + (1|site), type = "random") %>% 
  rename("site" = group)

ggplot() +
  geom_point(data = CWM_fixed_filtered, aes(x = z.mean_urchins, y = tdmc_mean_fixed, col = site, shape = site), size = 3) +
  # scale_color_gradient2(low = "green", mid = "blue", high = "yellow") +
  # new_scale("col") +
  geom_line(data = predicted_fixed, aes(x = x, y = predicted, col = site), size = 1) + 
  theme_bw() +
  labs(x = "Urchin biomass (z-score standardized)",
       y = "Community weighted mean thallus dry matter content") +
  scale_x_continuous(expand = c(0, 0)) +
  # scale_y_continuous(limits = c(0, 0.45)) +
  theme(legend.position = c(0.9, 0.85)) +
  facet_wrap(~site)

  
tdmc_model_tbl <- model_tdmc_01g %>% 
  tbl_regression(intercept = TRUE) %>% 
   bold_p(t = 0.05)

MuMIn::r.squaredGLMM(model_tdmc_01g)
```

### iii. SA:P ratio
```{r sa-to-perimeter}
ggplot(CWM_fixed, aes(x = mean_urchins, y = sap_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(aes(col = site), method = "lm")

ggplot(CWM_fixed, aes(x = sap_mean_fixed)) +
  geom_histogram(binwidth = 1)

# full
model_sap_01a <- glmmTMB(sap_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no urchins
model_sap_01b <- glmmTMB(sap_mean_fixed ~ z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass)

# no substrate
model_sap_01c <- glmmTMB(sap_mean_fixed ~ z.mean_urchins + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no kelp
model_sap_01d <- glmmTMB(sap_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_umol + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# no irradiance
model_sap_01e <- glmmTMB(sap_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_urchins*z.mean_kelp +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass) 

# plot residuals
plot(simulateResiduals(model_sap_01a))
plot(simulateResiduals(model_sap_01b))
plot(simulateResiduals(model_sap_01c))
plot(simulateResiduals(model_sap_01d))
plot(simulateResiduals(model_sap_01e))

summary(model_sap_01a)
summary(model_sap_01b)
summary(model_sap_01c)
summary(model_sap_01d)
summary(model_sap_01e)

AIC(model_sap_01a, model_sap_01b, model_sap_01c, model_sap_01d, model_sap_01e)

predicted_sap <- ggpredict(model_sap_01d, terms = ~ z.mean_subs + (1|site), type = "random") %>% 
  rename("site" = group)

ggplot() +
  geom_point(data = CWM_fixed_filtered, aes(x = z.mean_subs, y = sap_mean_fixed, col = site)) +
  geom_line(data = predicted_sap, aes(x = x, y = predicted, col = site), size = 1) + 
  theme_bw() +
  labs(x = "Hard substrate cover (z-score standardized)",
       y = "Community weighted mean surface area:perimeter ratio") +
  scale_x_continuous(expand = c(0, 0)) +
  # scale_y_continuous(limits = c(0, 0.45)) +
  theme(legend.position = c(0.95, 0.85))

  
sap_model_tbl <- model_sap_01b %>% 
  tbl_regression(intercept = TRUE) %>% 
   bold_p(t = 0.05)

MuMIn::r.squaredGLMM(model_sap_01b)
```

### iv. FvFm
```{r fvfm}
ggplot(CWM_fixed_filtered, aes(x = site, y = fvfm_mean_fixed)) +
  geom_boxplot() +
  geom_jitter(alpha = 0.5)
  
ggplot(CWM_fixed_filtered, aes(x = z.mean_kelp, y = fvfm_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(aes(col = site), method = "lm")

ggplot(CWM_fixed_filtered, aes(x = fvfm_mean_fixed)) +
  geom_histogram(binwidth = 0.01)

# full
model_fvfm_01a <- glmmTMB(fvfm_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass) 

# best model
model_fvfm_01b <- glmmTMB(fvfm_mean_fixed ~ z.mean_umol + (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass) 

# second best model
model_fvfm_01c <- glmmTMB(fvfm_mean_fixed ~ z.mean_kelp + z.mean_umol + z.mean_urchins + z.mean_urchins*z.mean_kelp + (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass) 

# potentially useful model
model_fvfm_01d <- glmmTMB(fvfm_mean_fixed ~ z.mean_kelp + z.mean_urchins + z.mean_urchins*z.mean_kelp + (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "log"), na.action = na.pass) 

# plot residuals
model_fvfm_01a_resid <- simulateResiduals(model_fvfm_01a)
plot(model_fvfm_01a_resid)

model_fvfm_01b_resid <- simulateResiduals(model_fvfm_01b)
plot(model_fvfm_01b_resid)

model_fvfm_01c_resid <- simulateResiduals(model_fvfm_01c)
plot(model_fvfm_01c_resid)

plot(simulateResiduals(model_fvfm_01d))

summary(model_fvfm_01a)
summary(model_fvfm_01b)
summary(model_fvfm_01c)
summary(model_fvfm_01d)

AIC(model_fvfm_01a, model_fvfm_01b, model_fvfm_01c, model_fvfm_01d)

plot(check_collinearity(model_fvfm_01a))

predicted_fvfm <- ggpredict(model_fvfm_01b, terms = c("z.mean_umol", "site"), type = "random")

ggplot() +
  geom_point(data = CWM_fixed_filtered, aes(x = z.mean_umol, y = fvfm_mean_fixed, col = site)) +
  geom_line(data = predicted_fvfm, aes(x = x, y = predicted, col = group)) +
  geom_ribbon(data = predicted_fvfm, aes(x = x, ymax = conf.high, ymin = conf.low, fill = group), alpha = 0.1)
```

### v. max height
```{r max-height}
ggplot(CWM_fixed, aes(x = urchin_density, y = max_height_cm_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(aes(col = site), method = "lm") +
  theme(legend.position = "none") +
  facet_wrap(~site)

ggplot(CWM_fixed, aes(x = max_height_cm_fixed)) +
  geom_histogram(binwidth = 1)

# full
model_max_height_cm_01a <- glmmTMB(max_height_cm_fixed ~ urchin_density + mean_subs + mean_daily_umol + kelp_dry + (1|site) + (1|year),
        data = CWM_fixed, family = gaussian(link = "identity")) 

# site as fixed instead of random effect
model_max_height_cm_01b <- glmmTMB(max_height_cm_fixed ~ urchin_density + mean_subs + mean_daily_umol + kelp_dry + site + (1|year),
        data = CWM_fixed, family = gaussian(link = "identity")) 

# no site
model_max_height_cm_01c <- glmmTMB(max_height_cm_fixed ~ urchin_density + mean_subs + mean_daily_umol + kelp_dry + (1|year),
        data = CWM_fixed, family = gaussian(link = "identity")) 

# plot residuals
model_max_height_cm_01a_resid <- simulateResiduals(model_max_height_cm_01a)
plot(model_max_height_cm_01a_resid)

model_max_height_cm_01b_resid <- simulateResiduals(model_max_height_cm_01b)
plot(model_max_height_cm_01b_resid)

model_max_height_cm_01c_resid <- simulateResiduals(model_max_height_cm_01c)
plot(model_max_height_cm_01c_resid)

summary(model_max_height_cm_01a)
summary(model_max_height_cm_01b)
summary(model_max_height_cm_01c)

AIC(model_max_height_cm_01a, model_max_height_cm_01b, model_max_height_cm_01c)
```

### vi. STA

```{r}
ggplot(CWM_fixed_filtered, aes(x = mean_umol, y = sta_mean_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(method = "lm")

ggplot(CWM_fixed, aes(x = sta_mean_fixed)) +
  geom_histogram(binwidth = 1)

# full
model_sta_mean_01a <- glmmTMB(sta_mean_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass)

model_sta_mean_01b <- glmmTMB(sta_mean_fixed ~ z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_kelp*z.mean_umol + (1|site) + (1|year),
        data = CWM_fixed_filtered, family = gaussian(link = "identity"), na.action = na.pass)

# plot residuals
plot(simulateResiduals(model_sta_mean_01a))

plot(simulateResiduals(model_sta_mean_01b))

summary(model_sta_mean_01a)
summary(model_sta_mean_01b)

AIC(model_sta_mean_01a, model_sta_mean_01b)

predicted_sta <- ggpredict(model_sta_mean_01b, terms = ~ z.mean_umol + (1|site), type = "random")

ggplot() +
  geom_point(data = CWM_fixed_filtered, aes(x = z.mean_umol, y = sta_mean_fixed, col = site, shape = site), size = 3) +
  geom_line(data = predicted_sta, aes(x = x, y = predicted, col = group), size = 1) +
  # geom_ribbon(data = predicted_sta, aes(x = x, ymax = conf.high, ymin = conf.low, fill = group), alpha = 0.05) +
  labs(x = "Irradiance (z-score standardized)",
       y = "Community weighted mean of STA") +
  theme_bw() +
  theme(legend.position = c(0.95, 0.85))

sta_model_tbl <- model_sta_mean_01b %>% 
  tbl_regression(intercept = TRUE) %>% 
  bold_p(t = 0.05)
  
MuMIn::r.squaredGLMM(model_sta_mean_01a)

tbl_merge(tbls = list(tdmc_model_tbl, sta_model_tbl, sap_model_tbl),
          tab_spanner = c("TDMC", "STA", "SA:P"))
```

### vii. leathery macrophytes

```{r}
ggplot(CWM_fixed, aes(x = mean_urchins, y = growth_form_leathery_macrophyte_fixed)) +
  geom_point(aes(col = site)) +
  geom_smooth(method = "lm")

model_lm_01a <- glmmTMB(growth_form_leathery_macrophyte_fixed ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_urchins*z.mean_subs + z.mean_urchins*z.mean_kelp + z.mean_kelp*z.mean_subs + (1|site) + (1|year), 
                        ziformula = z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_urchins*z.mean_subs + z.mean_urchins*z.mean_kelp + z.mean_kelp*z.mean_subs + (1|site) + (1|year),
               data = CWM_fixed, family = beta_family(link = "logit"), na.action = na.pass)

MuMIn::dredge(model_lm_01a) %>% 
  gt()

model_lm_01b <- glmmTMB(growth_form_leathery_macrophyte_fixed ~ z.mean_urchins + (1|site) + (1|year), 
               data = CWM_fixed, family = beta_family(link = "logit"), na.action = na.pass)

plot(simulateResiduals(model_lm_01a))
plot(simulateResiduals(model_lm_01b))

summary(model_lm_01a)
summary(model_lm_01b)

AIC(model_lm_01a, model_lm_01b)

predicted_lm <- ggpredict(model_lm_01b, terms = c("z.mean_urchins"))

ggplot() +
  geom_point(data = CWM_fixed, aes(x = z.mean_urchins, y = growth_form_leathery_macrophyte_fixed, col = site)) +
  geom_line(data = predicted_lm, aes(x = x, y = predicted)) +
  geom_ribbon(data = predicted_lm, aes(x = x, ymax = conf.high, ymin = conf.low), alpha = 0.2)
```

### viii. coloring NMDS results by CWM

```{r}
CWM_NMDS <- community_mds_sites_euclidean %>% 
  dplyr::select(transect_ID, NMDS1, NMDS2) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  mutate_at(4:31, as.numeric)

nmds_plot_euclidean <- ggplot() +
  geom_point(data = CWM_NMDS, aes(x = NMDS1, y = NMDS2, color = tdmc_mean_fixed, shape = site), size = 2) +
  scale_colour_gradient(low = "yellow", high = "darkblue") +
  new_scale_color() + 
  # scale_color_manual(values = pal) +
  geom_vline(xintercept = c(0), color = "grey70", linetype = 2) +
  geom_hline(yintercept = c(0), color = "grey70", linetype = 2) +
  # stat_ellipse(data = CWM_NMDS %>% filter(site != "NA"), aes(x = NMDS1, y = NMDS2, color = site), level = 0.95, size = 1) +
  # geom_segment(data = community_mds_spp_euclidean, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm"))) +
  # geom_text(data = community_mds_spp_euclidean, aes(x = NMDS1, y = NMDS2, label = scientific_name)) +
  geom_segment(data = fit_vect_euclidean, aes(x = 0, y = 0, xend = NMDS1, yend = NMDS2), arrow = arrow(length = unit(0.2, "cm")), col = "blue") +
  geom_text(data = fit_vect_euclidean, aes(x = NMDS1, y = NMDS2, label = rownames(fit_vect_euclidean)), col = "blue") +
  theme_bw() +
  guides(colour = guide_legend(ncol = 1)) +
  theme(legend.box = "vertical", legend.margin = margin()) 

nmds_plot_euclidean
```


## b. specific CWM using `FD::dbFD()` 

### i. Bullito

Only one species, skip.

```{r bull, include = FALSE, eval = FALSE}
# get a site specific trait by species matrix
bull_tbspp <- specif_tbspp("Bullito")

# get a site specific transect by species matrix
bull_commat <- specif_commat("bull", bull_tbspp)

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(bull_tbspp) == colnames(bull_commat)

# calculate CWM for site
CWM_bull <- functcomp(x = bull_tbspp, a = bull_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("sample_ID") %>% 
  rename_with(.fn = ~ paste0(., "_specif", sep = ""), .cols = 2:ncol(.)) %>% 
  left_join(., CWM_fixed, by = "transect_ID")
```

### ii. Arroyo Quemado

Only one species, skip.

```{r aque, include = FALSE, eval = FALSE}
# get a site specific trait by species matrix
aque_tbspp <- specif_tbspp("Arroyo Quemado")

# get a site specific transect by species matrix
aque_commat <- specif_commat("aque", aque_tbspp)

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(aque_tbspp) == colnames(aque_commat)

# calculate CWM for site
CWM_aque <- functcomp(x = aque_tbspp, a = aque_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("transect_ID") %>% 
  rename_with(.fn = ~ paste0(., "_specif", sep = ""), .cols = 2:ncol(.)) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  mutate(site = "aque")
```

### iii. Isla Vista

```{r ivee}
# get a site specific trait by species matrix
ivee_tbspp <- specif_tbspp("Isla Vista")

# get a site specific transect by species matrix
ivee_commat <- specif_commat("ivee", ivee_tbspp)

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(ivee_tbspp) == colnames(ivee_commat)

# calculate CWM for site
CWM_ivee <- functcomp(x = ivee_tbspp, a = ivee_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("transect_ID") %>% 
  rename_with(.fn = ~ paste0(., "_specif", sep = ""), .cols = 2:ncol(.)) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>%  
  mutate(site = "ivee")
```

### iv. Mohawk

```{r mohk}
# get a site specific trait by species matrix
mohk_tbspp <- specif_tbspp("Mohawk")

# get a site specific transect by species matrix
mohk_commat <- specif_commat("mohk", mohk_tbspp)

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(mohk_tbspp) == colnames(mohk_commat)

# calculate CWM for site
CWM_mohk <- functcomp(x = mohk_tbspp, a = mohk_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("transect_ID") %>% 
  rename_with(.fn = ~ paste0(., "_specif", sep = ""), .cols = 2:ncol(.)) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  mutate(site = "mohk")
```

### v. Carpinteria

Only one species, skip.

```{r carp, include = FALSE, eval = FALSE}
# get a site specific trait by species matrix
carp_tbspp <- specif_tbspp("Carpinteria")

# get a site specific transect by species matrix
carp_commat <- specif_commat("carp", carp_tbspp)

# making sure the rownames in the trait matrix == columns in the community matrix
rownames(carp_tbspp) == colnames(carp_commat)

# calculate CWM for site
CWM_carp <- functcomp(x = carp_tbspp, a = carp_commat, 
                  CWM.type = c("all")) %>% 
  rownames_to_column("transect_ID") %>% 
  rename_with(.fn = ~ paste0(., "_specif", sep = ""), .cols = 2:ncol(.)) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  mutate(site = "carp")
```

## c. putting all the data frames together

```{r all-together}
df <- CWM_mohk %>% 
  select(transect_ID, fvfm_mean_specif, tdmc_mean_specif, sta_mean_specif, thickness_mm_mean_specif, site) %>% 
  rbind(., (CWM_ivee %>% select(transect_ID, fvfm_mean_specif, tdmc_mean_specif, sta_mean_specif, thickness_mm_mean_specif, site))) %>% 
  rbind(., (CWM_aque %>% select(transect_ID, fvfm_mean_specif, tdmc_mean_specif, sta_mean_specif, thickness_mm_mean_specif, site))) %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  rename("site" = site.x)
```


## d. variance decomp using `cati::traitflex.anova()`

```{r variance-decomp}
tdmc_tf <- traitflex.anova(~ site, specif.avg = tdmc_mean_specif, const.avg = tdmc_mean_fixed, data = df)
tdmc_tf
plot(tdmc_tf)

thickness_tf <- traitflex.anova(~ site, specif.avg = thickness_mm_mean_specif, const.avg = thickness_mm_mean_fixed, data = df)
thickness_tf
plot(thickness_tf)

sta_tf <- traitflex.anova(~ site, specif.avg = sta_mean_specif, const.avg = sta_mean_fixed, data = df)
sta_tf
plot(sta_tf)
```

## e. functional diversity

### i. continuous and categorical traits

```{r}
tbspp_mat_num <- ind_traits %>% 
  filter(sp_code %in% algae_proposal) %>% 
  filter(!(sp_code %in% c("CO"))) %>% 
  # weird Nienburgia?
  filter(!(specimen_ID == "20210721-BULL-023")) %>% 
  mutate(growth_form_num = case_when(
    growth_form == "leathery_macrophyte" ~ 1,
    growth_form == "corticated_macrophytes" ~ 2,
    growth_form == "corticated_foliose" ~ 3
  ), 
  pigment_type_num = case_when(
    pigment_type == "red" ~ 1,
    pigment_type == "brown" ~ 2
  ), 
  longevity_num = case_when(
    longevity == "annual" ~ 1,
    longevity == "perennial" ~ 3,
    longevity == "annual or perennial" ~ 2
  )) %>% 
  dplyr::select(specimen_ID, maximum_height, sta_mean, tdmc_mean, sav_mean, sap_mean, fvfm_mean, total_volume, total_dry, thickness_mm_mean, frond_length_mean, frond_width_mean, growth_form_num, pigment_type_num, longevity_num) %>% 
  left_join(., metadata_ind, by = "specimen_ID") %>% 
  group_by(sp_code) %>% 
  summarize(across(2:15, ~ mean(.x, na.rm = TRUE))) %>% 
  column_to_rownames("sp_code")

commat_dbFD <- biomass %>% 
  dplyr::select(sample_ID, sp_code, dry_gm2) %>% 
  pivot_wider(names_from = sp_code, values_from = dry_gm2) %>% 
    # filter(sample_ID %in% CWM_fixed_filtered$transect_ID) %>% 
  column_to_rownames("sample_ID") %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(c(rownames(tbspp_mat_num))) %>% 
  # take out 0 sites
  mutate(total = rowSums(across(where(is.numeric)))) %>% 
  filter(total > 0) %>% 
  dplyr::select(-total) %>% 
  as.matrix()

fd_cont <- dbFD(x = tbspp_mat_num, a = commat_dbFD)

# FEVe: Could not be calculated for communities with <3 functionally singular species. 
# FDis: Equals 0 in communities with only one functionally singular species. 
# FRic: To respect s > t, FRic could not be calculated for communities with <3 functionally singular species. 
# FRic: Dimensionality reduction was required. The last 5 PCoA axes (out of 7 in total) were removed. 
# FRic: Quality of the reduced-space representation = 0.738934 
# FDiv: Could not be calculated for communities with <3 functionally singular species. 

fd_cont

# species richness
spp_rich <- specnumber(commat_dbFD) %>% 
  as.data.frame() %>% 
  rename("spprich" = '.') %>% 
  rownames_to_column("sample_ID")

spp_div <- diversity(commat_dbFD, index = "shannon") %>% 
  as.data.frame() %>% 
  rename("shandiv" = '.') %>% 
  rownames_to_column("sample_ID")

fric_df <- fd_cont$FRic %>% 
  as.data.frame() %>% 
  rename("fric" = '.') %>% 
  rownames_to_column("sample_ID")

fdis_df <- fd_cont$FDis %>% 
  as.data.frame() %>% 
  rename("fdis" = '.') %>% 
  rownames_to_column("sample_ID")

fdiv_df <- fd_cont$FDiv %>% 
  as.data.frame() %>% 
  rename("fdiv" = '.') %>% 
  rownames_to_column("sample_ID")

fd_sum_df <- fd_cont$FEve %>% 
  as.data.frame() %>% 
  rename("feve" = '.') %>% 
  rownames_to_column("sample_ID") %>% 
  left_join(., fric_df, by = "sample_ID") %>% 
  left_join(., fdiv_df, by = "sample_ID") %>% 
  left_join(., fdis_df, by = "sample_ID") %>% 
  left_join(., community_metadata, by = "sample_ID") %>% 
  left_join(., spp_rich, by = "sample_ID") %>% 
  left_join(., spp_div, by = "sample_ID") %>% 
  drop_na()

ggplot(fd_sum_df,
       aes(x = spprich,
           y = fric)) +
  geom_point()

mod <- glmmTMB(fric ~ spprich + (1|site/transect),
               data = fd_sum_df)

simulateResiduals(mod, plot = TRUE)


# richness
fric_lmm <- glmmTMB(fric ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|transect) + (1|year),
        data = fd_sum_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(fric_lmm))
summary(fric_lmm)

fric_lmm %>% 
  tbl_regression() %>% 
  bold_p(t = 0.05)

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_kelp, y = fric, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fric), method = "lm")

ggplot(fd_sum_df, aes(x = year, y = fric, col = site, fill = site)) +
  geom_point() +
  geom_smooth(method = "lm")

# evenness
feve_lmm <- glmmTMB(feve ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|transect) + (1|year),
        data = fd_sum_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(feve_lmm))
summary(feve_lmm)

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_kelp, y = feve, col = z.mean_umol)) +
  geom_smooth(aes(x = z.mean_kelp, y = feve), method = "lm")

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_umol, y = feve, col = site)) +
  geom_smooth(aes(x = z.mean_umol, y = feve), method = "lm")

# divergence
fdiv_lmm <- glmmTMB(fdiv ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|site) + (1|year),
        data = fd_sum_df, family = gaussian(link = "identity"), na.action = na.pass) 
plot(simulateResiduals(fdiv_lmm))
summary(fdiv_lmm)
fdiv_lmm %>% 
  tbl_regression() %>% 
  bold_p(t = 0.05)
MuMIn::r.squaredGLMM(fdiv_lmm)

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_kelp, y = fdiv, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fdiv), method = "lm") +
  theme_bw() +
  labs(x = "Kelp biomass (z-score standardized)",
       y = "Functional divergence") +
  theme(axis.title = element_text(size = 10))


# dispersion
fdis_lmm <- glmmTMB(fdis ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|site) + (1|year),
        data = fd_sum_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(fdis_lmm))
summary(fdis_lmm)

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_kelp, y = fdis, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fdis), method = "lm") 

fd_sum_df %>% 
  ggplot(aes(x = shandiv, y = fric)) +
  geom_point(aes(col = site)) +
  geom_smooth(method = "lm")

div_lm <- lm(fric ~ shandiv, data = fd_sum_df)
summary(div_lm)
cor.test(fd_sum_df$fric, fd_sum_df$shandiv, method = "pearson")
```

### ii. categorical traits on whole data set

```{r}
tbspp_mat_cat <- coarse_traits %>%
  dplyr::select(sp_code, growth_form, pigment_type, life_habit, longevity, posture) %>% 
  filter(sp_code != "NEO") %>% 
  column_to_rownames("sp_code") %>% 
  mutate_all(as.factor)

commat_dbFD_cat <- biomass %>% 
  dplyr::select(-sample_ID) %>% 
  unite("sample_ID", site, year, remove = FALSE) %>% 
  unite("transect_ID", site, year, transect, remove = FALSE) %>% 
  filter(sp_code != "NEO") %>% 
  # filter(sp_code %in% algae_proposal) %>% 
  dplyr::select(transect_ID, sp_code, dry_gm2) %>% 
  pivot_wider(names_from = sp_code, values_from = dry_gm2) %>% 
  filter(transect_ID %in% CWM_fixed_filtered$transect_ID) %>% 
  column_to_rownames("transect_ID") %>% 
  mutate_all(~replace(., is.na(.), 0)) %>% 
  dplyr::select(c(rownames(tbspp_mat_cat))) %>% 
  # take out 0 sites
  mutate(total = rowSums(across(where(is.numeric)))) %>% 
  filter(total > 0) %>% 
  dplyr::select(-total) %>% 
  # take out 0 species
  as.matrix()

# species richness
spp_rich <- specnumber(commat_dbFD_cat) %>% 
  as.data.frame() %>% 
  rename("spprich" = '.') %>% 
  rownames_to_column("transect_ID")

spp_div <- diversity(commat_dbFD_cat, index = "shannon") %>% 
  as.data.frame() %>% 
  rename("shandiv" = '.') %>% 
  rownames_to_column("transect_ID")

cat_fd <- dbFD(x = tbspp_mat_cat, a = commat_dbFD_cat, corr = c("none"))

# Species x species distance was not Euclidean, but no correction was applied. Only the PCoA axes with positive eigenvalues were kept. 
# FEVe: Could not be calculated for communities with <3 functionally singular species. 
# FRic: Only categorical and/or ordinal trait(s) present in 'x'. FRic was measured as the number of unique trait combinations, NOT as the convex hull volume. 
# FDiv: Cannot be computed when only categorical and/or ordinal trait(s) present in 'x'. 

cat_fd

fric_df <- cat_fd$FRic %>% 
  as.data.frame() %>% 
  rename("fric" = '.') %>% 
  rownames_to_column("transect_ID")

fdis_df <- cat_fd$FDis %>% 
  as.data.frame() %>% 
  rename("fdis" = '.') %>% 
  rownames_to_column("transect_ID")

raoq_df <- cat_fd$RaoQ %>% 
  as.data.frame() %>% 
  rename("raoq" = '.') %>% 
  rownames_to_column("transect_ID")

fd_cat_df <- cat_fd$FEve %>% 
  as.data.frame() %>% 
  rename("feve" = '.') %>% 
  rownames_to_column("transect_ID") %>% 
  left_join(., fric_df, by = "transect_ID") %>% 
  left_join(., raoq_df, by = "transect_ID") %>% 
  left_join(., fdis_df, by = "transect_ID") %>% 
  left_join(., CWM_fixed_filtered, by = "transect_ID") %>% 
  left_join(., spp_rich, by = "transect_ID") %>% 
  left_join(., spp_div, by = "transect_ID") %>% 
  drop_na(feve)

# richness
fric_lmm <- glmmTMB(fric ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|site) + (1|year),
        data = fd_cat_df, family = gaussian(link = "identity"), na.action = na.pass) 
plot(simulateResiduals(fric_lmm))
summary(fric_lmm)

ggplot(fd_cat_df) +
  geom_point(aes(x = z.mean_kelp, y = fric, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fric), method = "lm")

# evenness
feve_lmm <- glmmTMB(feve ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|site) + (1|year),
        data = fd_cat_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(feve_lmm))
summary(feve_lmm)

ggplot(fd_sum_df) +
  geom_point(aes(x = z.mean_kelp, y = feve, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = feve), method = "lm")

# divergence
raoq_lmm <- glmmTMB(raoq ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|site) + (1|year),
        data = fd_cat_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(raoq_lmm))
summary(raoq_lmm)
raoq_lmm %>% 
  tbl_regression() %>% 
  bold_p(t = 0.05)
MuMIn::r.squaredGLMM(raoq_lmm)

ggplot(fd_cat_df) +
  geom_point(aes(x = z.mean_kelp, y = raoq, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = raoq), method = "lm", se = FALSE) +
  scale_y_continuous(limits = c(0, 0.3)) +
  theme(legend.position = c(0.9, 0.8)) +
  theme_bw()

# dispersion
fdis_lmm <- glmmTMB(fdis ~ z.mean_urchins + z.mean_subs + z.mean_kelp + z.mean_umol + z.mean_urchins*z.mean_kelp + z.mean_urchins*z.mean_umol + z.mean_subs*z.mean_umol + z.mean_kelp*z.mean_umol +
                                 + (1|transect) + (1|year),
        data = fd_cat_df, family = gaussian(link = "log"), na.action = na.pass) 
plot(simulateResiduals(fdis_lmm))
summary(fdis_lmm)

ggplot(fd_cat_df) +
  geom_point(aes(x = z.mean_kelp, y = fdis, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fdis), method = "lm")

ggplot(fd_cat_df) +
  geom_jitter(aes(x = z.mean_kelp, y = spprich, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = spprich), method = "lm")

ggplot(fd_cat_df) +
  geom_jitter(aes(x = z.mean_kelp, y = fric, col = site)) +
  geom_smooth(aes(x = z.mean_kelp, y = fric), method = "lm")

fd_cat_df %>% 
  ggplot(aes(x = spprich, y = fric)) +
  geom_point(aes(col = site)) +
  geom_smooth()

rich_lm <- lm(raoq ~ shandiv, data = fd_cat_df)
summary(rich_lm)

```



```{r include = FALSE, eval = FALSE}
# Use `rmarkdown::render(here::here("code", "04_benthic-data.Rmd"))` in the console to knit the document.
```
